Chainweb.BlockHash

Description

TODO

BlockHash

newtype BlockHash

Note:

Serialization as JSON key doesn't include the chain id. Note, however that the chain id is included in the hash.

Serialization as JSON property includes the chain id, because it can't be recovered from the hash. Including it gives extra type safety across serialization roundtrips.

Constructors

BlockHash MerkleLogHash

encodeBlockHash :: MonadPut m => BlockHash -> m ()

decodeBlockHash :: MonadGet m => m BlockHash

randomBlockHash :: MonadIO m => m BlockHash

nullBlockHash :: BlockHash

blockHashToText :: BlockHash -> Text

blockHashFromText :: MonadThrow m => Text -> m BlockHash

Block Hash Record

newtype BlockHashRecord

Constructors

BlockHashRecord

Fields

_getBlockHashRecord :: HashMap ChainId BlockHash

getBlockHashRecord :: Iso' BlockHashRecord (HashMap ChainId BlockHash)

encodeBlockHashRecord :: MonadPut m => BlockHashRecord -> m ()

decodeBlockHashRecord :: MonadGet m => m BlockHashRecord

decodeBlockHashRecordChecked :: MonadThrow m => MonadGet m => HasChainId p => Expected [p] -> m BlockHashRecord

blockHashRecordToVector :: BlockHashRecord -> Vector BlockHash

blockHashRecordFromVector :: HasChainGraph g => HasChainId c => g -> c -> Vector BlockHash -> BlockHashRecord

blockHashRecordChainIdx :: BlockHashRecord -> ChainId -> Maybe Int

Exceptions