Chainweb.BlockHeader

Description

Block Height

newtype BlockHeight

BlockHeight

Constructors

BlockHeight

Fields

_height :: Word64

encodeBlockHeight :: MonadPut m => BlockHeight -> m ()

decodeBlockHeight :: MonadGet m => m BlockHeight

encodeBlockHeightBe :: MonadPut m => BlockHeight -> m ()

decodeBlockHeightBe :: MonadGet m => m BlockHeight

Block Weight

newtype BlockWeight

Constructors

BlockWeight HashDifficulty

encodeBlockWeight :: MonadPut m => BlockWeight -> m ()

decodeBlockWeight :: MonadGet m => m BlockWeight

encodeBlockWeightBe :: MonadPut m => BlockWeight -> m ()

decodeBlockWeightBe :: MonadGet m => m BlockWeight

Block Payload Hash

newtype BlockPayloadHash

Constructors

BlockPayloadHash MerkleLogHash

encodeBlockPayloadHash :: MonadPut m => BlockPayloadHash -> m ()

decodeBlockPayloadHash :: MonadGet m => m BlockPayloadHash

hashPayload :: HasChainId p => ChainwebVersion -> p -> ByteString -> BlockPayloadHash

Nonce

newtype Nonce

FIXME: is 64 bit enough for the nonce. It seems that it may not be sufficient for the current hashpower of the bitcoin network.

Constructors

Nonce Word64

encodeNonce :: MonadPut m => Nonce -> m ()

encodeNonceToWord64 :: Nonce -> Word64

decodeNonce :: MonadGet m => m Nonce

BlockCreationTime

newtype BlockCreationTime

Constructors

BlockCreationTime

Fields

_bct :: Time Micros

encodeBlockCreationTime :: MonadPut m => BlockCreationTime -> m ()

decodeBlockCreationTime :: MonadGet m => m BlockCreationTime

EpochStartTime

newtype EpochStartTime

Constructors

EpochStartTime (Time Micros)

encodeEpochStartTime :: MonadPut m => EpochStartTime -> m ()

decodeEpochStartTime :: MonadGet m => m EpochStartTime

epochStart

Arguments

:: BlockHeader
parent header

-> BlockCreationTime
block creation time of new block

-> EpochStartTime
epoch start time of new block

Compute the epoch start value for a new BlockHeader

FeatureFlags

newtype FeatureFlags

Constructors

FeatureFlags Word64

encodeFeatureFlags :: MonadPut m => FeatureFlags -> m ()

decodeFeatureFlags :: MonadGet m => m FeatureFlags

POW Target

powTarget

Arguments

:: BlockHeader
parent header

-> BlockCreationTime
block creation time of new block

-> HashTarget
POW target of new block

Compute the POW target for a new BlockHeader.

BlockHeader

data BlockHeader :: Type where

BlockHeader

Some redundant, aggregated information is included in the block and the block hash. This enables nodes to be checked inductively with respect to existing blocks without recalculating the aggregated value from the genesis block onward.

The POW hash is not include, since it can be derived from the Nonce and the other fields of the BlockHeader.

IMPORTANT: Fields in this record must have pairwise distinct types.

Constructors

BlockHeader

Fields

:: { _blockNonce :: !Nonce

authoritative

, _blockCreationTime :: !BlockCreationTime

the time when the block was creates as recorded by the miner of the block. The value must be strictly monotonically increasing with in the chain of blocks. The smallest allowed increment is smallestBlockTimeIncrement. Nodes are supposed to ignore blocks with values that are in the future and reconsider a block when its value is in the past.

The block creation time is used to determine the block difficulty for future blocks.

Nodes are not supposed to consider the creation time when choosing between two valid (this include that creation times must not be in the future) forks.

This creates an incentive for nodes to maintain an accurate clock with respect to a (unspecified) commonly accepted time source, such as the public NTP network.

It is possible that a miner always chooses the smallest possible creation time value. It is not clear what advantage a miner would gain from doing so, but attack models should consider and investigate such behavior.

On the other hand miners may choose to compute forks with creation time long in the future. By doing so, the difficulty on such a fork would decrease allowing the miner to compute very long chains very quickly. However, those chains would become valid only after a long time passed. The algorithm for computing the difficulty must ensure this strategy doesn't give an advantage to an attacker that would increase the success probability for an attack.

, _blockParent :: !BlockHash

authoritative

, _blockAdjacentHashes :: !BlockHashRecord

authoritative

, _blockTarget :: !HashTarget

authoritative

, _blockPayloadHash :: !BlockPayloadHash

authoritative

, _blockChainId :: !ChainId

, _blockWeight :: !BlockWeight

the accumulated weight of the chain. It is redundant information that is subject to the inductive property that the block weight of a block is the block weight of the parent plus the difficulty of the block.

, _blockHeight :: !BlockHeight

block height records the length of the chain. It is redundant information and thus subject the inductive property that the block height of a block is the block height of its parent plus one.

, _blockChainwebVersion :: !ChainwebVersion

the Chainweb version is a constant for the chain. A chain is uniquely identified by its genesis block. Thus this is redundant information and thus subject to the inductive property that the Chainweb version of a block equals the Chainweb version of its parent.

, _blockEpochStart :: !EpochStartTime

The start time of the current difficulty adjustment epoch. Epochs divide the sequence of blocks in the chain into continuous ranges of blocks. Each epoch is defined by the minimal block height of the blocks in the epoch.

, _blockFlags :: !FeatureFlags

An 8-byte bitmask reserved for the future addition of boolean "feature flags".

, _blockHash :: !BlockHash

the hash of the block. It includes all of the above block properties.

} -> BlockHeader

blockNonce :: Lens' BlockHeader Nonce

blockChainId :: Lens' BlockHeader ChainId

blockHeight :: Lens' BlockHeader BlockHeight

blockWeight :: Lens' BlockHeader BlockWeight

blockChainwebVersion :: Lens' BlockHeader ChainwebVersion

blockAdjacentHashes :: Lens' BlockHeader BlockHashRecord

blockCreationTime :: Lens' BlockHeader BlockCreationTime

blockHash :: Lens' BlockHeader BlockHash

blockParent :: Lens' BlockHeader BlockHash

blockPayloadHash :: Lens' BlockHeader BlockPayloadHash

blockTarget :: Lens' BlockHeader HashTarget

blockEpochStart :: Lens' BlockHeader EpochStartTime

blockFlags :: Lens' BlockHeader FeatureFlags

_blockPow :: BlockHeader -> PowHash

The Proof-Of-Work hash includes all data in the block except for the _blockHash. The value (interpreted as BlockHashNat must be smaller than the value of _blockTarget (interpreted as BlockHashNat).

blockPow :: Getter BlockHeader PowHash

_blockAdjacentChainIds :: BlockHeader -> HashSet ChainId

blockAdjacentChainIds :: Getter BlockHeader (HashSet ChainId)

encodeBlockHeader :: MonadPut m => BlockHeader -> m ()

encodeBlockHeaderWithoutHash :: MonadPut m => BlockHeader -> m ()

decodeBlockHeader :: MonadGet m => m BlockHeader

Decode a BlockHeader and trust the result

decodeBlockHeaderWithoutHash :: MonadGet m => m BlockHeader

Decode a BlockHeader and trust the result

decodeBlockHeaderChecked :: MonadThrow m => MonadGet m => m BlockHeader

Decode and check that

chain id is in graph

all adjacentParent match adjacents in graph

decodeBlockHeaderCheckedChainId :: MonadThrow m => MonadGet m => HasChainId p => Expected p -> m BlockHeader

Decode and check that

chain id is in graph

all adjacentParent match adjacents in graph

chainId matches the expected chain id

newtype ObjectEncoded a

By default a binary encoding of block headers is used as JSON encoding. In some circumstance, like logging and configuration files, a textual encoding is desired.

Constructors

ObjectEncoded

Fields

_objectEncoded :: a

timeBetween :: BlockCreationTime -> BlockCreationTime -> Micros

The number of microseconds between the creation time of two BlockHeaders.

getAdjacentHash :: MonadThrow m => HasChainId p => p -> BlockHeader -> m BlockHash

computeBlockHash :: BlockHeader -> BlockHash

adjacentChainIds :: HasChainId p => ChainGraph -> p -> HashSet ChainId

IsBlockHeader

class IsBlockHeader t where

Any type which can purely produce a BlockHeader, or purely construct one.

Methods

isoBH :: Iso' t BlockHeader

Genesis BlockHeader

isGenesisBlockHeader :: BlockHeader -> Bool

Create a new BlockHeader

newBlockHeader

Arguments

:: BlockHashRecord
Adjacent parent hashes

-> BlockPayloadHash
payload hash

-> Nonce
Randomness to affect the block hash

-> Time Micros
Creation time of the block

-> BlockHeader
parent block header

-> BlockHeader

Testing

testBlockHeader

Arguments

:: BlockHashRecord
Adjacent parent hashes

-> Nonce
Randomness to affect the block hash

-> BlockHeader
parent block header

-> BlockHeader

testBlockHeaders :: BlockHeader -> [BlockHeader]

Given a BlockHeader of some initial parent, generate an infinite stream of BlockHeaders which form a legal chain.

Should only be used for testing purposes.

testBlockHeadersWithNonce :: Nonce -> BlockHeader -> [BlockHeader]

Given a BlockHeader of some initial parent, generate an infinite stream of BlockHeaders which form a legal chain.

Should only be used for testing purposes.

testBlockPayload :: BlockHeader -> BlockPayloadHash

CAS Constraint

type BlockHeaderCas cas = CasConstraint cas BlockHeader