Chainweb.BlockHeader.Genesis

Description

Hard-coded Genesis blocks for various versions of Chainweb.

Genesis Blocks

Creation

genesisBlockHeader :: HasChainId p => ChainwebVersion -> p -> BlockHeader

A block chain is globally uniquely identified by its genesis hash. Internally, we use the ChainwebVersion value and the ChainId as identifiers. We thus include the ChainwebVersion value and the ChainId into the genesis block hash.

We assume that there is always only a single ChainwebVersion in scope and identify chains only by their internal ChainId.

genesisBlockHeader' :: HasChainId p => ChainwebVersion -> p -> BlockCreationTime -> Nonce -> BlockHeader

Like genesisBlockHeader, but with slightly more control.

genesisBlockHeaders :: ChainwebVersion -> HashMap ChainId BlockHeader

This is an expensive call, try not to repeat it.

Querying

genesisBlockPayload :: ChainwebVersion -> ChainId -> PayloadWithOutputs

genesisParentBlockHash :: HasChainId p => ChainwebVersion -> p -> BlockHash

The genesis block hash includes the Chainweb version and the ChainId within the Chainweb version.

It is the _blockParent of the genesis block

genesisBlockTarget :: HashTarget

By definition, Genesis Blocks are "mined" on the easiest difficulty. No subsequent block mining can have a HashTarget easier (re: higher) than this. Equivalent to maxTarget.

genesisTime :: ChainwebVersion -> BlockCreationTime

The moment of creation of a Genesis Block. For test chains, this is the Linux Epoch. Production chains are otherwise fixed to a specific timestamp.