Chainweb.BlockHeader.Validation

Description

TODO

Validation Failures

data ValidationFailure

A data type that describes a failure of validating a block header.

Constructors

ValidationFailure

Fields

_validateFailureParent :: !(Maybe BlockHeader)

_validateFailureHeader :: !BlockHeader

_validationFailureFailures :: ![ValidationFailureType]

data ValidationFailureType

An enumeration of possible validation failures for a block header.

Constructors

MissingParent
Parent isn't in the database

CreatedBeforeParent
Claims to be created at a time prior to its parent's creation

VersionMismatch
Claims to use a version of chainweb different from that of its parent

ChainMismatch
Claims to use a chain-id different from that of its parent

IncorrectHash
The hash of the header properties as computed by computeBlockHash does not match the hash given in the header

IncorrectPow
The POW hash of the header does not match the POW target of the block.

IncorrectHeight
The given height is not one more than the parent height

IncorrectWeight
The given weight is not the sum of the target difficulty and the parent's weight

IncorrectTarget
The given target of the block is not correct. The target of the first block of an epoch is the adjusted target of the previous epoch. For all other blocks the target equals the target of the parent block.

IncorrectEpoch
The epoch start time value of the block is incorrect. The epoch start time of the first block of an epoch equals the block creation time of that block. For all other blocks the epoch start time equls the epoch start time of the parent block.

IncorrectGenesisParent
The block is a genesis block, but doesn't have its parent set to its own hash.

IncorrectGenesisTarget
The block is a genesis block, but doesn't have the correct POW target.

BlockInTheFuture
The creation time of the block is in the future.

IncorrectPayloadHash
The validation of the payload hash failed.

MissingPayload
The payload for the block is missing.

definiteValidationFailures :: [ValidationFailureType]

The list of validation failures that are definite and independent of any external context. A block for which validation fails with one of these failures must be dicarded.

No node on the chainweb-web network should propgate blocks with these failures. If a block is received that causes a definite validation failures is receveived from a chainweb-node, that chainweb-node should be blacklisted/removed in the peer database.

isDefinite :: [ValidationFailureType] -> Bool

Predicate that checks whether a validation failure is definite.

ephemeralValidationFailures :: [ValidationFailureType]

The list of ephemaral validation failures. These failure depend on a local context and validation may succeed in the future.

In case of BlockInTheFuture validation will eventually succeed.

isEphemeral :: [ValidationFailureType] -> Bool

A predicate that checks whether a validation failure is ephemeral.

Validation functions

validateBlockHeader

Arguments

:: Time Micros
The current clock time

-> BlockHeader
parent block header. The genesis header is considered its own parent.

-> BlockHeader
The block header to be checked

-> [ValidationFailureType]
A list of ways in which the block header isn't valid

Validate properties of the block header, producing a list of the validation failures.

This doesn't include checks for

MissingPayload

MissingParent

IncorrectPayloadHash

validateBlockHeaderM

Arguments

:: MonadThrow m

=> Time Micros
The current clock time

-> BlockHeader
parent block header. The genesis header is considered its own parent.

-> BlockHeader
The block header to be checked

-> m ()

Validate properties of the block header, throwing an exception detailing the failures if any.

MissingPayload

MissingParent

IncorrectPayloadHash

isValidBlockHeader

Arguments

:: Time Micros
The current clock time

-> BlockHeader
parent block header. The genesis header is considered its own parent.

-> BlockHeader
The block header to be checked

-> Bool

Check whether a BlockHeader satisfies all validaiton properties.

MissingPayload

MissingParent

IncorrectPayloadHash

validateIntrinsic

Arguments

:: Time Micros
The current clock time

-> BlockHeader
block header to be validated

-> [ValidationFailureType]
A list of ways in which the block header isn't valid

Validates properties of a block which are checkable from the block header without observing the remainder of the database.

validateInductive

Arguments

:: BlockHeader
parent block header. The genesis header is considered its own parent.

-> BlockHeader
Block header under scrutiny

-> [ValidationFailureType]
A list of ways in which the block header isn't valid

Validate properties of a block with respect to a given parent.

validateBlockParentExists :: Monad m => (BlockHash -> m (Maybe BlockHeader)) -> BlockHeader -> m (Either ValidationFailureType BlockHeader)

Validate that the parent exist with the given lookup function.

Returns the parent if it exists. Throws an error otherwise.

validateBlocksM

Arguments

:: MonadThrow m

=> Time Micros
The current clock time

-> (BlockHash -> m (Maybe BlockHeader))

-> HashMap BlockHash BlockHeader

-> m ()

Validate a set of blocks that may depend on each other.

MissingPayload

IncorrectPayloadHash

Intrinsic BlockHeader Properties

prop_block_pow :: BlockHeader -> Bool

prop_block_hash :: BlockHeader -> Bool

prop_block_genesis_parent :: BlockHeader -> Bool

prop_block_genesis_target :: BlockHeader -> Bool

prop_block_target

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool

Inductive BlockHeader Properties

prop_block_epoch

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool

prop_block_height

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool

prop_block_weight

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool

prop_block_chainwebVersion

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool

prop_block_chainId

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool

prop_block_creationTime

Arguments

:: BlockHeader
parent header of the checked block header

-> BlockHeader
the block header that is checked

-> Bool