Chainweb.BlockHeaderDB

Description

TODO

Chain Database Handle

data Configuration

Configuration of the chain DB.

Constructors

Configuration

Fields

_configRoot :: !BlockHeader

_configRocksDb :: !RocksDb

data BlockHeaderDb

initBlockHeaderDb :: Configuration -> IO BlockHeaderDb

Initialize a database handle

closeBlockHeaderDb :: BlockHeaderDb -> IO ()

Close a database handle and release all resources

withBlockHeaderDb :: RocksDb -> ChainwebVersion -> ChainId -> (BlockHeaderDb -> IO b) -> IO b

pruneForks

Arguments

:: Logger logger

=> logger

-> BlockHeaderDb

-> Natural
The depth at which the roots are collected for marking block headers that are kept. Any fork that was active that the given depth is kept.

The depth is computed based on the entry of maximum rank. This entry is not necessarily included in the overall consensus of the chainweb. In particular the the higest ranking entry is not necessarily the entry of largest POW weight.

Usually this number would be defined in terms of the chainweb diameter and/or the difficulty adjustment window.

-> (BlockHeader -> Bool -> IO ())
Deletion call back. This hook is called after the entry is deleted from the database. It's main purpose is to delete any resources that were related to the deleted header and that are not needed any more. The Boolean argument indicates whether the payload of the block is shared with any block header that isn't marked for deletion.

-> IO Int

Prunes most block headers and block payloads from forks that are older than the given number of blocks.

This function guarantees that the predecessors of all remaining nodes also remain in the database and that there are are no tangling references.

This function doesn't guarantee to delete all blocks on forks. A small number of fork blocks may not get deleted.

The function takes a callback that is invoked on each deleted block header. The callback takes a parameter that indicates whether the block payload hash is shared with any non-deleted block header. There is a small rate of false positives of block payload hashes that are marked in use.

This doesn't update the the cut db or the payload db.

seekTreeDb :: BlockHeaderDb -> Maybe (NextItem BlockHash) -> Maybe MinRank -> IO (RocksDbTableIter RankedBlockHash RankedBlockHeader)

If k is not Nothing, seekTreeDb d k mir seeks key k in db. If the key doesn't exist it throws TreeDbKeyNotFound. Otherwise if k was exclusive, it advance the iterator to the next item. If minr is not Nothing and the iterator is at a position smaller than minr an invalid iterator is returned. Otherwise the iterator is returned.

If k is Nothing and minr is not Nothing it seeks up to minr and returns the iterator.

If both k and minr are Nothing it returns an iterator that points to the first entry in d.

Orphan instances

TreeDb BlockHeaderDb
TODO provide more efficient branchEntries implementation that uses iterators.

Instance details

Associated Types

type DbEntry BlockHeaderDb

Methods

lookup :: BlockHeaderDb -> DbKey BlockHeaderDb -> IO (Maybe (DbEntry BlockHeaderDb))

keys :: BlockHeaderDb -> Maybe (NextItem (DbKey BlockHeaderDb)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> (Stream (Of (DbKey BlockHeaderDb)) IO (Natural, Eos) -> IO a) -> IO a

entries :: BlockHeaderDb -> Maybe (NextItem (DbKey BlockHeaderDb)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> (Stream (Of (DbEntry BlockHeaderDb)) IO (Natural, Eos) -> IO a) -> IO a

branchKeys :: BlockHeaderDb -> Maybe (NextItem (DbKey BlockHeaderDb)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> HashSet (LowerBound (DbKey BlockHeaderDb)) -> HashSet (UpperBound (DbKey BlockHeaderDb)) -> (Stream (Of (DbKey BlockHeaderDb)) IO (Natural, Eos) -> IO a) -> IO a

branchEntries :: BlockHeaderDb -> Maybe (NextItem (DbKey BlockHeaderDb)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> HashSet (LowerBound (DbKey BlockHeaderDb)) -> HashSet (UpperBound (DbKey BlockHeaderDb)) -> (Stream (Of (DbEntry BlockHeaderDb)) IO (Natural, Eos) -> IO a) -> IO a

insertStream :: BlockHeaderDb -> Stream (Of (DbEntry BlockHeaderDb)) IO a -> IO a

insert :: BlockHeaderDb -> DbEntry BlockHeaderDb -> IO ()

maxEntry :: BlockHeaderDb -> IO (DbEntry BlockHeaderDb)

maxRank :: BlockHeaderDb -> IO Natural