Chainweb.BlockHeaderDB.RestAPI

Description

Collections of hashes or block headers are returned in an order that is compatible with the order of the block heights of the respective block headers. This allows guarantees that all dependencies of a block header are returned before the block header itself is returned.

Block hashes and block headers are base64-URL encoded.

If a filter parameter is given the result is the intersection of the headers that are defined by the filter and the headers from the rest of the query.

The branch parameter specifies a sequence of block headers such that each header is the parent of the succeeding header in the sequence. The last header of the sequence is the second parameter value. The first header in the sequence is the largest header that is smaller or equal than both parameter values.

All functions return 404 if provided with a block hash argument or parameter that doesn't exist in the database.

There are default values for minheight and page limit applied to some queries minrank. These can be discovered through GET chainweb/ApiVersion/InstanceId/api

API types

type BlockHashPage = Page (NextItem BlockHash) BlockHash

A page of BlockHashes

type BlockHeaderPage = Page (NextItem BlockHash) BlockHeader

A page of BlockHeaders

Encodings

data JsonBlockHeaderObject

The default JSON instance of BlockHeader is an unpadded base64Url encoding of the block header bytes. There a newtype wrapper that provides an alternative encoding with as an JSON object.

The mime type application/json;blockheader-encoding=object is used in the HTTP accept header and HTTP content-type header to indicate that block headers are encoded as JSON objects

BlockHeaderDb with typelevel ChainId and ChainwebVersion parameters

newtype BlockHeaderDb_ (v :: ChainwebVersionT) (c :: ChainIdT)

Constructors

BlockHeaderDb_ BlockHeaderDb

data SomeBlockHeaderDb

Constructors

forall v c.(KnownChainwebVersionSymbol v, KnownChainIdSymbol c) => SomeBlockHeaderDb (BlockHeaderDb_ v c)

someBlockHeaderDbVal :: ChainwebVersion -> ChainId -> BlockHeaderDb -> SomeBlockHeaderDb

BlockHeaderDb API

type BlockHeaderDbApi v c = HashesApi v c :<|> (HeadersApi v c :<|> (HeaderApi v c :<|> (HeaderPutApi v c :<|> (BranchHashesApi v c :<|> BranchHeadersApi v c))))

BlockHeaderDb Api

blockHeaderDbApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (BlockHeaderDbApi v c)

Multichain APIs

someBlockHeaderDbApi :: ChainwebVersion -> ChainId -> SomeApi

someBlockHeaderDbApis :: ChainwebVersion -> [ChainId] -> SomeApi

BlockHeader Event Stream

newtype HeaderStream

Constructors

HeaderStream Bool

data HeaderUpdate

Constructors

HeaderUpdate

Fields

_huHeader :: !(ObjectEncoded BlockHeader)

_huTxCount :: !Int

_huPowHash :: !Text

_huTarget :: !Text

type HeaderStreamApi (v :: ChainwebVersionT) = 'ChainwebEndpoint v :> HeaderStreamApi_

A stream of all new BlockHeaders that are accepted into the true Cut.

headerStreamApi :: forall (v :: ChainwebVersionT). Proxy (HeaderStreamApi v)

someHeaderStreamApi :: ChainwebVersion -> SomeApi

Sub APIs

type BranchHashesApi (v :: ChainwebVersionT) (c :: ChainIdT) = 'ChainwebEndpoint v :> (ChainEndpoint c :> Reassoc BranchHashesApi_)

GET /chainweb/<ApiVersion>/<InstanceId>/chain/<ChainId>/hash/branch

Returns a set of branches. A branch is obtained by traversing the block header tree database starting at some entry along the parent relation, i.e. in direction towards the root or decending order with respect to the children relation.

If a lower bound for the traversal is given, then no node is returned that is in a branch (is equal to, parent, or grantparent) of the lower bound. This means the query stops at the fork point of the upper bound and the lower bound and returns all nodes between from the upper bound down to fork point. The fork point itself isn't included.

Simultaneously traversing more than a single branch results in a tree that is a sub-graph, but not necessarily a sub-tree, of the database tree. Search stops at the first matching lower bound.

branchHashesApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (BranchHashesApi v c)

type BranchHeadersApi (v :: ChainwebVersionT) (c :: ChainIdT) = 'ChainwebEndpoint v :> (ChainEndpoint c :> Reassoc BranchHeadersApi_)

GET /chainweb/<ApiVersion>/<InstanceId>/chain/<ChainId>/header/branch

Returns a set of branches. A branch is obtained by traversing the block header tree database starting at some entry along the parent relation, i.e. in direction towards the root or decending order with respect to the children relation.

If a lower bound for the traversal is given, then no node is returned that is in a branch (is equal to, parent, or grantparent) of the lower bound. This means the query stops at the fork point of the upper bound and the lower bound and returns all nodes between from the upper bound down to fork point. The fork point itself isn't included.

Simultaneously traversing more than a single branch results in a tree that is a sub-graph, but not necessarily a sub-tree, of the database tree. Search stops at the first matching lower bound.

branchHeadersApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (BranchHeadersApi v c)

type HeaderApi (v :: ChainwebVersionT) (c :: ChainIdT) = 'ChainwebEndpoint v :> (ChainEndpoint c :> HeaderApi_)

GET /chainweb/<ApiVersion>/<InstanceId>/chain/<ChainId>/header/<BlockHash>

Returns a single block headers for a given block hash.

headerApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (HeaderApi v c)

type HeaderPutApi (v :: ChainwebVersionT) (c :: ChainIdT) = 'ChainwebEndpoint v :> (ChainEndpoint c :> HeaderPutApi_)

PUT /chainweb/<ApiVersion>/<InstanceId>/chain/<ChainId>/header

Adds a block header to the block header tree database. Returns a failure with status code 400 if the block header can't be addded because of a validation failure or missing dependencies.

headerPutApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (HeaderPutApi v c)

type HeadersApi (v :: ChainwebVersionT) (c :: ChainIdT) = 'ChainwebEndpoint v :> (ChainEndpoint c :> Reassoc HeadersApi_)

GET /chainweb/<ApiVersion>/<InstanceId>/chain/<ChainId>/header

Returns block headers in the block header tree database in ascending order with respect to the children relation.

Note that for block headers on different branches, the order isn't determined. Therefore a block header of higher block height can be returned before a block header of lower block height.

headersApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (HeadersApi v c)

type HashesApi (v :: ChainwebVersionT) (c :: ChainIdT) = 'ChainwebEndpoint v :> (ChainEndpoint c :> Reassoc HashesApi_)

GET /chainweb/<ApiVersion>/<InstanceId>/chain/<ChainId>/hash

Returns hashes in the block header tree database in ascending order with respect to the children relation.

Note that for block hashes on different branches, the order isn't determined. Therefore a block hash of higher block height can be returned before a block hash of lower block height.

hashesApi :: forall (v :: ChainwebVersionT) (c :: ChainIdT). Proxy (HashesApi v c)

Orphan instances

MimeRender OctetStream BlockHeader
Orphan instance to encode BlockHeaders as OctetStream

Instance details

Methods

mimeRender :: Proxy OctetStream -> BlockHeader -> ByteString

MimeUnrender OctetStream BlockHeader
Orphan instance to encode BlockHeaders as OctetStream

Instance details

Methods

mimeUnrender :: Proxy OctetStream -> ByteString -> Either String BlockHeader

mimeUnrenderWithType :: Proxy OctetStream -> MediaType -> ByteString -> Either String BlockHeader