Chainweb.BlockHeaderDB.RestAPI.Client

Description

Documentation

headerClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => DbKey BlockHeaderDb -> ClientM BlockHeader

headerClient :: ChainwebVersion -> ChainId -> DbKey BlockHeaderDb -> ClientM BlockHeader

headerClientContentType_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT) (ct :: Type) x. KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Accept ct => SupportedRespBodyContentType ct x BlockHeader => HeaderApi v c ~ x => BlockHash -> ClientM BlockHeader

headerClientJson :: ChainwebVersion -> ChainId -> DbKey BlockHeaderDb -> ClientM BlockHeader

headerClientJsonPretty :: ChainwebVersion -> ChainId -> BlockHash -> ClientM BlockHeader

headerClientJsonBinary :: ChainwebVersion -> ChainId -> BlockHash -> ClientM BlockHeader

headerPutClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => DbEntry BlockHeaderDb -> ClientM NoContent

headerPutClient :: ChainwebVersion -> ChainId -> DbEntry BlockHeaderDb -> ClientM NoContent

headerPutClientContentType_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT) (ct :: Type). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Accept ct => SupportedReqBodyContentType ct (HeaderPutApi v c) BlockHeader => DbEntry BlockHeaderDb -> ClientM NoContent

headerPutClientJson :: ChainwebVersion -> ChainId -> DbEntry BlockHeaderDb -> ClientM NoContent

headerPutClientJsonPretty :: ChainwebVersion -> ChainId -> DbEntry BlockHeaderDb -> ClientM NoContent

headerPutClientBinary :: ChainwebVersion -> ChainId -> DbEntry BlockHeaderDb -> ClientM NoContent

hashesClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> ClientM BlockHashPage

hashesClient :: ChainwebVersion -> ChainId -> Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> ClientM BlockHashPage

headersClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> ClientM BlockHeaderPage

headersClient

Arguments

:: ChainwebVersion
The remote chainweb that you wish to query from.

-> ChainId
The remote chain within the web that you wish to query from.

-> Maybe Limit
The number of responses per-Page to return.

-> Maybe (NextItem BlockHash)
The first header you want to see within the Page. Page contains a field _pageNext, which can be used to produce the value needed for subsequent calls.

-> Maybe MinRank
Filter: no header of BlockHeight lower than this will be returned.

-> Maybe MaxRank
Filter: no header of BlockHeight higher than this will be returned.

-> ClientM BlockHeaderPage

headersClientContentType_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT) (ct :: Type) x. KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Accept ct => HeadersApi v c ~ x => SupportedRespBodyContentType ct x BlockHeaderPage => Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> ClientM BlockHeaderPage

headersClientJson

Arguments

:: ChainwebVersion
The remote chainweb that you wish to query from.

-> ChainId
The remote chain within the web that you wish to query from.

-> Maybe Limit
The number of responses per-Page to return.

-> Maybe (NextItem BlockHash)
The first header you want to see within the Page. Page contains a field _pageNext, which can be used to produce the value needed for subsequent calls.

-> Maybe MinRank
Filter: no header of BlockHeight lower than this will be returned.

-> Maybe MaxRank
Filter: no header of BlockHeight higher than this will be returned.

-> ClientM BlockHeaderPage

headersClientJsonPretty

Arguments

:: ChainwebVersion
The remote chainweb that you wish to query from.

-> ChainId
The remote chain within the web that you wish to query from.

-> Maybe Limit
The number of responses per-Page to return.

-> Maybe (NextItem BlockHash)
The first header you want to see within the Page. Page contains a field _pageNext, which can be used to produce the value needed for subsequent calls.

-> Maybe MinRank
Filter: no header of BlockHeight lower than this will be returned.

-> Maybe MaxRank
Filter: no header of BlockHeight higher than this will be returned.

-> ClientM BlockHeaderPage

branchHashesClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHashPage

branchHashesClient :: ChainwebVersion -> ChainId -> Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHashPage

branchHeadersClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHeaderPage

branchHeadersClient :: ChainwebVersion -> ChainId -> Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHeaderPage

branchHeadersClientContentType_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT) (ct :: Type) api. KnownChainwebVersionSymbol v => KnownChainIdSymbol c => HeadersApi v c ~ api => SupportedRespBodyContentType ct api BlockHeaderPage => Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHeaderPage

branchHeadersClientJson :: ChainwebVersion -> ChainId -> Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHeaderPage

branchHeadersClientJsonPretty :: ChainwebVersion -> ChainId -> Maybe Limit -> Maybe (NextItem BlockHash) -> Maybe MinRank -> Maybe MaxRank -> BranchBounds BlockHeaderDb -> ClientM BlockHeaderPage