Chainweb.BlockHeaderDB.Types

Documentation

data BlockHeaderDb

Constructors

BlockHeaderDb

Fields

_chainDbId :: !ChainId

_chainDbChainwebVersion :: !ChainwebVersion

_chainDbCas :: !(RocksDbTable RankedBlockHash RankedBlockHeader)

Ranked block hashes provide fast access and iterating by block height. Blocks of similar height are stored and cached closely together. This table is an instance of IsCas.

_chainDbRankTable :: !(RocksDbTable BlockHash BlockHeight)

This index supports lookup of a block hash for which the height isn't known

newtype RankedBlockHeader

Constructors

RankedBlockHeader

Fields

_getRankedBlockHeader :: BlockHeader

data RankedBlockHash

Constructors

RankedBlockHash

Fields

_rankedBlockHashHeight :: !BlockHeight

_rankedBlockHash :: !BlockHash

newtype BlockRank

Constructors

BlockRank

Fields

_getBlockRank :: BlockHeight