Chainweb.ChainId

Description

The defininitions in this module are also exported via Chainweb.Version.

Documentation

data ChainIdException

Constructors

ChainIdMismatch (Expected ChainId) (Actual ChainId)

data ChainId :: Type

ChainId within the context of a Chainweb instance.

The set of valid ChainIds is determined by the ChainwebVersion. In almost all use cases there should be a context that is an instance of HasChainwebVersion can be used get the set of chain ids.

In the context of a particular chain the respective ChainId can be obtained via instances of HasChainId.

How to create values of type ChainId

To fold or traverse over all chain ids, use chainIds.

To deserialize a chain id, use mkChainId.

For a random chain id consider using randomChainId.

For some arbitrary but fixed chain id consider using someChainId.

class HasChainId a where

Minimal complete definition

_chainId | chainId

Methods

_chainId :: a -> ChainId

chainId :: Getter a ChainId

checkChainId :: MonadThrow m => HasChainId expected => HasChainId actual => Expected expected -> Actual actual -> m ChainId

chainIdToText :: ChainId -> Text

chainIdFromText :: MonadThrow m => Text -> m ChainId

Serialization

encodeChainId :: MonadPut m => ChainId -> m ()

decodeChainId :: MonadGet m => m ChainId

decodeChainIdChecked :: MonadGet m => MonadThrow m => HasChainId p => Expected p -> m ChainId

Typelevel ChainID

newtype ChainIdT

Constructors

ChainIdT Symbol

type family ChainIdSymbol n :: Symbol

chainIdSymbolVal :: KnownChainIdSymbol n => Proxy n -> Text

data SomeChainIdT

Constructors

forall (a :: ChainIdT).KnownChainIdSymbol a => SomeChainIdT (Proxy a)

class KnownSymbol (ChainIdSymbol n) => KnownChainIdSymbol (n :: ChainIdT)

Minimal complete definition

chainIdSymbolVal

someChainIdVal :: ChainId -> SomeChainIdT

Singletons

data family Sing :: k -> Type

Kind-indexed family of singleton types.

In order to define singletons for types of a data kind A, one defines an instance that has a value constructor for each type constructor of A. The following is an example how to define singletons for types of kind Maybe a:

data instance Sing :: forall a . Maybe a -> Type where
   SNothing :: Sing 'Nothing
   SJust :: Sing a -> Sing ('Just a)

type SChainId (n :: ChainIdT) = Sing n

Testing

unsafeChainId :: Word32 -> ChainId

This function should be be rarely needed. Please consult the documentation of ChainId for alternative ways to obtain ChainId values.

chainIdInt :: Integral i => ChainId -> i