Chainweb.Chainweb

Description

This module provides the tools to initialize the different components of the chainweb consensus and run a chainweb-consensus node.

The overall pattern is that for each component C there is

a function withC that manages the resources of the component and passes a handle that encapsulates those resources to an inner computation, and

a function runC that takes the resource handle of C and starts the component.

This pattern allows to bootstrap components with mutual dependencies and resolve timing and ordering constraints for the startup of services.

Logging functions are initialized in the enviroment and passed as call backs to the components.

Pact Configuration

data TransactionIndexConfig

Constructors

TransactionIndexConfig

defaultTransactionIndexConfig :: TransactionIndexConfig

pTransactionIndexConfig :: MParser TransactionIndexConfig

Configuration

data ChainwebConfiguration

Constructors

ChainwebConfiguration

Fields

_configChainwebVersion :: !ChainwebVersion

_configNodeId :: !NodeId

_configMiner :: !(EnableConfig MinerConfig)

_configCoordinator :: !Bool

_configHeaderStream :: !Bool

_configReintroTxs :: !Bool

_configP2p :: !P2pConfiguration

_configTransactionIndex :: !(EnableConfig TransactionIndexConfig)

_configIncludeOrigin :: !Bool

_configThrottleRate :: !Natural

_configMempoolP2p :: !(EnableConfig MempoolP2pConfig)

_configPruneChainDatabase :: !Bool

_configBlockGasLimit :: !GasLimit

configNodeId :: Lens' ChainwebConfiguration NodeId

configChainwebVersion :: Lens' ChainwebConfiguration ChainwebVersion

configMiner :: Lens' ChainwebConfiguration (EnableConfig MinerConfig)

configCoordinator :: Lens' ChainwebConfiguration Bool

configHeaderStream :: Lens' ChainwebConfiguration Bool

configReintroTxs :: Lens' ChainwebConfiguration Bool

configP2p :: Lens' ChainwebConfiguration P2pConfiguration

configTransactionIndex :: Lens' ChainwebConfiguration (EnableConfig TransactionIndexConfig)

configBlockGasLimit :: Lens' ChainwebConfiguration GasLimit

defaultChainwebConfiguration :: ChainwebVersion -> ChainwebConfiguration

pChainwebConfiguration :: MParser ChainwebConfiguration

validateChainwebConfiguration :: ConfigValidation ChainwebConfiguration l

Chainweb Resources

data Chainweb logger cas

Constructors

Chainweb

Fields

_chainwebHostAddress :: !HostAddress

_chainwebChains :: !(HashMap ChainId (ChainResources logger))

_chainwebCutResources :: !(CutResources logger cas)

_chainwebMiner :: !(Maybe (MinerResources logger cas))

_chainwebCoordinator :: !(Maybe (MiningCoordination logger cas))

_chainwebHeaderStream :: !HeaderStream

_chainwebLogger :: !logger

_chainwebPeer :: !(PeerResources logger)

_chainwebPayloadDb :: !(PayloadDb cas)

_chainwebManager :: !Manager

_chainwebPactData :: [(ChainId, PactServerData logger cas)]

_chainwebThrottler :: !(Throttle Address)

_chainwebConfig :: !ChainwebConfiguration

chainwebChains :: forall logger cas. Lens' (Chainweb logger cas) (HashMap ChainId (ChainResources logger))

chainwebCutResources :: forall logger cas. Lens' (Chainweb logger cas) (CutResources logger cas)

chainwebHostAddress :: forall logger cas. Lens' (Chainweb logger cas) HostAddress

chainwebMiner :: forall logger cas. Lens' (Chainweb logger cas) (Maybe (MinerResources logger cas))

chainwebCoordinator :: forall logger cas. Lens' (Chainweb logger cas) (Maybe (MiningCoordination logger cas))

chainwebHeaderStream :: forall logger cas. Lens' (Chainweb logger cas) HeaderStream

chainwebLogger :: forall logger cas. Lens' (Chainweb logger cas) logger

chainwebSocket :: Getter (Chainweb logger cas) Socket

chainwebPeer :: forall logger cas. Lens' (Chainweb logger cas) (PeerResources logger)

chainwebPayloadDb :: forall logger cas. Lens' (Chainweb logger cas) (PayloadDb cas)

chainwebPactData :: forall logger cas. Lens' (Chainweb logger cas) [(ChainId, PactServerData logger cas)]

chainwebThrottler :: forall logger cas. Lens' (Chainweb logger cas) (Throttle Address)

chainwebConfig :: forall logger cas. Lens' (Chainweb logger cas) ChainwebConfiguration

Mempool integration

type ChainwebTransaction = Command PayloadWithText

chainwebTransactionConfig :: TransactionConfig ChainwebTransaction

validatingMempoolConfig :: ChainId -> ChainwebVersion -> GasLimit -> MVar PactExecutionService -> InMemConfig ChainwebTransaction

withChainweb :: Logger logger => ChainwebConfiguration -> logger -> RocksDb -> Maybe FilePath -> Bool -> (Chainweb logger RocksDbCas -> IO a) -> IO a

runChainweb :: forall logger cas. Logger logger => PayloadCas cas => Chainweb logger cas -> IO ()

Starts server and runs all network clients

Miner

runMiner :: forall logger cas. Logger logger => PayloadCas cas => ChainwebVersion -> MinerResources logger cas -> IO ()