Chainweb.Chainweb.PeerResources

Description

TODO

Documentation

data PeerResources logger

Constructors

PeerResources

Fields

_peerResConfig :: !P2pConfiguration

configuration of this peer

_peerResPeer :: !Peer

local peer of this chainweb node

_peerResSocket :: !Socket

a socket that is used by the server of this peer

_peerResDb :: !PeerDb

the peer db

_peerResManager :: !Manager

the connection manager

_peerLogger :: !logger

peerResConfig :: forall logger. Lens' (PeerResources logger) P2pConfiguration

peerResSocket :: forall logger. Lens' (PeerResources logger) Socket

peerResPeer :: forall logger. Lens' (PeerResources logger) Peer

peerResDb :: forall logger. Lens' (PeerResources logger) PeerDb

peerResManager :: forall logger. Lens' (PeerResources logger) Manager

peerServerSettings :: Peer -> Settings

peerLogger :: forall logger logger. Lens (PeerResources logger) (PeerResources logger) logger logger

Allocate Peer Resources

withPeerResources :: Logger logger => ChainwebVersion -> P2pConfiguration -> logger -> (logger -> PeerResources logger -> IO a) -> IO a

Allocate Peer resources. All P2P networks of a chainweb node share a single Peer and the associated underlying network resources.

Additionally, the continuation is provided with a logger the records the peer info in each log message.

The following resources are allocated:

Resolve port and allocating a socket for the port,

Generate a new certifcate, if none is provided in the configuration, and

adjust the P2PConfig with the new values.

Internal Utils

withSocket :: P2pConfiguration -> ((P2pConfiguration, Socket) -> IO a) -> IO a

withPeerDb :: HashSet NetworkId -> P2pConfiguration -> (PeerDb -> IO a) -> IO a

Run a computation with a PeerDb

withConnectionManger :: Logger logger => logger -> X509CertChainPem -> X509KeyPem -> PeerDb -> (Manager -> IO a) -> IO a