Chainweb.Counter

Description

TODO

Documentation

data Counter (s :: Symbol)

newCounter :: IO (Counter s)

inc :: Counter s -> IO ()

incBy :: Integral a => Counter s -> a -> IO ()

Counter Map

data CounterMap (s :: Symbol)

newCounterMap :: IO (CounterMap s)

incKey :: CounterMap s -> Text -> IO ()

incKeyBy :: Integral a => CounterMap s -> Text -> a -> IO ()

Rolling of Counters

data CounterValue

roll :: IsCounter c => c -> IO CounterValue

Logging of Counters

data CounterLog

The Counters are rolled at the time the message is logged (not when it is processed by the backend). This adds a small amount of overhead, but ensures that the timestamp of the log matches the time when the logs are rolled. This means that logs are rolled even when the message ends up being discarded by the backend, which means that counts are lost along with the discarded messages. This is consistent with the semantics of discarding log messages.

type LogFunctionCounter = forall f. Foldable f => LogLevel -> f CounterValue -> IO ()

logFunctionCounter :: Logger l => l -> LogFunctionCounter