Chainweb.Crypto.MerkleLog

Description

This module provides a framework for hashing structured data and creating inclusion proofs for hashed data.

An example of how to use this module can be found in the file .docsmerklelog-example.md. The impatient reader is encouraged to skip right to the example.

Background: The Problem of Hashing Structured Data

When authenticating structured data it is common practice to create hashes by using some binary encoding of the whole structure or of the individual components. In the latter case, the serialized components are usually concatenated and hashed. This ad-hoc approach to hashing binary data is error prone and makes API specifications more complicated.

Moreover, the resulting hashes can't be used for compact self-contained inclusion proofs. Usually a lot of unrelated additional data has to be included in the proof to make it self-contained. Another way is to define a separate structure as input for a Merkle tree which requires to maintain an additional index for the roots of the trees and possibly also for querying the relevant input in the right order.

Finally, authenticated values in block chains and values for content- addressed key value stores in general, often include the key in the data structure, which creates a cyclic dependency of a value on itself. When creating such a structure this is usually dealt with by lazily tying a knot, instantiating the key with a default value (e.g. Nothing) and replacing it later with the actual key (e.g. 'Just key'), or by defining two functions for creating the hash and another for creating the structure that take the same data as input plus the hash. Often those functions take a large number of parameters and it can be hard to keep both in sync when maintaining the code. It is easy to add a parameter to the structure, but forget to update the hash computation to include the new value, which is difficult to detect in tests and can lead to flawed security.

What this Module Offers

This module defines types and idioms for a unified way to compute hashes on structured data. This makes it easier to maintain the consistency of the hash computations and generates hashes in a way that supports compact Merkle proofs. The module introduces a syntax for defining the structure of hashes that makes it obvious how a hash is created and helps to to make sure that the code complies with the specification.

The tools in the module also automate some of the tedious and error prone work of creating and verifying inclusion proofs. The precise specification of the construction of these proofs is key for a block chain application, and this module should help to ensure that the specification is uniform across all proofs.

Merkle Log

A merkle log represents a view of a data structure that

defines how a hash is computed and

supports efficient creation of compact, self-contained inclusion proofs.

Entries in a merkle log must be

serializable to a binary representation and

arranged in a arbitrary but fixed order.

The types of entries that can be included as leaves of the same MerkleTree must live in a closed universe. Each universe allows fixes the hash algorithm that is used for all Merkle trees over the universe. Using a closed universe prevents attacks that are based on using hashes of data from different domains with equal representation.

In our implementation we assume that each value for which a Merkle Log is created can be represented as a short, finite length header and a body that is a sequence of values of the same type. The HasMerkleLog class provides the representation of a type as a Merkle log and the functions to transform a value to and from that representation.

Documentation

Merkle Tree Inputs

A framework for computing hashes from structures in a well-defined way.

A structure that can be hashed is called a log. The elements of the structure that are inputs to the hash are called log entries. In order to use a type as a log one has to define two things:

For each log entry a serialization and deserialization method.

A decomposition of a log type into an ordered sequence of log entries.

A Merkle Universe is created as follows:

Define a data kind with a nullary type constructor for each type in the Merkle universe.

Define an instance for MerkleUniverse for the new kind and assign each type constructor in the universe a typelevel Nat value that represents a Word16 value.

A log entry is defined as follows:

Consider creating a specific newtype wrapper for the entry type, in particular, if the entry is of a generic type, like, for instance, Int or ByteString.

Define an IsMerkleLogEntry instance or derive it using the 'deriving via' extension if available. For the Tag associated type family pick the value from the Merkle universe type that corresponds to the entry type.

A log type is defines as follows:

Define all constructor fields as log entries.

Define a HasMerkleLogEntry instance for the type.

class HashAlgorithm (HashAlg k) => MerkleUniverse k

A Kind that represents a closed universe of types that can be included as leaves of the same MerkleTree.

The MerkleLogHash type family defines the hash function that is used for Merkle trees within the universe.

The MerkleTagVal type family is used to assing each type-constructor in the universe a type-level Nat that represents a Word16 value.

Associated Types

type HashAlg k :: Type

type MerkleTagVal k (a :: k) :: Nat

tagVal :: forall u (t :: u). KnownNat (MerkleTagVal u t) => Word16

Term level representation of the MerkleTagVal of a type in a Merkle universe.

Merkle Log Entries

class InUniverse u (Tag b) => IsMerkleLogEntry u b | b -> u where

Class of types that can be used as entries in a merkle tree.

The Tag associated type family tags each type that is an instance of IsMerkleLogEntry with a type from the respective Merkle universe.

The functions of the type class specify whether the entry corresponds to the root of a nested Merkle tree or corresponds to an input node.

Associated Types

type Tag b :: u

Methods

toMerkleNode :: b -> MerkleNodeType (HashAlg u) ByteString

fromMerkleNode :: MerkleNodeType (HashAlg u) ByteString -> Either SomeException b

data MerkleLogEntries :: Type -> [Type] -> Type -> Type where

A data type that represents the Merkle log for a structure as

a fixed size polymorphic list of header values and

a monomorphic sequence of body values.

Both the header and the body may possibly be empty. The type of the former is represented by an empty type-level list, while the latter is represented by Void.

Constructors

MerkleLogBody :: IsMerkleLogEntry u b => Vector b -> MerkleLogEntries u '[] b

(:+:) :: IsMerkleLogEntry u h => h -> MerkleLogEntries u t b -> MerkleLogEntries u (h ': t) b infixr 5

emptyBody :: IsMerkleLogEntry u b => MerkleLogEntries u '[] b

mapLogEntries :: forall u h s b. (forall x. IsMerkleLogEntry u x => x -> b) -> MerkleLogEntries u h s -> Vector b

entriesHeaderSize :: MerkleLogEntries u l s -> Int

entriesBody :: MerkleLogEntries u l s -> Vector s

Merkle Log

data MerkleLog u (h :: [Type]) (b :: Type)

A merkle log represents values of IsMerkleLog types in a generic way that supports computing the root hash and a merkle tree.

Constructors

MerkleLog

Fields

_merkleLogRoot :: !(MerkleRoot (HashAlg u))

The root hash of the Merkle tree of the log.

_merkleLogEntries :: !(MerkleLogEntries u h b)

The entries of the Merkle log.

Note: For creating proofs this isn't needed. Should we make this lazy, too? For large entries it may be somewhat costly to pull it from the store.

_merkleLogTree :: MerkleTree (HashAlg u)

The merkle tree for the entries of the log. It is required to compute the root and for creating a merkle proofs.

For some types, computing the merkle tree can be expensive. Therefore it is instantiated lazily and only forced if actually required.

class (MerkleUniverse u, HashAlgorithm (HashAlg u)) => HasMerkleLog u b | b -> u where

Class of types which can be represented as a Merkle tree log.

An instance of HasMerkleLog can be encoded as

a header that consists of a fixed size polymorphic list IsMerkleEntry instances, and

a body that consists of a monomorphic sequence of IsMerkleEntry instances.

Associated Types

type MerkleLogHeader b :: [Type]

type MerkleLogBody b :: Type

Methods

toLog

Arguments

:: b

-> MerkleLog u (MerkleLogHeader b) (MerkleLogBody b)
Transform a value into a Merkle log.

Often the root of the Merkle tree is used as identifier and is stored as part of the input structure. In those cases the function merkleLog can be used to create the MerkleLog from the root and the entries without forcing the computation of the Merkle tree field.

fromLog

Arguments

:: MerkleLog u (MerkleLogHeader b) (MerkleLogBody b)

-> b
Recover a value from a Merkle log.

type MkLogType u b = MerkleLog u (MerkleLogHeader b) (MerkleLogBody b)

merkleLog :: forall u h b. HashAlgorithm (HashAlg u) => MerkleRoot (HashAlg u) -> MerkleLogEntries u h b -> MerkleLog u h b

Create a MerkleLog from a MerkleRoot and a sequence of MerkleLogEntrys. The _merkleLogTree fields is instantiated lazily.

newMerkleLog :: forall u h b. MerkleUniverse u => MerkleLogEntries u h b -> MerkleLog u h b

IsMerkleLog values often include a hash of the value itself, which represents cyclic dependency of a value on itself. This function allows to create such an value from its representation as a sequence of merkle log entries.

class HasHeader c b where

Internal: Get (first) header entry of given type from a Merkle log.

TODO: * check that the header value is unique (requires traversing the list until then end after a value is found)

Associated Types

type HeaderPos c b :: Nat

Methods

header :: b -> c

headerPos :: Int

body :: MerkleLog a l s -> Vector s

Get the body sequence of a Merkle log.

headerSize :: MerkleLog a l s -> Int

Get the number of entries in the header of a Merkle log.

bodySize :: MerkleLog a l s -> Int

Get the number of entries in the body of a Merkle log.

computeMerkleLogRoot :: forall u b. HasMerkleLog u b => b -> MerkleRoot (HashAlg u)

Compute the Merkle root hash for an instance of HasMerkleLog.

merkleLogHash b computes the merkle tree, which is linear in the size of the b. For large logs the hash or the full MerkleTree should e cached.

Merkle Log Proofs

headerProof :: forall c u b m. MonadThrow m => HasHeader c (MkLogType u b) => IsMerkleLogEntry u c => HasMerkleLog u b => b -> m (MerkleProof (HashAlg u))

Create an inclusion proof for a value in the header of a Merkle log.

NOTE: The call to toLog can be potentially expensive if the b body of the log is large and the tree isn't memoized. However, for most applications, either the header is empty or the body is small and recomputing it is cheap.

We could consider placing the header at the end of the tree. In that case we could cache (all but at most logarithmic many hashes) of the body tree and recompute just the part of the tree that depends on the header. However, for large trees, we store a full cached version of the tree in any case, so we should use that instead. Another option would be to use two separate trees, but, again, our current use case wouldn't justify the overhead.

headerTree :: forall c u b. HasHeader c (MkLogType u b) => IsMerkleLogEntry u c => HasMerkleLog u b => b -> (MerkleNodeType (HashAlg u) ByteString, Int, MerkleTree (HashAlg u))

Create the parameters for creating nested inclusion proofs with the merkleProof_ of the merkle-log package.

This function returns the proof subject, the position of the subject, and the Merkle tree and should be used for the leaf tree in the nested proof.

headerTree_ :: forall c u b. HasHeader c (MkLogType u b) => HasMerkleLog u b => b -> (Int, MerkleTree (HashAlg u))

Create the parameters for creating nested inclusion proofs with the merkleProof_ of the merkle-log package.

This function returns the position of the input and the Merkle tree, but not the subject. It should be used for inner trees in the nested proof.

bodyProof

Arguments

:: forall u b m. MonadThrow m

=> HasMerkleLog u b

=> b

-> Int
the index in the body of the log

-> m (MerkleProof (HashAlg u))

Create an inclusion proof for a value in the body of a Merkle log.

TODO: it is not clear if the result of toLog (which contains the cached tree) is memoized on the heap. (i.e. depends on where b is coming from). We should either make sure that it is memoized or provide a version that takes a MerkleLog value.

bodyTree

Arguments

:: forall u b. HasMerkleLog u b

=> b

-> Int
the index in the body of the log

-> (MerkleNodeType (HashAlg u) ByteString, Int, MerkleTree (HashAlg u))

Create the parameters for creating nested inclusion proofs with the merkleProof_ of the merkle-log package.

This function returns the proof subject, the position of the subject, and the Merkle tree and should be used for the leaf tree in the nested proof.

bodyTree_

Arguments

:: forall u b. HasMerkleLog u b

=> b

-> Int
the index in the body of the log

-> (Int, MerkleTree (HashAlg u))

Create the parameters for creating nested inclusion proofs with the merkleProof_ of the merkle-log package.

This function returns the position of the input and the Merkle tree, but not the subject. It should be used for inner trees in the nested proof.

proofSubject :: forall u b m. MonadThrow m => IsMerkleLogEntry u b => MerkleProof (HashAlg u) -> m b

Extract the proof subject from a MerkleProof value.

Utils for Defining Instances

decodeMerkleInputNode :: MonadThrow m => (forall n. MonadGet n => n b) -> MerkleNodeType a ByteString -> m b

encodeMerkleInputNode :: (forall m. MonadPut m => b -> m ()) -> b -> MerkleNodeType a ByteString

decodeMerkleTreeNode :: MonadThrow m => Coercible (MerkleRoot alg) a => MerkleNodeType alg x -> m a

encodeMerkleTreeNode :: Coercible a (MerkleRoot alg) => a -> MerkleNodeType alg x

IsMerkleLogEntry instance for use with deriving via

newtype ByteArrayMerkleLogEntry u (t :: u) b

Support for deriving IsMerkleLogEntry for types that are an instance of ByteArray via the DerivingVia extension.

Constructors

ByteArrayMerkleLogEntry b

newtype MerkleRootLogEntry (t :: u)

Support for deriving IsMerkleLogEntry for types that are newtype wrappers of MerkleRoot via the DerivingVia extension.

Constructors

MerkleRootLogEntry (MerkleRoot (HashAlg u))

newtype Word8MerkleLogEntry (t :: u)

Support for deriving IsMerkleLogEntry for types that are newtype wrappers of Word8 via the DerivingVia extension.

Constructors

Word8MerkleLogEntry

Fields

_getWord8LogEntry :: Word8

newtype Word16BeMerkleLogEntry (t :: u)

Support for deriving IsMerkleLogEntry for types that are newtype wrappers of Word16 via the DerivingVia extension.

Constructors

Word16BeMerkleLogEntry

Fields

_getWord16BeLogEntry :: Word16

newtype Word32BeMerkleLogEntry (t :: u)

Support for deriving IsMerkleLogEntry for types that are newtype wrappers of Word32 via the DerivingVia extension.

Constructors

Word32BeMerkleLogEntry

Fields

_getWord32BeLogEntry :: Word32

newtype Word64BeMerkleLogEntry (t :: u)

Support for deriving IsMerkleLogEntry for types that are newtype wrappers of Word64 via the DerivingVia extension.

Constructors

Word64BeMerkleLogEntry

Fields

_getWord64BeLogEntry :: Word64

Exceptions

data MerkleLogException

Constructors

MerkleLogWrongNodeTypeException (Expected Text) (Actual Text)

MerkleLogWrongTagException (Expected Text) (Actual Text)

MerkleLogDecodeException Text

expectedInputNodeException :: MerkleLogException

expectedTreeNodeException :: MerkleLogException