Chainweb.Cut

Description

TODO

Documentation

data Cut

A cut is a distributed state of a concurrent computation. Formally, a cut satisfies the property

prop_cutBraiding

A cut also satisfies the properties that

the graph is valid,

the graph corresponds to the ChainwebVersion,

the set of ChainIds of the cut are exactly the vertices of the graph,

all blockHeaders are valid (which is guaranteed by genesisBlockHeader) with respect to the graph.

_cutMap :: Cut -> HashMap ChainId BlockHeader

cutMap :: Getter Cut (HashMap ChainId BlockHeader)

_cutHeight :: Cut -> BlockHeight

cutHeight :: Getter Cut BlockHeight

_cutWeight :: Cut -> BlockWeight

cutWeight :: Getter Cut BlockWeight

_cutAdjPairs :: Cut -> HashSet (AdjPair BlockHeader)

All adjacent pairs of a cut

cutAdjPairs :: Getter Cut (HashSet (AdjPair BlockHeader))

cutAdjs :: HasChainId cid => Cut -> cid -> HashMap ChainId BlockHeader

Assumes that the cid exists in the graph.

lookupCutM :: MonadThrow m => HasChainId cid => cid -> Cut -> m BlockHeader

forkDepth :: WebBlockHeaderDb -> Cut -> Cut -> IO Natural

limitCut :: HasCallStack => WebBlockHeaderDb -> BlockHeight -> Cut -> IO Cut

Find a Cut that is a predecessor of the given one, and that has a block height that is smaller or equal the given height.

If the requested limit is larger or equal to the current height, the given cut is returned.

Otherwise, the requested height limit is divided by the number of chains (rounded down) and the result is used such that for each chain the predecessor of the given cut at the respective height is returned.

Exceptions

data CutException

Constructors

InvalidCutPair (AdjPair BlockHeader)

NonMonotonicCutExtension (Expected BlockHash) (Actual BlockHash) BlockHeader

InvalidCutExtension BlockHeader

Genesis Cut

genesisCut :: ChainwebVersion -> Cut

TODO

This guarantees that

the graph is valid,

the graph corresponds to the ChainwebVersion,

the set of ChainIds of the cut are the exactly the vertices of the graph,

all blockHeaders are valid (which is guaranteed by genesisBlockHeader).

These properties are maintained as inductive invariants for all Cuts.

Checks

checkBraidingOfCut :: MonadThrow m => Cut -> m ()

Check that a cut is correctly braided.

This check is only needed for external cuts. Note, that for imported cuts this must be checked recursively. This can be done by doing a join that starts with a meet with a local cut.

checkBraidingOfCutPairs :: MonadThrow m => Foldable f => f (AdjPair BlockHeader) -> m ()

Check that a set of adjacent pairs of a cut is correctly braided.

checkBraidingOfCutPair :: MonadThrow m => AdjPair BlockHeader -> m ()

isBraidingOfCutPair :: MonadThrow m => AdjPair BlockHeader -> m Bool

Returns whether an adjacent pair in a cut is correctly braided.

throws ChainNotAdjacentException

Extending Cuts

isMonotonicCutExtension :: MonadThrow m => Cut -> BlockHeader -> m Bool

Extends a Cut monotonically, i.e. the replaced block header is the parent of the added block header.

Checks

block header is from the ChainGraph of the Cut

result has valid braiding

result is a cut

update is monotonic

This includes a check that inductively maintains checkBraidingOfCut.

monotonicCutExtension :: MonadThrow m => Cut -> BlockHeader -> m Cut

tryMonotonicCutExtension :: MonadThrow m => Cut -> BlockHeader -> m (Maybe Cut)

Join

data Join a

Constructors

Join

Fields

_joinBase :: !Cut

_joinQueue :: !(JoinQueue a)

join :: Ord a => WebBlockHeaderDb -> (DiffItem BlockHeader -> DiffItem (Maybe a)) -> Cut -> Cut -> IO (Join a)

join_ :: forall a. Ord a => WebBlockHeaderDb -> (DiffItem BlockHeader -> DiffItem (Maybe a)) -> HashMap ChainId BlockHeader -> HashMap ChainId BlockHeader -> IO (Join a)

This merges two maps from ChainIds to BlockHeaders such that the result is a Cut. Note, however, that the resulting cut contains only the chain ids from the intersection of the input maps.

applyJoin :: MonadThrow m => Join a -> m Cut

prioritizeHeavier :: Cut -> Cut -> DiffItem BlockHeader -> DiffItem (Maybe Int)

prioritizeHeavier_ :: Foldable f => Ord (f BlockHeader) => f BlockHeader -> f BlockHeader -> DiffItem BlockHeader -> DiffItem (Maybe Int)

Note: consider the weight of the recursive dependencies for the priority of a block header. For that we would have to annotate the block headers before putting them in the queue. To traverse only once, we'd have to traverse the zipped cuts by height and not by chainid, which could easily be done by merging/zipping the branch-diff streams.

joinIntoHeavier :: WebBlockHeaderDb -> Cut -> Cut -> IO Cut

Merge two Cuts. If at least one of the input cuts had a valid braiding the result is guaranteed to have a valid braiding for all blocks included in cut and their ancestors.

This is because the merge starts with the intersection of both cuts, using branchDiff_ on each chain, and constructs the merge cut using tryMonotonicCutExtension. If one of the inputs is correctly braided, so is the intersection. tryMonotonicCutExtension is guaranteed to maintain that property.

joinIntoHeavier_ :: WebBlockHeaderDb -> HashMap ChainId BlockHeader -> HashMap ChainId BlockHeader -> IO Cut

Meet

meet :: WebBlockHeaderDb -> Cut -> Cut -> IO Cut

Intersection of cuts

Misc internal tools

zipCuts :: Cut -> Cut -> [(ChainId, BlockHeader, BlockHeader)]

If the cuts are from different graphs only the chain ids of the intersection are included in the result.