Chainweb.Cut.CutHashes

Description

Cut Id

data CutId

This is used to uniquly identify a cut.

TODO: should we use a MerkelHash for this, that that we could proof in clusion of a block header in a cut?

cutIdBytes :: CutId -> ShortByteString

encodeCutId :: MonadPut m => CutId -> m ()

decodeCutId :: MonadGet m => m CutId

cutIdToText :: CutId -> Text

cutIdFromText :: MonadThrow m => Text -> m CutId

cutIdToTextShort :: CutId -> Text

HasCutId

class HasCutId c where

Minimal complete definition

cutId | _cutId

Methods

_cutId :: c -> CutId

cutId :: Getter c CutId

CutHashes

data CutHashes

This data structure is used to inform other components and chainweb nodes about new cuts along with some properties of the cut.

Generally, the Cut is represented by providing the BlockHash of the respective BlockHeader for each chain in the Cut.

Optionally, a node may attach the PayloadData and/or the BlockHeader for some of the block of the Cut.

Constructors

CutHashes

Fields

_cutHashes :: !(HashMap ChainId (BlockHeight, BlockHash))

_cutOrigin :: !(Maybe PeerInfo)

Nothing is used for locally mined Cuts

_cutHashesWeight :: !BlockWeight

_cutHashesHeight :: !BlockHeight

_cutHashesChainwebVersion :: !ChainwebVersion

_cutHashesId :: !CutId

_cutHashesHeaders :: !(HashMap BlockHash BlockHeader)

optional block headers

_cutHashesPayloads :: !(HashMap BlockPayloadHash PayloadData)

optional block payloads

cutHashes :: Lens' CutHashes (HashMap ChainId (BlockHeight, BlockHash))

cutHashesChainwebVersion :: Lens' CutHashes ChainwebVersion

cutHashesId :: Lens' CutHashes CutId

cutOrigin :: Lens' CutHashes (Maybe PeerInfo)

cutHashesWeight :: Lens' CutHashes BlockWeight

cutHashesHeight :: Lens' CutHashes BlockHeight

cutHashesHeaders :: Lens' CutHashes (HashMap BlockHash BlockHeader)

cutHashesPayloads :: Lens' CutHashes (HashMap BlockPayloadHash PayloadData)

cutToCutHashes :: Maybe PeerInfo -> Cut -> CutHashes

Compute a CutHashes structure from a Cut. The result doesn't include any block headers or payloads.

type CutHashesCas cas = CasConstraint cas CutHashes