Chainweb.CutDB

Description

TODO

CutConfig

data CutDbConfig

Constructors

CutDbConfig

Fields

_cutDbConfigInitialCut :: !Cut

_cutDbConfigInitialCutFile :: !(Maybe FilePath)

_cutDbConfigBufferSize :: !Natural

_cutDbConfigLogLevel :: !LogLevel

_cutDbConfigTelemetryLevel :: !LogLevel

_cutDbConfigUseOrigin :: !Bool

cutDbConfigInitialCut :: Lens' CutDbConfig Cut

cutDbConfigInitialCutFile :: Lens' CutDbConfig (Maybe FilePath)

cutDbConfigBufferSize :: Lens' CutDbConfig Natural

cutDbConfigLogLevel :: Lens' CutDbConfig LogLevel

cutDbConfigTelemetryLevel :: Lens' CutDbConfig LogLevel

cutDbConfigUseOrigin :: Lens' CutDbConfig Bool

defaultCutDbConfig :: ChainwebVersion -> CutDbConfig

farAheadThreshold :: Int

We ignore cuts that are two far ahead of the current best cut that we have. There are two reasons for this:

It limits the effect of a DOS attack where an attack sends large fake cuts that would consume a lot of resources before we notice that the cut was fake.

It helps a node catching up with the overall consensus of the network by doing the catchup via several moderately fixed size steps instead of one giant step. The latter would consume an unlimited amount of resources, possibly leading to resource exhaustion on the local system. Also if the node is restarted during the catch-up process it resumes from the intermediate steps instead of starting all over again.

For the latter to work it is important that the local node only pulls cuts that are at most forAheadThreshold blocks in the ahead. Otherwise the pulled blocks would be immediately rejected by the cut processing pipeline processCuts below in the module and the node would never be able to join the consensus of the network.

NOTE: THIS NUMBER MUST BE STRICTLY LARGER THAN THE RESPECTIVE LIMIT IN Sync

Cut Hashes Table

cutHashesTable :: RocksDb -> RocksDbCas CutHashes

CutDb

data CutDb cas

This is a singleton DB that contains the latest chainweb cut as only entry.

cutDbWebBlockHeaderDb :: Getter (CutDb cas) WebBlockHeaderDb

cutDbWebBlockHeaderStore :: Getter (CutDb cas) WebBlockHeaderStore

cutDbBlockHeaderDb :: HasChainId cid => cid -> Fold (CutDb cas) BlockHeaderDb

Access the blockerheaderdb via the cutdb for a given chain id

cutDbPayloadCas :: Getter (CutDb cas) (PayloadDb cas)

cutDbPayloadStore :: Getter (CutDb cas) (WebBlockPayloadStore cas)

cutDbStore :: Getter (CutDb cas) (RocksDbCas CutHashes)

member :: CutDb cas -> ChainId -> BlockHash -> IO Bool

cut :: Getter (CutDb cas) (IO Cut)

Get the current Cut, which represent the latest chainweb state.

This the main API method of chainweb-consensus.

_cut :: CutDb cas -> IO Cut

Get the current Cut, which represent the latest chainweb state.

This the main API method of chainweb-consensus.

_cutStm :: CutDb cas -> STM Cut

An STM version of _cut.

_cut db is generally more efficient than as atomically (_cut db).

cutStm :: Getter (CutDb cas) (STM Cut)

An STM version of cut.

_cut db is generally more efficient than as atomically (_cut db).

awaitNewCut :: CutDb cas -> Cut -> IO Cut

A common idiom to spin while waiting for a guaranteed new Cut, different from the given one.

awaitNewCutByChainId :: CutDb cas -> ChainId -> Cut -> IO Cut

As in awaitNewCut, but only updates when the specified ChainId has grown.

cutStream :: MonadIO m => CutDb cas -> Stream (Of Cut) m r

Stream of most recent cuts. This stream does not generally include the full history of cuts. When no cuts are demanded from the stream or new cuts are produced faster than they are consumed from the stream, the stream skips over cuts and always returns the latest cut in the db.

addCutHashes :: CutDb cas -> CutHashes -> IO ()

withCutDb :: PayloadCas cas => CutDbConfig -> LogFunction -> WebBlockHeaderStore -> WebBlockPayloadStore cas -> RocksDbCas CutHashes -> (CutDb cas -> IO a) -> IO a

startCutDb :: PayloadCas cas => CutDbConfig -> LogFunction -> WebBlockHeaderStore -> WebBlockPayloadStore cas -> RocksDbCas CutHashes -> IO (CutDb cas)

Start a CutDB. This loads the initial cut from the database (falling back to the configured initial cut loading fails) and starts the cut validation pipeline.

TODO: Instead of falling back to the configured initial cut, which usually is the genesis cut, we could instead try to walk back in history until we find a cut that succeed to load.

stopCutDb :: CutDb cas -> IO ()

Stop the cut validation pipeline.

cutDbQueueSize :: CutDb cas -> IO Natural

blockStream :: MonadIO m => CutDb cas -> Stream (Of BlockHeader) m r

blockDiffStream :: MonadIO m => CutDb cas -> Stream (Of (Either BlockHeader BlockHeader)) m r

cutStreamToHeaderStream :: forall m cas r. MonadIO m => CutDb cas -> Stream (Of Cut) m r -> Stream (Of BlockHeader) m r

Given a stream of cuts, produce a stream of all blocks included in those cuts. Blocks of the same chain are sorted by block height.

cutStreamToHeaderDiffStream :: forall m cas r. MonadIO m => CutDb cas -> Stream (Of Cut) m r -> Stream (Of (Either BlockHeader BlockHeader)) m r

Given a stream of cuts, produce a stream of all blocks included in those cuts. Blocks of the same chain are sorted by block height.

Some CutDb

newtype CutDbT cas (v :: ChainwebVersionT)

CutDb with type level ChainwebVersion

Constructors

CutDbT (CutDb cas)

data SomeCutDb cas

Constructors

forall v.KnownChainwebVersionSymbol v => SomeCutDb (CutDbT cas v)

someCutDbVal :: ChainwebVersion -> CutDb cas -> SomeCutDb cas