Chainweb.Difficulty

Description

TODO

PowHashNat

newtype PowHashNat

A type that maps block hashes to unsigned 256 bit integers by projecting onto the first 8 bytes (least significant in little endian encoding) and interpreting them as little endian encoded unsigned integer value.

Arithmetic is defined as unsigned bounded integer arithmetic. Overflows result in an exception and may result in program abort.

Constructors

PowHashNat Word256

powHashNat :: PowHash -> PowHashNat

encodePowHashNat :: MonadPut m => PowHashNat -> m ()

decodePowHashNat :: MonadGet m => m PowHashNat

encodePowHashNatBe :: MonadPut m => PowHashNat -> m ()

decodePowHashNatBe :: MonadGet m => m PowHashNat

HashTarget

newtype HashTarget

HashTarget

target = maxBound / (network hash rate * block time) = maxBound / difficulty

network hash rate is interpolated from observered past block times.

Constructors

HashTarget PowHashNat

showTargetHex :: HashTarget -> Text

A visualization of a HashTarget as binary.

showTargetBits :: HashTarget -> Text

A visualization of a HashTarget as binary.

checkTarget :: HashTarget -> PowHash -> Bool

The critical check in Proof-of-Work mining: did the generated hash match the target?

maxTarget :: HashTarget

By maximum, we mean "easiest".

maxTargetWord :: Word256

difficultyToTarget :: HashDifficulty -> HashTarget

Given the same ChainwebVersion, forms an isomorphism with targetToDifficulty.

targetToDifficulty :: HashTarget -> HashDifficulty

Given the same ChainwebVersion, forms an isomorphism with difficultyToTarget.

encodeHashTarget :: MonadPut m => HashTarget -> m ()

decodeHashTarget :: MonadGet m => m HashTarget

HashDifficulty

newtype HashDifficulty

Hash Difficulty

difficulty = maxBound / target = network hash rate * block time

Constructors

HashDifficulty PowHashNat

encodeHashDifficulty :: MonadPut m => HashDifficulty -> m ()

decodeHashDifficulty :: MonadGet m => m HashDifficulty

encodeHashDifficultyBe :: MonadPut m => HashDifficulty -> m ()

decodeHashDifficultyBe :: MonadGet m => m HashDifficulty

Difficulty Adjustment

adjust :: ChainwebVersion -> WindowWidth -> TimeSpan Micros -> HashTarget -> HashTarget

A new HashTarget, based on the rate of mining success over the previous N blocks.

Epoch-based Difficulty Adjustment

This function represents a Bitcoin-inspired, "epoch-based" adjustment algorithm. For every N blocks (as defined by WindowWidth), we perform an adjustment.

Terminology

BlockHeader stores a 256-bit measure of difficulty: HashTarget. More precisely, HashTarget is a derivation (seen below) of the HashDifficulty. HashDifficulty in itself is roughly a measure of the number of hashes necessary to "solve" a block. For non-POW testing scenarios that use trivial targets (i.e. maxBound), then difficulty is exactly the number of necessary hashes. For POW mining, this is offset. See maxTarget.

A HashDifficulty of 1 is considered the "easiest" difficulty, and represents a HashTarget of maxTarget. There must never be a difficulty of 0.

Given the same ChainwebVersion, the functions targetToDifficulty and difficultyToTarget form an isomorphism between the above mentioned types.

Justification

We define the maximum possible hash target (the "easiest" target) as follows:

\[ \begin{align*} \text{MaxBound} &= 2^{256} - 1 \\ \text{MaxTarget} &= \frac{\text{MaxBound}}{2^{\text{offset}}} \end{align*} \]

where offset is some number of bits, 0 for trivial scenarios and some experimentally discovered \(N\) for real POW mining scenarios. For Bitcoin, \(N = 32\).

Given some difficulty \(D\), its corresponding HashTarget can be found by:

\[ \text{Target} = \frac{\text{MaxTarget}}{D} \]

During adjustment, we seek to solve for some new \(D\). From the above, it follows that the expected number of hashes necessary to "solve" a block becomes:

\[ \text{Expected} = \frac{D * \text{MaxBound}}{\text{MaxTarget}} \]

If we expect a block to be solved every \(R\) seconds, we find our total Network Hash Rate:

\[ \text{HashRate} = \frac{\text{Expected}}{R} \]

But, as a block chain is a dynamic system, the real time it took to mine some block would likely not be exactly \(R\). This implies:

\[ \begin{align*} \frac{\text{Expected}}{R} &= \text{HashRate} = \frac{\text{Expected}'}{M} \\ \frac{D * \text{MaxBound}}{R * \text{MaxTarget}} &= \text{HashRate} = \frac{D' * \text{MaxBound}}{M * \text{MaxTarget}} \\ \frac{D}{R} &= \text{HashRate} = \frac{D'}{M} \end{align*} \]

where \(D'\) is the known difficulty from the previous block, \(M\) is the average time in seconds it took to calculate the previous \(B\) blocks. The value of \(B\) is assumed to be configurable.

Given this, our new \(D\) is a simple ratio:

\[ D = \frac{D' * R}{M} \]

HashRate will of course not stay fixed as the network grows. Luckily, the difference in \(M\) values will naturally correct for this in the calculation of a new \(D\).

Precision

In real systems, the difference between \(M\) and \(R\) may be minute. To ensure that:

differences are not lost to integer-math rounding errors

adjustment actually occurs

small, incremental adjustments are allowed to build into greater change over time

Word256-based overflows do not occur

the algorithm is simple

we use the infinite-precision Rational type in the calculation of the new \(D\). Only when being converted to a final HashTarget is the non-integer precision discarded.

Note: Use of Rational is likely not our final solution, and complicates any cross-language spec we would write regarding adjustment algorithm expectations. For now, however, Rational is stable for a Haskell-only environment.

Test Properties

properties :: [(String, Property)]

prop_littleEndian :: Bool

Orphan instances

NFData Word128

Instance details

Methods

rnf :: Word128 -> ()

NFData Word256

Instance details

Methods

rnf :: Word256 -> ()