Chainweb.Graph

Description

A chain graph is

directed

regular

symmetric

irreflexive

Exceptions

data ChainGraphException :: Type where

This exceptions are not about the properties of the graph itself but about properties of enties (BlockHeader graph) that are constrained by this graph. So, maybe we should move this and the respective checks to the place where those enties are defined and rename these exceptions accordingly. However, keeping it here remove code duplication.

Constructors

ChainNotInChainGraphException :: Expected (HashSet ChainId) -> Actual ChainId -> ChainGraphException

AdjacentChainMismatch :: Expected (HashSet ChainId) -> Actual (HashSet ChainId) -> ChainGraphException

ChainNotAdjacentException :: Expected ChainId -> Actual (HashSet ChainId) -> ChainGraphException

Chain Graph

data ChainGraph

chainGraphKnown :: Getter ChainGraph KnownGraph

toChainGraph :: (a -> ChainId) -> KnownGraph -> DiGraph a -> ChainGraph

This function is unsafe, it throws an error if the graph isn't a valid chain graph. That's OK, since chaingraphs are hard-coded in the code and won't change dynamically, except for during testing.

validChainGraph :: DiGraph ChainId -> Bool

A valid chain graph is symmetric, regular, and the out-degree is at least 1 if the graph has at least two vertices.

These properties imply that the graph is strongly connected.

adjacentChainIds :: HasChainId p => ChainGraph -> p -> HashSet ChainId

class HasChainGraph a where

Minimal complete definition

_chainGraph | chainGraph

Methods

_chainGraph :: a -> ChainGraph

chainGraph :: Getter a ChainGraph

Undirected Edges

data AdjPair a

pattern Adj :: HasChainId a => a -> a -> AdjPair a

_getAdjPair :: AdjPair a -> (a, a)

adjs :: ChainGraph -> HashSet (AdjPair ChainId)

adjsOfVertex :: HasChainId p => ChainGraph -> p -> HashSet (AdjPair ChainId)

Graph Properties

shortestPath :: ChainId -> ChainId -> ChainGraph -> [ChainId]

diameter :: ChainGraph -> Natural

size :: ChainGraph -> Natural

The size of the graph refers to the number of edges |E| of a given graph.

order :: ChainGraph -> Natural

The order of the graph refers to the number of vertices |V| of a given graph.

degree :: ChainGraph -> Natural

Checks with a given chain graph

isWebChain :: HasChainGraph g => HasChainId p => g -> p -> Bool

graphChainIds :: ChainGraph -> HashSet ChainId

checkWebChainId :: MonadThrow m => HasChainGraph g => HasChainId p => g -> p -> m ()

Given a ChainGraph g, checkWebChainId p checks that p is a vertex in g.

checkAdjacentChainIds :: MonadThrow m => HasChainGraph g => HasChainId cid => HasChainId adj => g -> cid -> Expected (HashSet adj) -> m (HashSet adj)

Given a ChainGraph g, checkAdjacentChainIds cid as checks that the ChainId cid is in g and the set of adjacents chain ids of cid is the expected set as.

Specific, Known Graphs

data KnownGraph

Graphs which have known, specific, intended meaning for Chainweb.

Constructors

Singleton

Pair

Triangle

Peterson

Twenty

HoffmanSingle

knownGraph :: KnownGraph -> ChainGraph

singletonChainGraph :: ChainGraph

pairChainGraph :: ChainGraph

triangleChainGraph :: ChainGraph

petersonChainGraph :: ChainGraph

twentyChainGraph :: ChainGraph

hoffmanSingletonGraph :: ChainGraph