Chainweb.HostAddress

Description

Port numbers must be within the range [0,2^16-1].

All hostnames are considered fully qualified and thus the final dot is omitted.

For hostnames we follow the specification for "Server-based Naming Authority" for URIs from RFC2396 section 3.2.2.:

     hostport      = host [ ":" port ]
     host          = hostname | IPv4address
     hostname      = *( domainlabel "." ) toplabel [ "." ]
     domainlabel   = alphanum | alphanum *( alphanum | "-" ) alphanum
     toplabel      = alpha | alpha *( alphanum | "-" ) alphanum

     IPv4address   = 1*digit "." 1*digit "." 1*digit "." 1*digit
     port          = *digit

1*digit designates the decimal representation of an octet. The specification takes the form of hostnames from section 2.1 RFC1123, but limiting the rightmost (top-most) label to the from given in section 3 of RFC1034, which allows to disambiguate domain names and IPv4 addresses.

IPv6 Addresses are partially supported. IPv6 address are parsed as described in RFC4291, but embedding of IPv4 addresses is not supported. IPv6 addresses are printed exactly as they where parsed. No normalization is performed. In particular the recommendations from RFC5952 are not considered. For host addresses RFC3986 and RFC 5952 are followed by requiring that IPv6 literals are enclosed in square brackets. Anything else from RFC3986, which is concerning URIs is ignored.

Additional restriction for hostname apply from RFC1123: labels must have not more than 63 octets, letters are case-insenstive. The maximum length must not exceed 254 octets, excluding the (optional) terminating dot.

See https://cs.uwaterloo.ca/twiki/view/CF/HostNamingRules for an extensive overview of different standards for host names.

Non-ascii characters are encoded via Punycode and are of no concern in this implementation.

Port Numbers

data Port

portToText :: Port -> Text

portFromText :: MonadThrow m => Text -> m Port

pPort :: Maybe String -> OptionParser Port

readPortBytes :: MonadThrow m => ByteString -> m Port

Hostnames

data Hostname

hostnameBytes :: Hostname -> ByteString

localhost :: Hostname

localhostIPv4 :: Hostname

Using explicit IP addresses and not to "localhost" greatly improves networking performance and Mac OS X.

localhostIPv6 :: Hostname

Using explicit IP addresses and not to "localhost" greatly improves networking performance and Mac OS X.

readHostnameBytes :: MonadThrow m => ByteString -> m Hostname

hostnameToText :: Hostname -> Text

hostnameFromText :: MonadThrow m => Text -> m Hostname

unsafeHostnameFromText :: HasCallStack => Text -> Hostname

pHostname :: Maybe String -> OptionParser Hostname

HostAddresses

data HostAddress

Constructors

HostAddress

Fields

_hostAddressHost :: !Hostname

_hostAddressPort :: !Port

hostAddressPort :: Lens' HostAddress Port

hostAddressHost :: Lens' HostAddress Hostname

hostAddressBytes :: HostAddress -> ByteString

readHostAddressBytes :: MonadThrow m => ByteString -> m HostAddress

hostAddressToText :: HostAddress -> Text

hostAddressFromText :: MonadThrow m => Text -> m HostAddress

unsafeHostAddressFromText :: HasCallStack => Text -> HostAddress

arbitraryHostAddress :: Gen HostAddress

pHostAddress :: Maybe String -> MParser HostAddress

pHostAddress' :: Maybe String -> OptionParser HostAddress

hostAddressToBaseUrl :: Scheme -> HostAddress -> BaseUrl

Arbitrary Values

arbitraryPort :: Gen Port

arbitraryDomainName :: Gen Hostname

arbitraryHostname :: Gen Hostname

arbitraryIpV4 :: Gen Hostname

TODO should we exclude network, broadcast, otherwise special values?

arbitraryIpV6 :: Gen Hostname

HostPreference Utils

hostPreferenceToText :: HostPreference -> Text

hostPreferenceFromText :: MonadThrow m => Text -> m HostPreference

Properties

properties :: [(String, Property)]

Orphan instances

HasTextRepresentation HostPreference

Instance details

Methods

toText :: HostPreference -> Text

fromText :: MonadThrow m => Text -> m HostPreference