Chainweb.Logging.Amberdata

Description

Data structures and functions for logging BlocHeader data to Amberdata

Documentation

data AmberdataBlock

Constructors

AmberdataBlock

Fields

_amberdataNumber :: !BlockHeight

_amberdataHash :: !BlockHash

_amberdataTimestamp :: !BlockCreationTime

_amberdataParentHash :: !BlockHash

_amberdataNonce :: !Nonce

_amberdataSize :: !Word

Bytes

_amberdataNumTransactions :: !Word

_amberdataMeta :: !ChainId

_amberdataDifficulty :: !BlockWeight

data AmberdataConfig

Constructors

AmberdataConfig

Fields

_ambredataConfigHost :: HostAddress

_amberdataApiKey :: AmberdataApiKey

_amberdataBlockchainId :: AmberdataBlockchainId

_amberdataChainId :: Maybe ChainId

_amberdataDebug :: Bool

defaultAmberdataConfig :: AmberdataConfig

validateAmberdataConfig :: ConfigValidation AmberdataConfig []

pAmberdataConfig :: MParser AmberdataConfig

amberdataBlockMonitor :: (PayloadCas cas, Logger logger) => Maybe ChainId -> logger -> CutDb cas -> IO ()

withAmberDataBlocksBackend :: Manager -> AmberdataConfig -> (Backend (JsonLog AmberdataBlock) -> IO b) -> IO b

A backend for JSON log messages that sends all logs to the Amberdata /blocks endpoint. Messages are sent in a fire-and-forget fashion. If a connection fails, the messages are dropped without notice.