Chainweb.Logging.Config

Description

Chainweb Logging Configuration

Logging Config

data ClusterId

An user provided label that is used to tag log messages from nodes.

It intended purpose is to allow users to defines arbitrary groups of nodes and tag the respective log messages.

data LogConfig

General logging config

Constructors

LogConfig

Fields

_logConfigLogger :: !LoggerConfig

_logConfigBackend :: !BackendConfig

_logConfigTelemetryBackend :: !(EnableConfig BackendConfig)

_logConfigAmberdataBackend :: !(EnableConfig AmberdataConfig)

_logConfigClusterId :: !(Maybe ClusterId)

_logConfigFilter :: !LogFilter

logConfigLogger :: Lens' LogConfig LoggerConfig

logConfigBackend :: Lens' LogConfig BackendConfig

logConfigTelemetryBackend :: Lens' LogConfig (EnableConfig BackendConfig)

logConfigClusterId :: Lens' LogConfig (Maybe ClusterId)

logConfigAmberdataBackend :: Lens' LogConfig (EnableConfig AmberdataConfig)

defaultLogConfig :: LogConfig

validateLogConfig :: ConfigValidation LogConfig []

pLogConfig :: MParser LogConfig

pLogConfig_

Arguments

:: Text
prefix for this and all subordinate command line options.

-> MParser LogConfig

A version of pLogConfig that takes a prefix for the command line option.