Chainweb.Mempool.Mempool

Description

Mempool

The mempool contains an in-memory store of all of the pending transactions that have been submitted by users for inclusion into blocks, and is responsible for collecting the candidate transaction set when we prepare a new block for mining.

The in-memory mempool implementation (see InMem.hs and InMemTypes.hs) contains three datasets:

a priority search queue of pending transactions, keyed by pact transaction hash, and (currently) ordered by gas price (descending) with ties broken by gas limit (ascending). Basically -- we prefer highest bidder first, and all else being equal, smaller transactions.

a recent-modifications log, storing the last k updates to the mempool. This is in here so peers can sync with us by fetching starting with a high-water mark, rather than having to go through a complete re-sync of all transaction hashes.

Transaction lifecycle:

a transaction is created and signed by a user, and sent (via the pact /send endpoint) to the mempool, using mempoolInsert

miner calls mempoolGetBlock, at which point the top transactions (up to the gas limit) are run through the MempoolPreBlockCheck and discarded if they fail

the transaction makes it into a mined block

consensus calls pact's newBlock to create the next block; at this point, the Consensus module processes the fork and: a) removes any transactions that made it onto the winning fork so that they are no longer considered for inclusion in blocks or gossiped, b) reintroduces any transactions into the mempool that were on the losing fork but didn't make it onto the winning one.

The mempool API is defined as a record-of-functions in MempoolBackend.

Documentation

data MempoolBackend t

Mempool backend API. Here t is the transaction payload type.

Constructors

MempoolBackend

Fields

mempoolTxConfig :: !(TransactionConfig t)

mempoolBlockGasLimit :: GasLimit

mempoolMember :: Vector TransactionHash -> IO (Vector Bool)

Returns true if the given transaction hash is known to this mempool.

mempoolLookup :: Vector TransactionHash -> IO (Vector (LookupResult t))

Lookup transactions in the pending queue by hash.

mempoolInsert :: InsertType -> Vector t -> IO ()

Insert the given transactions into the mempool.

mempoolInsertCheck :: Vector t -> IO (Either (T2 TransactionHash InsertError) ())

Perform the pre-insert check for the given transactions. Short-circuits on the first Transaction that fails.

mempoolMarkValidated :: Vector TransactionHash -> IO ()

Remove the given hashes from the pending set.

mempoolGetBlock :: MempoolPreBlockCheck t -> BlockHeight -> BlockHash -> GasLimit -> IO (Vector t)

given maximum block size, produce a candidate block of transactions for mining.

TODO remove gas limit argument here

mempoolGetPendingTransactions :: Maybe HighwaterMark -> (Vector TransactionHash -> IO ()) -> IO HighwaterMark

given a previous high-water mark and a chunk callback function, loops through the pending candidate transactions and supplies the hashes to the callback in chunks. No ordering of hashes is presupposed. Returns the remote high-water mark.

mempoolClear :: IO ()

A hook to clear the mempool. Intended only for the in-mem backend and only for testing.

type MempoolPreBlockCheck t = BlockHeight -> BlockHash -> Vector t -> IO (Vector Bool)

data TransactionConfig t

Mempool operates over a transaction type t. Mempool needs several functions on t, e.g. "how do you encode and decode t over the wire?" and "how is t hashed?". This information is passed to mempool in a TransactionConfig.

Constructors

TransactionConfig

Fields

txCodec :: !(Codec t)

converting transactions to/from bytestring. Note: the generated bytestring is currently expected to be valid utf8 so that it can be safely JSON-encoded by servant (we should revisit this)

txHasher :: t -> TransactionHash

hash function to use when making transaction hashes.

txHashMeta :: !HashMeta

hash function metadata. Currently informational only -- for future use.

txGasPrice :: t -> GasPrice

getter for the transaction gas price.

txGasLimit :: t -> GasLimit

getter for transaction gas limit.

txMetadata :: t -> TransactionMetadata

getter for transaction metadata (creation and expiry timestamps)

newtype TransactionHash

Raw/unencoded transaction hashes.

TODO: production versions of this kind of DB should salt with a runtime-generated constant to avoid collision attacks; see the "hashing and security" section of the hashable docs.

Constructors

TransactionHash ShortByteString

data TransactionMetadata

Constructors

TransactionMetadata

Fields

txMetaCreationTime :: !(Time Micros)

txMetaExpiryTime :: !(Time Micros)

type MempoolTxId = Int64

data HashMeta

Mempools will check these values match in APIs

Constructors

HashMeta

Fields

hashmetaName :: !Text

hashmetaLenBytes :: !Int

data ValidatedTransaction t

Constructors

ValidatedTransaction

Fields

validatedHeight :: !BlockHeight

validatedHash :: !BlockHash

validatedTransaction :: !t

data LookupResult t

Constructors

Missing

Pending t

data MockTx

Mempool only cares about a few projected values from the transaction type (gas price, gas limit, hash, metadata), so our mock transaction type will only contain these (plus a nonce)

Constructors

MockTx

Fields

mockNonce :: !Int64

mockGasPrice :: !GasPrice

mockGasLimit :: !GasLimit

mockMeta :: !TransactionMetadata

type ServerNonce = Int

type HighwaterMark = (ServerNonce, MempoolTxId)

data InsertType

Constructors

CheckedInsert

UncheckedInsert

data InsertError

Constructors

InsertErrorDuplicate

InsertErrorInvalidTime

InsertErrorOversized

InsertErrorBadlisted

InsertErrorMetadataMismatch

InsertErrorTransactionsDisabled

InsertErrorOther Text

chainwebTransactionConfig :: TransactionConfig ChainwebTransaction

mockCodec :: Codec MockTx

A codec for transactions when sending them over the wire.

mockEncode :: MockTx -> ByteString

mockBlockGasLimit :: GasLimit

chainwebTestHasher :: ByteString -> TransactionHash

chainwebTestHashMeta :: HashMeta

noopMempool :: IO (MempoolBackend t)

noopMempoolPreBlockCheck :: MempoolPreBlockCheck t

syncMempools

Arguments

:: Show t

=> LogFunctionText

-> Int
polling interval in microseconds

-> MempoolBackend t
local mempool

-> MempoolBackend t
remote mempool

-> IO ()

syncMempools'

Arguments

:: Show t

=> LogFunctionText

-> Int
polling interval in microseconds

-> MempoolBackend t
local mempool

-> MempoolBackend t
remote mempool

-> IO ()

Pulls any missing pending transactions from a remote mempool.

The initial sync procedure:

get the list of pending transaction hashes from remote,

lookup any hashes that are missing, and insert them into local,

push any hashes remote is missing

After initial sync, will loop subscribing to updates from remote.

Loops until killed, or until the underlying mempool connection throws an exception.

newtype GasLimit

Constructors

GasLimit ParsedInteger