Chainweb.MerkleLogHash

Description

TODO

MerkleLogHash

newtype MerkleLogHash

Constructors

MerkleLogHash (MerkleRoot (HashAlg ChainwebHashTag))

type MerkleLogHashBytesCount = 32

TODO: consider parameterizing this value on the ChainwebVersion. (e.g. for Test we may want to use just () or Int)

merkleLogHashBytesCount :: Natural

merkleLogHash :: MonadThrow m => ByteString -> m MerkleLogHash

Smart constructor

unsafeMerkleLogHash :: HasCallStack => ByteString -> MerkleLogHash

encodeMerkleLogHash :: MonadPut m => MerkleLogHash -> m ()

decodeMerkleLogHash :: MonadGet m => m MerkleLogHash

nullHashBytes :: MerkleLogHash

oneHashBytes :: MerkleLogHash

randomMerkleLogHash :: MonadIO m => m MerkleLogHash

This must be used only for testing. The result hash is uniformily distributed, but not cryptographically safe.