Chainweb.Miner.Coordinator

Description

Types

newtype MiningState

Data shared between the mining threads represented by newWork and publish.

Constructors

MiningState (Map BlockPayloadHash (T3 Miner PrevTime PayloadWithOutputs))

data MiningStats

For logging during MiningState manipulation.

Constructors

MiningStats

Fields

_statsCacheSize :: Int

_states503s :: Int

_statsAvgTxs :: Int

newtype PrevTime

The BlockCreationTime of the parent of some current, "working" BlockHeader.

Constructors

PrevTime BlockCreationTime

data ChainChoice

Constructors

Anything

TriedLast ChainId

Suggestion ChainId

Functions

newWork :: ChainChoice -> Miner -> PactExecutionService -> Cut -> IO (T3 PrevTime BlockHeader PayloadWithOutputs)

Construct a new BlockHeader to mine on.

publish :: LogFunction -> MiningState -> CutDb cas -> BlockHeader -> IO ()

KILLSWITCH: This extra logic involving txSilenceEndDate is to be removed in a future version of Chainweb. It prevents this Node from generating any new Cuts after a specified date.