Chainweb.Miner.Core

Description

Documentation

newtype HeaderBytes

The encoded form of a BlockHeader.

Constructors

HeaderBytes

Fields

_headerBytes :: ByteString

newtype TargetBytes

The encoded form of a HashTarget.

Constructors

TargetBytes

Fields

_targetBytes :: ByteString

newtype ChainBytes

The encoded form of a ChainId.

Constructors

ChainBytes

Fields

_chainBytes :: ByteString

newtype WorkBytes

Encoding of ChainId + HashTarget + BlockHeader to be consumed by a remote Mining API.

Constructors

WorkBytes

Fields

_workBytes :: ByteString

workBytes :: ChainBytes -> TargetBytes -> HeaderBytes -> WorkBytes

Combine ChainBytes, TargetBytes and HeaderBytes in such a way that can be later undone by unWorkBytes.

unWorkBytes :: WorkBytes -> T3 ChainBytes TargetBytes HeaderBytes

NOTE: This makes a low-level assumption about the encoded size of HashTarget and ChainId!

data MiningResult

Constructors

MiningResult

Fields

_mrNonceBytes :: !ByteString

_mrNumNoncesTried :: !Word64

_mrEstimatedHashesPerSec :: !Word64

_mrStderr :: ByteString

usePowHash :: ChainwebVersion -> (forall a. HashAlgorithm a => Proxy a -> f) -> f

Select a hashing algorithm.

mine :: forall a. HashAlgorithm a => Proxy a -> Nonce -> TargetBytes -> HeaderBytes -> IO (T2 HeaderBytes Word64)

This Miner makes low-level assumptions about the chainweb protocol. It may break if the protocol changes.

TODO: Check the chainweb version to make sure this function can handle the respective version.

TODO: Remove the Proxy?

fastCheckTarget :: Ptr Word64 -> Ptr Word64 -> IO Bool

PowHashNat interprets POW hashes as unsigned 256 bit integral numbers in little endian encoding, hence we compare against the target from the end of the bytes first, then move toward the front 8 bytes at a time.

injectNonce :: Nonce -> Ptr Word8 -> IO ()

injectNonce makes low-level assumptions about the byte layout of a hashed BlockHeader. If that layout changes, this functions need to be updated. The assumption allows us to iterate on new nonces quickly.

callExternalMiner

Arguments

:: Path Absolute
miner path

-> [String]
miner extra args

-> Bool
save stderr?

-> ByteString
target hash

-> ByteString
block bytes

-> IO (Either String MiningResult)