Chainweb.Miner.Pact

Description

The definition of the Pact miner and the Pact miner reward.

Data

newtype MinerId

MinerId is a thin wrapper around Text to differentiate it from user addresses.

Constructors

MinerId Text

newtype MinerKeys

MinerKeys are a thin wrapper around a Pact KeySet to differentiate it from user keysets.

Constructors

MinerKeys KeySet

data Miner

Miner info data consists of a miner id (text), and its keyset (a pact type).

Constructors

Miner !MinerId !MinerKeys

data MinerRewards

Constructors

MinerRewards

Fields

_minerRewards :: !(HashMap BlockHeight ParsedDecimal)

The map of blockheight thresholds to miner rewards

_minerRewardHeights :: !(Vector BlockHeight)

A (sorted) vector of blockheights (head is most recent)

Combinators

toMinerData :: Miner -> MinerData

Convert from Pact Miner to Chainweb MinerData.

fromMinerData :: MonadThrow m => MinerData -> m Miner

Convert from Chainweb MinerData to Pact Miner.

readRewards :: HasChainGraph v => v -> MinerRewards

Rewards table mapping 3-month periods to their rewards according to the calculated exponential decay over 120 year period

rawMinerRewards :: ByteString

Read in the reward csv via TH for deployment purposes.

Optics

minerId :: Lens' Miner MinerId

A lens into the miner id of a miner.

minerKeys :: Lens' Miner MinerKeys

A lens into the miner keys of a miner.

minerRewardHeights :: Lens' MinerRewards (Vector BlockHeight)

A lens into the sorted vector of significant block heights pegged to a reward

minerRewards :: Getter MinerRewards (HashMap BlockHeight ParsedDecimal)

A getter into the map of heights to rewards

Defaults

noMiner :: Miner

A trivial Miner.

defaultMiner :: Miner

Keyset taken from cp examples in Pact The private key here was taken from `examples/cp` from the Pact repository

CLI Utils

pMiner :: Parser Miner