Chainweb.Miner.RestAPI

Description

Mining API

type MiningApi_ = ("mining" :> ("work" :> (QueryParam "chain" ChainId :> (ReqBody '[JSON] Miner :> Get '[OctetStream] WorkBytes)))) :<|> (("mining" :> ("solved" :> (ReqBody '[OctetStream] HeaderBytes :> Post '[JSON] NoContent))) :<|> ("mining" :> ("updates" :> (ReqBody '[OctetStream] ChainBytes :> Raw))))

work: To have a new BlockHeader assembled to mine upon. Miner information is required in order to properly distribute mining rewards.

solved: To yield a solved BlockHeader back to a Chainweb Node for it to be reassociated with its Cut and Payload, then published to the rest of the network.

type MiningApi (v :: ChainwebVersionT) = 'ChainwebEndpoint v :> Reassoc MiningApi_

miningApi :: forall (v :: ChainwebVersionT). Proxy (MiningApi v)

someMiningApi :: ChainwebVersion -> SomeApi