Chainweb.Miner.RestAPI.Server

Description

Documentation

workHandler :: Logger l => MiningCoordination l cas -> ChainwebVersion -> Maybe ChainId -> Miner -> Handler WorkBytes

KILLSWITCH: The logic here involving txSilenceEndDate is to be removed in a future version of Chainweb. This logic errors on remote requests for new mining work, such that this Node can no longer meaningfully participate in mining.

workHandler' :: forall l cas. Logger l => MiningCoordination l cas -> Maybe ChainId -> Miner -> IO WorkBytes

solvedHandler :: forall l cas. Logger l => MiningCoordination l cas -> HeaderBytes -> IO NoContent

updatesHandler :: CutDb cas -> ChainBytes -> Tagged Handler Application

miningServer :: forall l cas (v :: ChainwebVersionT). Logger l => MiningCoordination l cas -> ChainwebVersion -> Server (MiningApi v)

someMiningServer :: Logger l => ChainwebVersion -> MiningCoordination l cas -> SomeServer