Chainweb.Pact.Backend.Types

Description

Chainweb / Pact Types module for various database backends

Documentation

data CheckpointEnv

Constructors

CheckpointEnv

Fields

_cpeCheckpointer :: !Checkpointer

_cpeLogger :: !Logger

cpeCheckpointer :: Lens' CheckpointEnv Checkpointer

cpeLogger :: Lens' CheckpointEnv Logger

data Checkpointer

Constructors

Checkpointer

Fields

_cpRestore :: !(Maybe (BlockHeight, ParentHash) -> IO PactDbEnv')

prerequisite: (BlockHeight - 1, ParentHash) is a direct ancestor of the "latest block"

_cpSave :: !(BlockHash -> IO ())

commits pending modifications to block, with the given blockhash

_cpDiscard :: IO ()

discard pending block changes

_cpGetLatestBlock :: IO (Maybe (BlockHeight, BlockHash))

get the checkpointer's idea of the latest block

_cpBeginCheckpointerBatch :: IO ()

_cpCommitCheckpointerBatch :: IO ()

_cpDiscardCheckpointerBatch :: IO ()

_cpLookupBlockInCheckpointer :: !((BlockHeight, BlockHash) -> IO Bool)

is the checkpointer aware of the given block?

_cpGetBlockParent :: !((BlockHeight, BlockHash) -> IO (Maybe BlockHash))

_cpRegisterProcessedTx :: !(PactHash -> IO ())

_cpLookupProcessedTx :: !(PactHash -> IO (Maybe (T2 BlockHeight BlockHash)))

data Env'

Constructors

forall a. Env' (PactDbEnv (DbEnv a))

data EnvPersist'

Constructors

forall a. EnvPersist' (PactDbEnvPersist a)

data PactDbConfig

Constructors

PactDbConfig

Fields

_pdbcPersistDir :: Maybe FilePath

_pdbcLogDir :: FilePath

_pdbcPragmas :: [Pragma]

_pdbcGasLimit :: Maybe Int

_pdbcGasRate :: Maybe Int

pdbcGasLimit :: Lens' PactDbConfig (Maybe Int)

pdbcGasRate :: Lens' PactDbConfig (Maybe Int)

pdbcLogDir :: Lens' PactDbConfig FilePath

pdbcPersistDir :: Lens' PactDbConfig (Maybe FilePath)

pdbcPragmas :: Lens' PactDbConfig [Pragma]

data PactDbEnv'

Constructors

forall e. PactDbEnv' (PactDbEnv e)

data PactDbEnvPersist p

Constructors

PactDbEnvPersist

Fields

_pdepPactDb :: PactDb (DbEnv p)

_pdepEnv :: DbEnv p

pdepEnv :: forall p. Lens' (PactDbEnvPersist p) (DbEnv p)

pdepPactDb :: forall p. Lens' (PactDbEnvPersist p) (PactDb (DbEnv p))

data PactDbState

Constructors

PactDbState

Fields

_pdbsDbEnv :: EnvPersist'

pdbsDbEnv :: Iso' PactDbState EnvPersist'

data SQLiteRowDelta

Within a restore .. save block, mutations to the pact database are held in RAM to be written to the DB in batches at save time. For any given db write, we need to record the table name, the current tx id, the row key, and the row value.

Constructors

SQLiteRowDelta

Fields

_deltaTableName :: !ByteString

_deltaTxId :: !TxId

_deltaRowKey :: !ByteString

_deltaData :: !ByteString

data SQLiteDeltaKey

When we index SQLiteRowDelta values, we need a lookup key.

Constructors

SQLiteDeltaKey

Fields

_dkTable :: !ByteString

_dkRowKey :: !ByteString

type SQLitePendingTableCreations = HashSet ByteString

Between a restore..save bracket, we also need to record which tables were created during this block (so the necessary CREATE TABLE statements can be performed upon block save).

type SQLitePendingWrites = HashMap SQLiteDeltaKey (DList SQLiteRowDelta)

Pending writes to the pact db during a block, to be recorded in BlockState.

data SQLitePendingData

A collection of pending mutations to the pact db. We maintain two of these; one for the block as a whole, and one for any pending pact transaction. Upon pact transaction commit, the two SQLitePendingData values are merged together.

Constructors

SQLitePendingData

Fields

_pendingTableCreation :: !SQLitePendingTableCreations

_pendingWrites :: !SQLitePendingWrites

_pendingTxLogMap :: !TxLogMap

_pendingSuccessfulTxs :: !SQLitePendingSuccessfulTxs

pendingTableCreation :: Lens' SQLitePendingData SQLitePendingTableCreations

pendingWrites :: Lens' SQLitePendingData SQLitePendingWrites

pendingTxLogMap :: Lens' SQLitePendingData TxLogMap

pendingSuccessfulTxs :: Lens' SQLitePendingData SQLitePendingSuccessfulTxs

emptySQLitePendingData :: SQLitePendingData

data BlockState

Monad state for 'BlockHandler.

Constructors

BlockState

Fields

_bsTxId :: !TxId

_bsMode :: !(Maybe ExecutionMode)

_bsBlockHeight :: !BlockHeight

_bsPendingBlock :: !SQLitePendingData

_bsPendingTx :: !(Maybe SQLitePendingData)

initBlockState :: BlockState

bsBlockHeight :: Lens' BlockState BlockHeight

bsMode :: Lens' BlockState (Maybe ExecutionMode)

bsTxId :: Lens' BlockState TxId

bsPendingBlock :: Lens' BlockState SQLitePendingData

bsPendingTx :: Lens' BlockState (Maybe SQLitePendingData)

data BlockEnv p

Constructors

BlockEnv

Fields

_benvDb :: !(BlockDbEnv p)

_benvBlockState :: !BlockState

The current block state.

benvBlockState :: forall p. Lens' (BlockEnv p) BlockState

benvDb :: forall p p. Lens (BlockEnv p) (BlockEnv p) (BlockDbEnv p) (BlockDbEnv p)

data SQLiteEnv

Constructors

SQLiteEnv

Fields

_sConn :: !Database

_sConfig :: !SQLiteConfig

sConn :: Lens' SQLiteEnv Database

sConfig :: Lens' SQLiteEnv SQLiteConfig

newtype BlockHandler p a

Constructors

BlockHandler

Fields

runBlockHandler :: ReaderT (BlockDbEnv p) (StateT BlockState IO) a

type ParentHash = BlockHash

data BlockDbEnv p

Constructors

BlockDbEnv

Fields

_bdbenvDb :: !p

_logger :: !Logger

bdbenvDb :: forall p p. Lens (BlockDbEnv p) (BlockDbEnv p) p p

newtype SQLiteFlag

Constructors

SQLiteFlag

Fields

getFlag :: CInt

mempool

data MemPoolAccess

Constructors

MemPoolAccess

Fields

mpaGetBlock :: MempoolPreBlockCheck ChainwebTransaction -> BlockHeight -> BlockHash -> BlockHeader -> IO (Vector ChainwebTransaction)

mpaSetLastHeader :: BlockHeader -> IO ()

mpaProcessFork :: BlockHeader -> IO ()

pact service monad + types

data PactServiceEnv cas

Constructors

PactServiceEnv

Fields

_psMempoolAccess :: !(Maybe MemPoolAccess)

_psCheckpointEnv :: !CheckpointEnv

_psSpvSupport :: !SPVSupport

_psPublicData :: !PublicData

_psPdb :: !(PayloadDb cas)

_psBlockHeaderDb :: !BlockHeaderDb

_psGasModel :: !GasModel

_psMinerRewards :: !MinerRewards

newtype PactServiceState

Constructors

PactServiceState

Fields

_psStateValidated :: Maybe BlockHeader

type PactServiceM cas = ReaderT (PactServiceEnv cas) (StateT PactServiceState IO)

runPactServiceM :: PayloadCas cas => PactServiceState -> PactServiceEnv cas -> PactServiceM cas a -> IO (a, PactServiceState)

data PactServiceException

Constructors

PactServiceIllegalRewind

Fields

_attemptedRewindTo :: Maybe (BlockHeight, BlockHash)

_latestBlock :: Maybe (BlockHeight, BlockHash)

optics

psMempoolAccess :: forall cas. Lens' (PactServiceEnv cas) (Maybe MemPoolAccess)

psCheckpointEnv :: forall cas. Lens' (PactServiceEnv cas) CheckpointEnv

psSpvSupport :: forall cas. Lens' (PactServiceEnv cas) SPVSupport

psPublicData :: forall cas. Lens' (PactServiceEnv cas) PublicData

psStateValidated :: Iso' PactServiceState (Maybe BlockHeader)

psPdb :: forall cas cas. Lens (PactServiceEnv cas) (PactServiceEnv cas) (PayloadDb cas) (PayloadDb cas)

psBlockHeaderDb :: forall cas. Lens' (PactServiceEnv cas) BlockHeaderDb

psMinerRewards :: forall cas. Lens' (PactServiceEnv cas) MinerRewards

psGasModel :: forall cas. Lens' (PactServiceEnv cas) GasModel