Chainweb.Pact.Backend.Utils

Description

Documentation

runBlockEnv :: MVar (BlockEnv SQLiteEnv) -> BlockHandler SQLiteEnv a -> IO a

callDb :: (MonadCatch m, MonadReader (BlockDbEnv SQLiteEnv) m, MonadIO m) => Text -> (Database -> IO b) -> m b

withSavepoint :: SavepointName -> BlockHandler SQLiteEnv a -> BlockHandler SQLiteEnv a

withTrace :: Database -> (Utf8 -> IO ()) -> IO a -> IO a

beginSavepoint :: SavepointName -> BlockHandler SQLiteEnv ()

commitSavepoint :: SavepointName -> BlockHandler SQLiteEnv ()

rollbackSavepoint :: SavepointName -> BlockHandler SQLiteEnv ()

rollbackSavepoint n rolls back all database updates since the most recent savepoint with the name n and restarts the transaction.

NOTE that the savepoint is not removed from the savepoint stack. In order to also remove the savepoint rollbackSavepoint n >> commitSavepoint n can be used to release the (empty) transaction.

Cf. https://www.sqlite.org/lang_savepoint.html for details about savepoints.

data SavepointName

Constructors

BatchSavepoint

Block

DbTransaction

PreBlock

withBlockEnv :: a -> (MVar a -> IO b) -> IO b

readBlockEnv :: (a -> b) -> MVar a -> IO b

withSQLiteConnection :: String -> [Pragma] -> Bool -> (SQLiteEnv -> IO c) -> IO c

openSQLiteConnection :: String -> [Pragma] -> IO SQLiteEnv

closeSQLiteConnection :: SQLiteEnv -> IO ()

withTempSQLiteConnection :: [Pragma] -> (SQLiteEnv -> IO c) -> IO c

domainTableName :: Domain k v -> Utf8

convKeySetName :: KeySetName -> Utf8

convModuleName :: ModuleName -> Utf8

convNamespaceName :: NamespaceName -> Utf8

convRowKey :: RowKey -> Utf8

convPactId :: PactId -> Utf8

expectSingleRowCol :: Show a => String -> [[a]] -> IO a

expectSingle :: Show a => String -> [a] -> IO a

chainwebPragmas :: [Pragma]

open2 :: String -> IO (Either (Error, Utf8) Database)

collapseFlags :: [SQLiteFlag] -> SQLiteFlag

sqlite_open_readwrite :: SQLiteFlag

sqlite_open_create :: SQLiteFlag

sqlite_open_fullmutex :: SQLiteFlag

execMulti :: Traversable t => Database -> Utf8 -> t [SType] -> IO ()

Orphan instances

StringConv ByteString Utf8

Instance details

Methods

strConv :: Leniency -> ByteString -> Utf8

StringConv String Utf8

Instance details

Methods

strConv :: Leniency -> String -> Utf8

StringConv Text Utf8

Instance details

Methods

strConv :: Leniency -> Text -> Utf8

StringConv Utf8 ByteString

Instance details

Methods

strConv :: Leniency -> Utf8 -> ByteString

StringConv Utf8 String

Instance details

Methods

strConv :: Leniency -> Utf8 -> String

StringConv Utf8 Text

Instance details

Methods

strConv :: Leniency -> Utf8 -> Text