Chainweb.Pact.PactService

Description

Pact service for Chainweb

For Chainweb

initialPayloadState :: PayloadCas cas => ChainwebVersion -> ChainId -> PactServiceM cas ()

execNewBlock :: PayloadCas cas => MemPoolAccess -> BlockHeader -> Miner -> BlockCreationTime -> PactServiceM cas PayloadWithOutputs

Note: The BlockHeader param here is the PARENT HEADER of the new block-to-be

execValidateBlock :: PayloadCas cas => BlockHeader -> PayloadData -> PactServiceM cas PayloadWithOutputs

Validate a mined block. Execute the transactions in Pact again as validation. Note: The BlockHeader here is the header of the block being validated.

execTransactions :: Maybe BlockHash -> Miner -> Vector ChainwebTransaction -> PactDbEnv' -> PactServiceM cas Transactions

initPactService :: Logger logger => PayloadCas cas => ChainwebVersion -> ChainId -> logger -> PactQueue -> MemPoolAccess -> MVar (CutDb cas) -> BlockHeaderDb -> PayloadDb cas -> Maybe FilePath -> Maybe NodeId -> Bool -> IO ()

readCoinAccount

Arguments

:: PactDbEnv'
pact db backend (sqlite)

-> Text
account name

-> IO (Maybe (T2 Decimal (Guard (Term Name))))

Read row from coin-table defined in coin contract, retrieving balance and keyset associated with account name

readAccountBalance

Arguments

:: PactDbEnv'
pact db backend (sqlite)

-> Text
account name

-> IO (Maybe Decimal)

Read row from coin-table defined in coin contract, retrieving balance associated with account name

readAccountGuard

Arguments

:: PactDbEnv'
pact db backend (sqlite)

-> Text
account name

-> IO (Maybe (Guard (Term Name)))

Read row from coin-table defined in coin contract, retrieving guard associated with account name

For Side-tooling

execNewGenesisBlock :: PayloadCas cas => Miner -> Vector ChainwebTransaction -> PactServiceM cas PayloadWithOutputs

only for use in generating genesis blocks in tools

initPactService' :: Logger logger => PayloadCas cas => ChainwebVersion -> ChainId -> logger -> SPVSupport -> BlockHeaderDb -> PayloadDb cas -> Maybe FilePath -> Maybe NodeId -> Bool -> PactServiceM cas a -> IO a

minerReward :: MinerRewards -> BlockHeight -> IO ParsedDecimal

Calculate miner reward. We want this to error hard in the case where block times have finally exceeded the 120-year range. Rewards are calculated at regular blockheight intervals.

See: 'rewards/miner_rewards.csv'