Chainweb.Pact.RestAPI.Client

Description

This module defines the client API for the main Pact service and its spv continuation proof endpoints.

Documentation

pactSpvApiClient_

Arguments

:: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v

=> KnownChainIdSymbol c

=> SpvRequest
Contains the chain id of the target chain id used in the 'target-chain' field of a cross-chain-transfer. Also contains the request key of of the cross-chain transfer tx request.

-> ClientM TransactionOutputProofB64

pactSpvApiClient

Arguments

:: ChainwebVersion

-> ChainId
the chain id of the source chain id used in the execution of a cross-chain-transfer.

-> SpvRequest
Contains the chain id of the target chain id used in the 'target-chain' field of a cross-chain-transfer. Also contains the request key of of the cross-chain transfer tx request.

-> ClientM TransactionOutputProofB64

pactPollApiClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Poll -> ClientM PollResponses

pactPollApiClient :: ChainwebVersion -> ChainId -> Poll -> ClientM PollResponses

pactListenApiClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => ListenerRequest -> ClientM ListenResponse

pactListenApiClient :: ChainwebVersion -> ChainId -> ListenerRequest -> ClientM ListenResponse

pactSendApiClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => SubmitBatch -> ClientM RequestKeys

pactSendApiClient :: ChainwebVersion -> ChainId -> SubmitBatch -> ClientM RequestKeys

pactLocalApiClient_ :: forall (v :: ChainwebVersionT) (c :: ChainIdT). KnownChainwebVersionSymbol v => KnownChainIdSymbol c => Command Text -> ClientM (CommandResult Hash)

pactLocalApiClient :: ChainwebVersion -> ChainId -> Command Text -> ClientM (CommandResult Hash)