Chainweb.Pact.SPV

Description

Pact Service SPV support

spv support

pactSPV

Arguments

:: forall cas. ChainId
chain id of the running pact service

-> MVar (CutDb cas)
handle into the cutdb

-> SPVSupport

Spv support for pact

verifySPV

Arguments

:: forall cas. ChainId

-> MVar (CutDb cas)
handle into the cut db

-> Text
TXOUT or TXIN - defines the type of proof used in validation

-> Object Name
the TransactionOutputProof object to validate

-> IO (Either Text (Object Name))

SPV transaction verification support. Calls to 'verify-spv' in Pact will thread through this function and verify an SPV receipt, making the requisite calls to the SPV api and verifying the output proof.

verifyCont

Arguments

:: forall cas. ChainId

-> MVar (CutDb cas)
handle into the cut db

-> ContProof
bytestring of 'TransactionOutputP roof' object to validate

-> IO (Either Text PactExec)

SPV defpact transaction verification support. This call validates a pact endorsement in Pact, providing a validation that the yield data of a cross-chain pact is valid.

spv api utilities

getTxIdx :: HasCallStack => PayloadCas cas => BlockHeaderDb -> PayloadDb cas -> BlockHeight -> PactHash -> IO (Either Text Int)

Look up pact tx hash at some block height in the payload db, and return the tx index for proof creation.

Note: runs in O(n) - this should be revisited if possible