Chainweb.Pact.TransactionExec

Description

Pact command execution and coin-contract transaction logic for Chainweb

Transaction Execution

applyCmd

Arguments

:: Logger
Pact logger

-> PactDbEnv p
Pact db environment

-> Miner
The miner chosen to mine the block

-> GasModel
Gas model (pact Service config)

-> PublicData
Contains block height, time, prev hash + metadata

-> SPVSupport
SPV support (validates cont proofs)

-> Command PayloadWithText
command with payload to execute

-> ModuleCache
cached module state

-> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

The main entry point to executing transactions. From here, applyCmd assembles the command environment for a command and orchestrates gas buys/redemption, and executing payloads.

applyGenesisCmd

Arguments

:: Logger
Pact logger

-> PactDbEnv p
Pact db environment

-> PublicData
Contains block height, time, prev hash + metadata

-> SPVSupport
SPV support (validates cont proofs)

-> Command (Payload PublicMeta ParsedCode)
command with payload to execute

-> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

applyLocal

Arguments

:: Logger
Pact logger

-> PactDbEnv p
Pact db environment

-> PublicData
Contains block height, time, prev hash + metadata

-> SPVSupport
SPV support (validates cont proofs)

-> Command (Payload PublicMeta ParsedCode)
command with payload to execute

-> IO (CommandResult [TxLog Value])

applyExec :: CommandEnv p -> Interpreter p -> RequestKey -> ExecMsg ParsedCode -> [Signer] -> Hash -> [TxLog Value] -> NamespacePolicy -> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

Execute an ExecMsg and Return the result with module cache

applyExec' :: CommandEnv p -> Interpreter p -> ExecMsg ParsedCode -> [Signer] -> Hash -> NamespacePolicy -> IO EvalResult

Variation on applyExec that returns EvalResult as opposed to wrapping it up in a JSON result.

applyContinuation :: CommandEnv p -> Interpreter p -> RequestKey -> ContMsg -> [Signer] -> Hash -> [TxLog Value] -> NamespacePolicy -> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

Execute a ContMsg and return the command result and module cache

applyContinuation' :: CommandEnv p -> Interpreter p -> ContMsg -> [Signer] -> Hash -> NamespacePolicy -> IO EvalResult

Execute a ContMsg and return just eval result, not wrapped in a CommandResult wrapper

runPayload :: CommandEnv p -> Interpreter p -> Command (Payload PublicMeta ParsedCode) -> [TxLog Value] -> NamespacePolicy -> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

Gas Execution

buyGas :: CommandEnv p -> Command (Payload PublicMeta ParsedCode) -> Miner -> GasSupply -> ModuleCache -> IO (Either Text (T3 GasId [TxLog Value] ModuleCache))

Build and execute 'coin.buygas' command from miner info and user command info (see applyCmd)

see: 'pactcoin-contractcoin.pact#fund-tx'

mkBuyGasCmd

Arguments

:: MinerId
Id of the miner to fund

-> MinerKeys
Miner keyset

-> Text
Address of the sender from the command

-> GasSupply
The gas limit total * price

-> IO (ExecMsg ParsedCode)

Build the 'coin-contract.buygas' command

Coinbase Execution

applyCoinbase

Arguments

:: Logger
Pact logger

-> PactDbEnv p
Pact db environment

-> Miner
The miner chosen to mine the block

-> ParsedDecimal
Miner reward

-> PublicData
Contains block height, time, prev hash + metadata

-> BlockHash
hash of the mined block

-> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

mkCoinbaseCmd :: MinerId -> MinerKeys -> ParsedDecimal -> IO (ExecMsg ParsedCode)

Command Helpers

publicMetaOf :: Command (Payload PublicMeta c) -> PublicMeta

Retrieve public metadata from a command

networkIdOf :: Command (Payload a b) -> Maybe NetworkId

Retrieve the optional Network identifier from a command

gasFeeOf :: Gas -> GasPrice -> GasSupply

Calculate the gas fee (pact-generate gas cost * user-specified gas price), rounding to the nearest stu.

gasPriceOf :: forall c. Command (Payload PublicMeta c) -> GasPrice

Get the gas price of a public chain command payload

mkGasEnvOf :: Command (Payload PublicMeta c) -> GasModel -> GasEnv

Create a gas environment from a verified command

Utilities

buildExecParsedCode :: Maybe Value -> Text -> IO (ExecMsg ParsedCode)

Build the ExecMsg for some pact code fed to the function. The value parameter is for any possible environmental data that needs to go into the ExecMsg.

jsonErrorResult :: CommandEnv a -> RequestKey -> PactError -> [TxLog Value] -> Gas -> ModuleCache -> String -> IO (T2 (CommandResult [TxLog Value]) ModuleCache)

Present a failure as a pair of json result of Command Error and associated logs

logDebugRequestKey :: Logger -> RequestKey -> String -> IO ()

Log request keys at DEBUG when successful

logErrorRequestKey :: Exception e => Logger -> RequestKey -> e -> String -> IO ()

Log request keys and error message at ERROR when failed

mkMagicCapSlot :: Text -> CapSlot UserCapability

Builder for "magic" capabilities given a magic cap name