Chainweb.Payload

Description

TODO

Block Chain Data

newtype Transaction

An encoded transaction, including all of its inputs.

We don't care about the encoding of a transaction. The semantics of a transaction is only known to pact.

Constructors

Transaction

Fields

_transactionBytes :: ByteString

newtype TransactionOutput

Encoded output of a single transaction.

We don't care about the encoding of the output. The semantics of a output is only known to pact.

Constructors

TransactionOutput

Fields

_transactionOutputBytes :: ByteString

Hashes

newtype BlockPayloadHash

Constructors

BlockPayloadHash MerkleLogHash

encodeBlockPayloadHash :: MonadPut m => BlockPayloadHash -> m ()

decodeBlockPayloadHash :: MonadGet m => m BlockPayloadHash

newtype BlockTransactionsHash

Constructors

BlockTransactionsHash MerkleLogHash

encodeBlockTransactionsHash :: MonadPut m => BlockTransactionsHash -> m ()

decodeBlockTransactionsHash :: MonadGet m => m BlockTransactionsHash

newtype BlockOutputsHash

Constructors

BlockOutputsHash MerkleLogHash

encodeBlockOutputsHash :: MonadPut m => BlockOutputsHash -> m ()

decodeBlockOutputsHash :: MonadGet m => m BlockOutputsHash

Authoritative Data

data BlockPayload

The Payload of a block.

The transactions of a block at a given height in the chain are discovered by _blockPayloadTransactionsHash . _blockHeaderPayloadHash.

NOTES:

The block and its output are uniquely determined by the transactions. However, we want to include the output hash in the payload hash as a checksum and for inclusion proofs. In order to validate the payload hash without recomputing the outputs we need this extra level of indirection between the BlockHeaders and the Transactions.

This structure could be recomputed from the block transactions only if it would be possible to discover the order of blocks from the block transactions structure which isn't the case. Instead the order of the blocks is discovered from the BlockHeader chain and block transactions for a given block height are looked up by _blockPayloadTransactionsHash . _blockPayloadPayloadHash.

Constructors

BlockPayload

Fields

_blockPayloadPayloadHash :: !BlockPayloadHash

Hash of _blockPayloadTransactionsHash and _blockPayloadTransactionsHash. Primary key of BlockPayloadStore.

_blockPayloadTransactionsHash :: !BlockTransactionsHash

Root of TransactionTree of the block. Foreign key into BlockTransactionsStore and TransactionTreeStore.

_blockPayloadOutputsHash :: !BlockOutputsHash

Root of OutputsTree of the block. Foreign key into BlockOutputsStore and OutputTreeStore.

data BlockTransactions

The block transactions

Constructors

BlockTransactions

Fields

_blockTransactionsHash :: !BlockTransactionsHash

Root of TransactionTree of the block. Primary key of BlockTransactionsStore. Foreign key into TransactionTreeStore.

_blockTransactions :: !(Vector Transaction)

Ordered list of all transactions of the block.

_blockMinerData :: !MinerData

Miner data for rewards

Redundant Data / Caches

data BlockOutputs

All outputs of the transactions of a block.

NOTE: the block outputs are associated with the respective block in the BlockPayload structure for the block.

Constructors

BlockOutputs

Fields

_blockOutputsHash :: !BlockOutputsHash

Root of OutputTree of the block. Primary key of BlockOutputsStore. Foreign key into OutputTreeStore.

_blockOutputs :: !(Vector TransactionOutput)

Output of all transactions of a block in the order of the transactions in the block.

_blockCoinbaseOutput :: !CoinbaseOutput

Output of coinbase transaction.

data TransactionTree

Merkle tree for the transactions in a block.

Constructors

TransactionTree

Fields

_transactionTreeHash :: !BlockTransactionsHash

Root of _transactionTree. Primary key of 'TransactionTreeStore. Foreign key into BlockTransactionsStore.

_transactionTree :: !(MerkleTree SHA512t_256)

data OutputTree

Merkle Tree for transaction outputs of a block.

Constructors

OutputTree

Fields

_outputTreeHash :: !BlockOutputsHash

Root of _outputTree. Primary key of 'OutputTreeStore. Foreign key into BlockOutputsStore.

_outputTree :: !(MerkleTree SHA512t_256)

Create Data

type BlockTransactionsLog = MkLogType ChainwebHashTag BlockTransactions

newTransactionLog :: MinerData -> Vector Transaction -> BlockTransactionsLog

This forces the MerkleTree which can be an expensive operation.

newBlockTransactions :: MinerData -> Vector Transaction -> (TransactionTree, BlockTransactions)

This forces the MerkleTree which can be an expensive operation.

transactionLog :: BlockTransactions -> TransactionTree -> BlockTransactionsLog

Create a MerkleLog from a BlockTransactions and a cached TransactionTree.

It is an error if the TransactionTree doesn't match the BlockTransactions.

If you only have the BlockTransactions you can use toLog in which case the _merkleLogTree will be instantiated lazily.

newtype MinerData

Constructors

MinerData

Fields

_minerData :: ByteString

newtype CoinbaseOutput

Constructors

CoinbaseOutput

Fields

_coinbaseOutput :: ByteString

type BlockOutputsLog = MkLogType ChainwebHashTag BlockOutputs

newBlockOutputLog :: CoinbaseOutput -> Vector TransactionOutput -> BlockOutputsLog

This forces the MerkleTree which can be an expensive operation.

newBlockOutputs :: CoinbaseOutput -> Vector TransactionOutput -> (OutputTree, BlockOutputs)

This forces the MerkleTree which can be an expensive operation.

blockOutputLog :: BlockOutputs -> OutputTree -> BlockOutputsLog

Create a MerkleLog from BlockOutputs and a cached OutputTree.

It is an error if the OutputTree doesn't match the BlockOutputs.

If you only have the BlockOutputs you can use toLog in which case the _merkleLogTree will be instantiated lazily.

blockPayload :: BlockTransactions -> BlockOutputs -> BlockPayload

Note, that forcing the MerkleTree is cheap for BlockPayload.

newBlockPayload :: MinerData -> CoinbaseOutput -> Vector (Transaction, TransactionOutput) -> BlockPayload

API Payload Data

data PayloadData

This contains all non-redundant payload data for a block. It doesn't contain any data that can be recomputed.

This data structure is used maintly to transfer payloads over the wire.

Constructors

PayloadData

Fields

_payloadDataTransactions :: !(Vector Transaction)

_payloadDataMiner :: !MinerData

_payloadDataPayloadHash :: !BlockPayloadHash

_payloadDataTransactionsHash :: !BlockTransactionsHash

_payloadDataOutputsHash :: !BlockOutputsHash

payloadData :: BlockTransactions -> BlockPayload -> PayloadData

newPayloadData :: BlockTransactions -> BlockOutputs -> PayloadData

type PayloadDataCas cas = CasConstraint cas PayloadData

All Payload Data in a Single Structure

data PayloadWithOutputs

Constructors

PayloadWithOutputs

Fields

_payloadWithOutputsTransactions :: !(Vector (Transaction, TransactionOutput))

_payloadWithOutputsMiner :: !MinerData

_payloadWithOutputsCoinbase :: !CoinbaseOutput

_payloadWithOutputsPayloadHash :: !BlockPayloadHash

_payloadWithOutputsTransactionsHash :: !BlockTransactionsHash

_payloadWithOutputsOutputsHash :: !BlockOutputsHash

payloadWithOutputs :: HasCallStack => PayloadData -> CoinbaseOutput -> Vector TransactionOutput -> PayloadWithOutputs

Smart constructor for PayloadWithOutputs.

Precondition: the vector of transaction output has the same length (and is in the same order, i.e. the two vectors will be zipped) as the list of input transactions inside the PayloadData.

newPayloadWithOutputs :: MinerData -> CoinbaseOutput -> Vector (Transaction, TransactionOutput) -> PayloadWithOutputs

payloadWithOutputsToBlockObjects :: PayloadWithOutputs -> (BlockTransactions, BlockOutputs)

payloadWithOutputsToPayloadData :: PayloadWithOutputs -> PayloadData