Chainweb.Payload.PayloadStore.Types

Description

Transaction Database

newtype BlockTransactionsStore cas

Store of the BlockTransactions for all blocks.

Primary Key: _blockTransactionsHash

Constructors

BlockTransactionsStore (cas BlockTransactions)

newtype BlockPayloadStore cas

Store of the BlockPayloads for all blocks

Primary Key: _blockPayloadHash

Constructors

BlockPayloadStore (cas BlockPayload)

data TransactionDb cas

The authoritative CAS stores for a block chain.

Constructors

TransactionDb

Fields

_transactionDbBlockTransactions :: !(BlockTransactionsStore cas)

The block transactions of the block chain. This data is strictly needed to rebuild the payload data.

_transactionDbBlockPayloads :: !(BlockPayloadStore cas)

While the content of this store can be computed from the block transactions, it is needed as an index into the block transaction store. If it would be lost one would have to recompute all of it in order to look up the data for a single transation.

type TransactionDbCas cas = (CasConstraint cas BlockPayload, CasConstraint cas BlockTransactions)

transactionDbBlockTransactions :: forall cas. Lens' (TransactionDb cas) (BlockTransactionsStore cas)

transactionDbBlockPayloads :: forall cas. Lens' (TransactionDb cas) (BlockPayloadStore cas)

Caches

newtype OutputTreeStore cas

Store of the OutputTree Merkle trees for all blocks.

Constructors

OutputTreeStore (cas OutputTree)

newtype TransactionTreeStore cas

Store of the TransactionTree Merkle trees for all blocks.

Constructors

TransactionTreeStore (cas TransactionTree)

newtype BlockOutputsStore cas

Store of the BlockOutputs for all blocks.

Constructors

BlockOutputsStore (cas BlockOutputs)

data PayloadCache cas

The CAS caches for a block chain.

If an entry is missing it can be rebuild from the PayloadDb

Constructors

PayloadCache

Fields

_payloadCacheBlockOutputs :: !(BlockOutputsStore cas)

This is relatively expensive to rebuild.

_payloadCacheTransactionTrees :: !(TransactionTreeStore cas)

This are relatively cheap to rebuild. (tens of thousands of blocks per second)

_payloadCacheOutputTrees :: !(OutputTreeStore cas)

This are relatively cheap to rebuild. (tens of thousands blocks per second)

type PayloadCacheCas cas = (CasConstraint cas BlockOutputs, CasConstraint cas TransactionTree, CasConstraint cas OutputTree)

payloadCacheBlockOutputs :: forall cas. Lens' (PayloadCache cas) (BlockOutputsStore cas)

payloadCacheOutputTrees :: forall cas. Lens' (PayloadCache cas) (OutputTreeStore cas)

payloadCacheTransactionTrees :: forall cas. Lens' (PayloadCache cas) (TransactionTreeStore cas)

Payload Database

data PayloadDb cas

Constructors

PayloadDb

Fields

_transactionDb :: !(TransactionDb cas)

_payloadCache :: !(PayloadCache cas)

type PayloadCas cas = (CasConstraint cas BlockOutputs, CasConstraint cas TransactionTree, CasConstraint cas OutputTree, CasConstraint cas BlockTransactions, CasConstraint cas BlockPayload)

payloadCache :: forall cas. Lens' (PayloadDb cas) (PayloadCache cas)

transactionDb :: forall cas. Lens' (PayloadDb cas) (TransactionDb cas)

Exceptions

newtype PayloadNotFoundException

Constructors

PayloadNotFoundException BlockPayloadHash

insert new payload

addPayload :: PayloadCas cas => PayloadDb cas -> BlockTransactions -> TransactionTree -> BlockOutputs -> OutputTree -> IO ()

Insert block payload data in to the database.

addNewPayload :: PayloadCas cas => PayloadDb cas -> PayloadWithOutputs -> IO ()

Create block payload data from a sequence of transaction and insert it into the database.