Chainweb.RestAPI.NetworkID

Description

TODO

Documentation

data NetworkId

Constructors

ChainNetwork ChainId

MempoolNetwork ChainId

CutNetwork

networkIdToText :: NetworkId -> Text

networkIdFromText :: MonadThrow m => Text -> m NetworkId

unsafeNetworkIdFromText :: HasCallStack => Text -> NetworkId

pNetworkId :: OptionParser NetworkId

Typelevel NetworkId

data NetworkIdT

Constructors

ChainNetworkT ChainIdT

MempoolNetworkT ChainIdT

CutNetworkT

data SomeNetworkIdT where

SomeNetworkIdT encapsulates a known NetworkIdT, such that its type is hidden from the compiler.

Constructors

SomeChainNetworkT :: forall (a :: ChainIdT) (n :: NetworkIdT). (KnownChainIdSymbol a, n ~ 'ChainNetworkT a) => Proxy n -> SomeNetworkIdT

SomeMempoolNetworkT :: forall (a :: ChainIdT) (n :: NetworkIdT). (KnownChainIdSymbol a, n ~ 'MempoolNetworkT a) => Proxy n -> SomeNetworkIdT

SomeCutNetworkT :: Proxy 'CutNetworkT -> SomeNetworkIdT

someNetworkIdVal :: NetworkId -> SomeNetworkIdT

class KnownNetworkId (n :: NetworkIdT) where

Methods

networkIdVal :: Proxy n -> SomeNetworkIdT

Singletons

data family Sing :: k -> Type

Kind-indexed family of singleton types.

In order to define singletons for types of a data kind A, one defines an instance that has a value constructor for each type constructor of A. The following is an example how to define singletons for types of kind Maybe a:

data instance Sing :: forall a . Maybe a -> Type where
   SNothing :: Sing 'Nothing
   SJust :: Sing a -> Sing ('Just a)

type SNetwork (n :: NetworkIdT) = Sing n