Chainweb.RestAPI.Utils

Description

Utils for Chainweb REST APIs

Servant Utils

type Reassoc (api :: Type) = ReassocBranch api '[]

API Version

type Version = "0.0"

newtype ApiVersion

Constructors

ApiVersion Text

apiVersion :: ApiVersion

prettyApiVersion :: Text

Paging

type PageParams k = LimitParam :> NextParam k

Pages Parameters

limit :: Natural

next :: k

type LimitParam = QueryParam "limit" Limit

type NextParam k = QueryParam "next" k

Chainweb API Endpoints

newtype ChainwebEndpoint

Constructors

ChainwebEndpoint ChainwebVersionT

newtype NetworkEndpoint

Constructors

NetworkEndpoint NetworkIdT

type ChainEndpoint (c :: ChainIdT) = 'NetworkEndpoint ('ChainNetworkT c)

type MempoolEndpoint (c :: ChainIdT) = 'NetworkEndpoint ('MempoolNetworkT c)

type CutEndpoint = 'NetworkEndpoint 'CutNetworkT

Some API

The chain graph and thus the list of chain ids is encoded as runtime values. In order to combin these in statically defined APIs we reify the the chainweb version and chainid as types and wrap them existenially so that they can be passed around and be combined.

data SomeApi

Constructors

forall (a :: Type).(HasSwagger a, HasServer a '[], HasClient ClientM a) => SomeApi (Proxy a)

someApi :: forall proxy (a :: Type). HasServer a '[] => HasClient ClientM a => HasSwagger a => proxy a -> SomeApi

Some Server

data SomeServer

Constructors

forall (a :: Type).HasServer a '[] => SomeServer (Proxy a) (Server a)

someServerApplication :: SomeServer -> Application

Misc Utils

bindPortTcp :: Port -> HostPreference -> IO (Port, Socket)