Chainweb.SPV

Description

TODO

Documentation

data SpvException

Constructors

SpvExceptionTargetNotReachable

Fields

_spvExceptionMsg :: !Text

_spvExceptionSourceChainId :: !ChainId

_spvExceptionSourceHeight :: !BlockHeight

_spvExceptionTargetChainId :: !ChainId

_spvExceptionTargetHeight :: !BlockHeight

SpvExceptionInconsistentPayloadData

Fields

_spvExceptionMsg :: !Text

_spvExceptionMsgPayloadHash :: !BlockPayloadHash

SpvExceptionVerificationFailed

Fields

_spvExceptionMsg :: !Text

data TransactionProof a

Witness that a transaction is included in the head of a chain in a chainweb.

Constructors

TransactionProof ChainId (MerkleProof a)

proofChainId :: Getter (TransactionProof a) ChainId

Getter into the chain id of a TransactionProof

data TransactionOutputProof a

Witness that a transaction output is included in the head of a chain in a chainweb.

Constructors

TransactionOutputProof ChainId (MerkleProof a)

outputProofChainId :: Getter (TransactionOutputProof a) ChainId

Getter into the chain id of a TransactionOutputProof