Chainweb.SPV.CreateProof

Description

TODO

Documentation

createTransactionProof

Arguments

:: HasCallStack

=> PayloadCas cas

=> CutDb cas
Block Header Database

-> ChainId
target chain. The proof asserts that the subject is included in the head of this chain.

-> ChainId
source chain. This the chain of the subject

-> BlockHeight
The block height of the transaction

-> Int
The index of the transaction in the block

-> IO (TransactionProof SHA512t_256)

Creates a witness that a transaction in is included in the head of a chain in a chainweb.

createTransactionOutputProof

Arguments

:: HasCallStack

=> PayloadCas cas

=> CutDb cas
Block Header Database

-> ChainId
target chain. The proof asserts that the subject is included in the head of this chain.

-> ChainId
source chain. This the chain of the subject

-> BlockHeight
The block height of the transaction

-> Int
The index of the transaction in the block

-> IO (TransactionOutputProof SHA512t_256)

Creates a witness that a transaction in is included in the head of a chain in a chainweb.