Chainweb.SPV.RestAPI.Client

Description

TODO

Documentation

spvGetTransactionProofClient

Arguments

:: ChainwebVersion

-> ChainId
the target chain of the proof. This is the chain for which inclusion is proved.

-> ChainId
the source chain of the proof. This is the chain where the proof subject, the transaction for which inclusion is proven, is located.

-> BlockHeight
the block height of the proof subject, the transaction for which inclusion is proven.

-> Natural
the index of the proof subject, the transaction for which inclusion is proven.

-> ClientM (TransactionProof SHA512t_256)

spvGetTransactionOutputProofClient

Arguments

:: ChainwebVersion

-> ChainId
the target chain of the proof. This is the chain for which inclusion is proved.

-> ChainId
the source chain of the proof. This is the chain where the proof subject, the transaction output for which inclusion is proven, is located.

-> BlockHeight
the block height of the proof subject, the transaction output for which inclusion is proven.

-> Natural
the index of the proof subject, the transaction output for which inclusion is proven.

-> ClientM (TransactionOutputProof SHA512t_256)