Chainweb.SPV.RestAPI.Server

Description

TODO

Documentation

spvGetTransactionProofHandler

Arguments

:: PayloadCas cas

=> CutDb cas

-> ChainId
the target chain of the proof. This is the chain for which inclusion is proved.

-> ChainId
the source chain of the proof. This is the chain where the proof subject, the transaction for which inclusion is proven, is located.

-> BlockHeight
the block height of the proof subject, the transaction for which inclusion is proven.

-> Natural
the index of the proof subject, the transaction for which inclusion is proven.

-> Handler (TransactionProof SHA512t_256)

spvGetTransactionOutputProofHandler

Arguments

:: PayloadCas cas

=> CutDb cas

-> ChainId
the target chain of the proof. This is the chain for which inclusion is proved.

-> ChainId
the source chain of the proof. This is the chain where the proof subject, the transaction output for which inclusion is proven, is located.

-> BlockHeight
the block height of the proof subject, the transaction output for which inclusion is proven.

-> Natural
the index of the proof subject, the transaction output for which inclusion is proven.

-> Handler (TransactionOutputProof SHA512t_256)

spvServer :: forall cas v (c :: ChainIdT). PayloadCas cas => KnownChainIdSymbol c => CutDbT cas v -> Server (SpvApi v c)

spvApp :: forall cas v c. PayloadCas cas => KnownChainwebVersionSymbol v => KnownChainIdSymbol c => CutDbT cas v -> Application

spvApiLayout :: forall cas v c. KnownChainwebVersionSymbol v => KnownChainIdSymbol c => CutDbT cas v -> IO ()

someSpvServer :: forall cas c. PayloadCas cas => KnownChainIdSymbol c => SomeCutDb cas -> SomeServer

someSpvServers :: PayloadCas cas => ChainwebVersion -> CutDb cas -> SomeServer