Chainweb.SPV.VerifyProof

Description

TODO

Transaction Proofs

runTransactionProof :: TransactionProof SHA512t_256 -> BlockHash

Runs a transaction Proof. Returns the block hash on the target chain for which inclusion is proven.

verifyTransactionProof :: CutDb cas -> TransactionProof SHA512t_256 -> IO Transaction

Transaction Output Proofs

runTransactionOutputProof :: TransactionOutputProof SHA512t_256 -> BlockHash

Runs a transaction Proof. Returns the block hash on the target chain for which inclusion is proven.

verifyTransactionOutputProof :: CutDb cas -> TransactionOutputProof SHA512t_256 -> IO TransactionOutput