Chainweb.Store.CAS.FS

Documentation

data FsDB t

A content-addressed store for block payloads using a direct filesystem representation. Block payload hashes are converted to hex strings and mapped to files rooted at the given directory. We use a 3-deep nesting, e.g. block payload hash

 f22136124cd3e1d65a48487cecf310771b2fd1e83dc032e3d19724160ac0ff71

would be stored at

 $ROOT/f22/136/124cd3e1d65a48487cecf310771b2fd1e83dc032e3d19724160ac0ff71

Filesystem operations are abstracted by passing an FsOps record; this will allow us to plug in a backend using S3 later.

Constructors

FsDB

Fields

_fsDbRoot :: Path Absolute

_fsDbConfig :: !(PayloadConfig t)

_fsOps :: !FsOps

_fsTP :: !ThreadPool

data FsOps

Filesystem operations for the content-addressed store backend. Use systemFsOps to use the local filesystem.

Constructors

FsOps

Fields

opDoesFileExist :: FilePath -> IO Bool

does a file at this path exist?

opMakeAbsolutePath :: FilePath -> IO (Path Absolute)

makes the given path absolute.

opCreateDirectory :: FilePath -> IO ()

makes a directory and all of its parents.

opWriteAtomic :: FilePath -> ByteString -> IO ()

atomic-writes the given contents to the filesystem. On local filesystems this means a write to a temporary file and an atomic-rename.

opDeleteFile :: FilePath -> IO ()

deletes the given file, silently.

opReadFile :: FilePath -> IO ByteString

strictly reads the given path into a bytestring.

withDB

Arguments

:: FilePath
db root

-> PayloadConfig t
description of payloads

-> (DB t -> IO b)
user handler

-> IO b

Given a root directory and a payload config, generates a block payload DB and runs the given user handler with it.

withDB'

Arguments

:: FsOps
filesystem abstraction

-> FilePath
db root

-> PayloadConfig t
description of payloads

-> (DB t -> IO b)
user handler

-> IO b

Given a filesystem implementation, a root directory, and a payload config, generates a block payload DB and runs the given user handler with it.

systemFsOps :: IO FsOps