Chainweb.Sync.WebBlockHeaderStore

Description

TODO

Documentation

data WebBlockHeaderStore

In order to use this a processor for the queue is needed.

The module P2P.TaskQueue provides a P2P session that serves the queue.

TODO

Find a better name

Parameterize in cas

This is currently based on TreeDB (for API) and BlockHeaderDB, it would be possible to run this on top of any CAS and API that offers a simple GET.

Constructors

WebBlockHeaderStore

Fields

_webBlockHeaderStoreCas :: !WebBlockHeaderDb

_webBlockHeaderStoreMemo :: !(TaskMap (ChainValue BlockHash) (ChainValue BlockHeader))

_webBlockHeaderStoreQueue :: !(PQueue (Task ClientEnv (ChainValue BlockHeader)))

_webBlockHeaderStoreLogFunction :: !LogFunction

_webBlockHeaderStoreMgr :: !Manager

newWebBlockHeaderStore :: Manager -> WebBlockHeaderDb -> LogFunction -> IO WebBlockHeaderStore

getBlockHeader :: PayloadCas cas => BlockHeaderCas candidateHeaderCas => PayloadDataCas candidatePayloadCas => WebBlockHeaderStore -> WebBlockPayloadStore cas -> candidateHeaderCas BlockHeader -> candidatePayloadCas PayloadData -> ChainId -> Priority -> Maybe PeerInfo -> BlockHash -> IO BlockHeader

data WebBlockPayloadStore cas

Constructors

WebBlockPayloadStore

Fields

_webBlockPayloadStoreCas :: !(PayloadDb cas)

Cas for storing complete payload data including outputs.

_webBlockPayloadStoreMemo :: !(TaskMap BlockPayloadHash PayloadData)

Internal memo table for active tasks

_webBlockPayloadStoreQueue :: !(PQueue (Task ClientEnv PayloadData))

task queue for scheduling tasks with the task server

_webBlockPayloadStoreLogFunction :: !LogFunction

LogFunction

_webBlockPayloadStoreMgr :: !Manager

Manager object for making HTTP requests

_webBlockPayloadStorePact :: !WebPactExecutionService

handle to the pact execution service for validating transactions and computing outputs.

newEmptyWebPayloadStore :: PayloadCas cas => ChainwebVersion -> Manager -> WebPactExecutionService -> LogFunction -> PayloadDb cas -> IO (WebBlockPayloadStore cas)

newWebPayloadStore :: Manager -> WebPactExecutionService -> PayloadDb cas -> LogFunction -> IO (WebBlockPayloadStore cas)

Utils

memoInsert :: IsCas a => Hashable (CasKeyType (CasValueType a)) => a -> TaskMap (CasKeyType (CasValueType a)) (CasValueType a) -> CasKeyType (CasValueType a) -> (CasKeyType (CasValueType a) -> IO (CasValueType a)) -> IO (CasValueType a)

data PactExecutionService

Constructors

PactExecutionService

Fields

_pactValidateBlock :: BlockHeader -> PayloadData -> IO PayloadWithOutputs

_pactNewBlock :: Miner -> BlockHeader -> BlockCreationTime -> IO PayloadWithOutputs

_pactLocal :: ChainwebTransaction -> IO (Either PactException HashCommandResult)

_pactLookup :: Either ChainId BlockHeader -> Vector PactHash -> IO (Either PactException (Vector (Maybe (T2 BlockHeight BlockHash))))

Orphan instances

IsCas WebBlockHeaderCas

Instance details

Associated Types

type CasValueType WebBlockHeaderCas

Methods

casLookup :: WebBlockHeaderCas -> CasKeyType (CasValueType WebBlockHeaderCas) -> IO (Maybe (CasValueType WebBlockHeaderCas))

casInsert :: WebBlockHeaderCas -> CasValueType WebBlockHeaderCas -> IO ()

casDelete :: WebBlockHeaderCas -> CasKeyType (CasValueType WebBlockHeaderCas) -> IO ()

casLookupBatch :: WebBlockHeaderCas -> Vector (CasKeyType (CasValueType WebBlockHeaderCas)) -> IO (Vector (Maybe (CasValueType WebBlockHeaderCas)))

casInsertBatch :: WebBlockHeaderCas -> Vector (CasValueType WebBlockHeaderCas) -> IO ()

casDeleteBatch :: WebBlockHeaderCas -> Vector (CasKeyType (CasValueType WebBlockHeaderCas)) -> IO ()

casMember :: WebBlockHeaderCas -> CasKeyType (CasValueType WebBlockHeaderCas) -> IO Bool