Chainweb.Sync.WebBlockHeaderStore.Types

Description

TODO

Documentation

data WebBlockHeaderStore

In order to use this a processor for the queue is needed.

The module P2P.TaskQueue provides a P2P session that serves the queue.

TODO

Find a better name

Parameterize in cas

This is currently based on TreeDB (for API) and BlockHeaderDB, it would be possible to run this on top of any CAS and API that offers a simple GET.

Constructors

WebBlockHeaderStore

Fields

_webBlockHeaderStoreCas :: !WebBlockHeaderDb

_webBlockHeaderStoreMemo :: !(TaskMap (ChainValue BlockHash) (ChainValue BlockHeader))

_webBlockHeaderStoreQueue :: !(PQueue (Task ClientEnv (ChainValue BlockHeader)))

_webBlockHeaderStoreLogFunction :: !LogFunction

_webBlockHeaderStoreMgr :: !Manager

data WebBlockPayloadStore cas

Constructors

WebBlockPayloadStore

Fields

_webBlockPayloadStoreCas :: !(PayloadDb cas)

Cas for storing complete payload data including outputs.

_webBlockPayloadStoreMemo :: !(TaskMap BlockPayloadHash PayloadData)

Internal memo table for active tasks

_webBlockPayloadStoreQueue :: !(PQueue (Task ClientEnv PayloadData))

task queue for scheduling tasks with the task server

_webBlockPayloadStoreLogFunction :: !LogFunction

LogFunction

_webBlockPayloadStoreMgr :: !Manager

Manager object for making HTTP requests

_webBlockPayloadStorePact :: !WebPactExecutionService

handle to the pact execution service for validating transactions and computing outputs.

Utils

data PactExecutionService

Constructors

PactExecutionService

Fields

_pactValidateBlock :: BlockHeader -> PayloadData -> IO PayloadWithOutputs

_pactNewBlock :: Miner -> BlockHeader -> BlockCreationTime -> IO PayloadWithOutputs

_pactLocal :: ChainwebTransaction -> IO (Either PactException HashCommandResult)

_pactLookup :: Either ChainId BlockHeader -> Vector PactHash -> IO (Either PactException (Vector (Maybe (T2 BlockHeight BlockHash))))

data ChainValue a

Constructors

ChainValue !ChainId !a

newtype WebBlockHeaderCas

Constructors

WebBlockHeaderCas WebBlockHeaderDb