Chainweb.Time

Description

A time library with a focus on supporting different numeric types for representing time and time spans. This is useful for defining time operations and serialization down to the bit level, including potential arithmetic overflows.

Precision is fixed to microseconds.

TimeSpan

newtype TimeSpan a

The internal unit is microseconds.

Constructors

TimeSpan a

encodeTimeSpan :: MonadPut m => TimeSpan Micros -> m ()

encodeTimeSpanToWord64 :: TimeSpan Micros -> Word64

decodeTimeSpan :: MonadGet m => m (TimeSpan Micros)

castTimeSpan :: NumCast a b => TimeSpan a -> TimeSpan b

maybeCastTimeSpan :: MaybeNumCast a b => TimeSpan a -> Maybe (TimeSpan b)

ceilingTimeSpan :: RealFrac a => Integral b => TimeSpan a -> TimeSpan b

floorTimeSpan :: RealFrac a => Integral b => TimeSpan a -> TimeSpan b

scaleTimeSpan :: Integral a => Num b => a -> TimeSpan b -> TimeSpan b

addTimeSpan :: Num a => TimeSpan a -> TimeSpan a -> TimeSpan a

Time

newtype Time a

Time is measured as microseconds relative to UNIX Epoche

Constructors

Time (TimeSpan a)

minTime :: Bounded a => Time a

maxTime :: Bounded a => Time a

encodeTime :: MonadPut m => Time Micros -> m ()

encodeTimeToWord64 :: Time Micros -> Word64

decodeTime :: MonadGet m => m (Time Micros)

castTime :: NumCast a b => Time a -> Time b

maybeCastTime :: MaybeNumCast a b => Time a -> Maybe (Time b)

ceilingTime :: RealFrac a => Integral b => Time a -> Time b

floorTime :: RealFrac a => Integral b => Time a -> Time b

floorTimeBy :: Integral a => TimeSpan a -> Time a -> Time a

getCurrentTimeIntegral :: Integral a => IO (Time a)

Adhering to Time, this is the current number of microseconds since the epoch.

epoch :: Num a => Time a

timeMicrosQQ :: QuasiQuoter

TimeSpan values

microsecond :: Num a => TimeSpan a

millisecond :: Num a => TimeSpan a

second :: Num a => TimeSpan a

minute :: Num a => TimeSpan a

hour :: Num a => TimeSpan a

day :: Num a => TimeSpan a

Seconds

newtype Seconds

Constructors

Seconds Integer

secondsToTimeSpan :: Num a => Seconds -> TimeSpan a

timeSpanToSeconds :: Integral a => TimeSpan a -> Seconds

Assumes that the TimeSpan contains milliseconds.

secondsToText :: Seconds -> Text

secondsFromText :: MonadThrow m => Text -> m Seconds

Micros

newtype Micros

Will last for around ~300,000 years after the Linux epoch.

Constructors

Micros Int64

microsToTimeSpan :: Num a => Micros -> TimeSpan a

timeSpanToMicros :: Integral a => TimeSpan a -> Micros

Assumes that the TimeSpan contains milliseconds.

microsToText :: Micros -> Text

microsFromText :: MonadThrow m => Text -> m Micros

Math, constants

add :: LeftTorsor t => Diff t -> t -> t

diff :: LeftTorsor t => t -> t -> Diff t

kilo :: Num a => a

mega :: Num a => a