Chainweb.Transaction

Documentation

type ChainwebTransaction = Command PayloadWithText

newtype HashableTrans a

Hashable newtype of ChainwebTransaction

Constructors

HashableTrans

Fields

unHashable :: Command a

data PayloadWithText

A product type representing a `Payload PublicMeta ParsedCode` coupled with the Text that generated it, to make gossiping easier.

chainwebPayloadCodec :: Codec (Command PayloadWithText)

A codec for (Command PayloadWithText) transactions.

gasLimitOf :: forall c. Command (Payload PublicMeta c) -> GasLimit

Get the gas limit/supply of a public chain command payload

gasPriceOf :: forall c. Command (Payload PublicMeta c) -> GasPrice

Get the gas price of a public chain command payload

timeToLiveOf :: forall c. Command (Payload PublicMeta c) -> TTLSeconds

creationTimeOf :: forall c. Command (Payload PublicMeta c) -> TxCreationTime

mkPayloadWithText :: Payload PublicMeta ParsedCode -> PayloadWithText

modifyPayloadWithText :: (Payload PublicMeta ParsedCode -> Payload PublicMeta ParsedCode) -> PayloadWithText -> PayloadWithText

payloadBytes :: PayloadWithText -> ShortByteString

payloadObj :: PayloadWithText -> Payload PublicMeta ParsedCode