Chainweb.TreeDB

Description

Exceptions

data TreeDbException db

Constructors

TreeDbParentMissing (DbEntry db)

TreeDbKeyNotFound (DbKey db)

TreeDbInvalidRank (DbEntry db)

TreeDbAncestorMissing (DbEntry db) Natural Text

Query Parameters

newtype MinRank

Constructors

MinRank (Min Natural)

_getMinRank :: MinRank -> Natural

newtype MaxRank

Constructors

MaxRank (Max Natural)

_getMaxRank :: MaxRank -> Natural

newtype LowerBound k

Constructors

LowerBound

Fields

_getLowerBound :: k

newtype UpperBound k

Constructors

UpperBound

Fields

_getUpperBound :: k

data BranchBounds db

Constructors

BranchBounds

Fields

_branchBoundsLower :: !(HashSet (LowerBound (DbKey db)))

_branchBoundsUpper :: !(HashSet (UpperBound (DbKey db)))

Tree Database

class (Show e, Show (Key e), Eq e, Eq (Key e), Hashable e, Hashable (Key e)) => TreeDbEntry e where

Associated Types

type Key e :: Type

Methods

key :: e -> Key e

rank :: e -> Natural

parent :: e -> Maybe (Key e)

type DbKey db = Key (DbEntry db)

class (Typeable db, TreeDbEntry (DbEntry db)) => TreeDb db where

Minimal complete definition

lookup, entries, (insert | insertStream), maxEntry

Associated Types

type DbEntry db :: Type

Methods

lookup :: db -> DbKey db -> IO (Maybe (DbEntry db))

Lookup a single entry by its key.

Implementations are expected to provide \(\mathcal{O}(1)\) performance.

lookup db k == S.head_ (entries db k 1 Nothing Nothing)

keys :: db -> Maybe (NextItem (DbKey db)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> (Stream (Of (DbKey db)) IO (Natural, Eos) -> IO a) -> IO a

This stream returns a prefix of the keys of the nodes in the tree in ascending order starting from the given key or the genesis block key.

The stream is provide via continuation passing style. This supports implementations that allocate resources for the streams that must be released after the items in the stream are processed. The returned stream must only be used within the scope of the continuation.

If there is no further entry available it terminates and returns a number of returned items and a cursor. Implementations should block on IO at most a constant amount of time.

The default implementation is based on entries, which in some cases doesn't give good performance.

entries :: db -> Maybe (NextItem (DbKey db)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> (Stream (Of (DbEntry db)) IO (Natural, Eos) -> IO a) -> IO a

This stream returns a prefix of the entries of the nodes in the tree in ascending order starting from the given key or the genesis block.

The stream is provide via continuation passing style. This supports implementations that allocate resources for the streams that must be released after the items in the stream are processed. The returned stream must only be used within the scope of the continuation.

If there is no further entry available it terminates and returns a number of returned items and a cursor. Implementations should block on IO at most a constant amount of time.

branchKeys

Arguments

:: db

-> Maybe (NextItem (DbKey db))

-> Maybe Limit

-> Maybe MinRank

-> Maybe MaxRank

-> HashSet (LowerBound (DbKey db))
Lower limits

-> HashSet (UpperBound (DbKey db))
Upper limits

-> (Stream (Of (DbKey db)) IO (Natural, Eos) -> IO a)

-> IO a

branchKeys n l mir mar lower upper returns all nodes within the given range of minimum rank mir and maximun rank mar that are predecessors of nodes in upper and not predecessors of any node in lower, starting at the entry n. The number of items in the result is limited by l. Items are returned in descending order.

The stream is provide via continuation passing style. This supports implementations that allocate resources for the streams that must be released after the items in the stream are processed. The returned stream must only be used within the scope of the continuation.

The result stream may return less than the requested number of items.

branchEntries

Arguments

:: db

-> Maybe (NextItem (DbKey db))
Cursor

-> Maybe Limit
Maximum number of items that are returned

-> Maybe MinRank
Minimum rank for returned items

-> Maybe MaxRank
Maximum rank for returned items

-> HashSet (LowerBound (DbKey db))
Lower bounds for the returned items

-> HashSet (UpperBound (DbKey db))
Upper bounds for the returned items

-> (Stream (Of (DbEntry db)) IO (Natural, Eos) -> IO a)
continuation that is provided the stream of result items

-> IO a

branchEntries n l mir mar lower upper returns all nodes within the given range of minimum rank mir and maximun rank mar that are predecessors of nodes in upper and not predecessors of any node in lower, starting at the entry after n. The number of items in the result is limited by l. Items are returned in descending order.

The stream is provide via continuation passing style. This supports implementations that allocate resources for the streams that must be released after the items in the stream are processed. The returned stream must only be used within the scope of the continuation.

The result stream may return less than the requested number of items.

Also see getBranch for a less powerful but easier to use variant of this function.

insertStream :: db -> Stream (Of (DbEntry db)) IO a -> IO a

insert :: db -> DbEntry db -> IO ()

maxEntry :: db -> IO (DbEntry db)

The largest entry in the database. This is the last entry in the entries stream. It is also an entry of maximal rank.

maxRank :: db -> IO Natural

Maximum rank of all entries in the database.

maxRank db == head (branches db Nothing (Just 1))

Utils

root :: TreeDb db => db -> IO (DbEntry db)

toTree :: (TreeDb db, Ord (DbKey db)) => db -> IO (Tree (DbEntry db))

Interpret a given BlockHeaderDb as a native Haskell Tree. Should be used only for debugging purposes.

Limiting and Seeking a Stream

newtype Eos

Data type to indicate end of stream

Constructors

Eos

Fields

_getEos :: Bool

isEos :: Eos -> Bool

limitStream :: Monad m => Maybe Limit -> Stream (Of a) m () -> Stream (Of a) m (Natural, Eos)

seekStream :: Monad m => Eq k => (a -> k) -> Maybe (NextItem k) -> Stream (Of a) m r -> Stream (Of a) m r

seekStreamSet :: Monad m => Eq k => Hashable k => (a -> k) -> Maybe (NextItem (HashSet k)) -> Stream (Of a) m r -> Stream (Of a) m r

Query branches

getBranch :: forall db. TreeDb db => db -> HashSet (LowerBound (DbKey db)) -> HashSet (UpperBound (DbKey db)) -> Stream (Of (DbEntry db)) IO ()

getBranch db lower upper returns all nodes that are predecessors of nodes in upper and not predecessors of any node in lower. Entries are returned in descending order.

Lookups

lookupM :: forall db. TreeDb db => db -> DbKey db -> IO (DbEntry db)

lookupStreamM :: TreeDb db => db -> Stream (Of (DbKey db)) IO r -> Stream (Of (DbEntry db)) IO r

Lookup all entries in a stream of database keys and return the stream of entries. Throws if an entry is missing.

Stream a foldable value

foldableEntries :: forall e f. Foldable f => TreeDbEntry e => Maybe (NextItem (Key e)) -> Maybe Limit -> Maybe MinRank -> Maybe MaxRank -> f e -> Stream (Of e) IO (Natural, Eos)

Create a Stream from a Foldable value.

Misc Utils

forkEntry :: TreeDb db => db -> DbEntry db -> DbEntry db -> IO (DbEntry db)

Returns the fork entry of two branches of the TreeDb.

branchDiff :: TreeDb db => db -> DbEntry db -> DbEntry db -> Stream (Of (These (DbEntry db) (DbEntry db))) IO ()

Compares two branches of a TreeDb. The fork entry is included as last item in the stream.

If you only need one branch of the fork you may use getBranch instead.

branchDiff_ :: TreeDb db => db -> DbEntry db -> DbEntry db -> Stream (Of (These (DbEntry db) (DbEntry db))) IO (DbEntry db)

Compares two branches of a TreeDb. The fork entry is returned as the result of the stream computation.

If you only need one branch of the fork you may use getBranch instead.

collectForkBlocks :: forall db. TreeDb db => db -> DbEntry db -> DbEntry db -> IO (DbEntry db, Vector (DbEntry db), Vector (DbEntry db))

Collects the blocks on the old and new branches of a fork. Returns (commonAncestor, oldBlocks, newBlocks).

seekAncestor :: forall db. TreeDb db => db -> DbEntry db -> Natural -> IO (Maybe (DbEntry db))

This has expected constant complexity on tree databases that have very long linear trunks and only short branches, which is the case, for instance, for block chains. Worst case complexity is \(O(n)\), and at least twice as much as calling getBranch directly, which is used as fallback.

The function returns Nothing only if the requested rank is larger than the rank of the given header.

properties

properties :: [(String, Property)]