Chainweb.TreeDB.Persist

Description

Reading a TreeDb from - or writing one to - the filesystem.

Writing

persist :: (TreeDb db, Serialize (DbEntry db)) => Path Absolute -> db -> IO ()

Persist the contents of some TreeDb to disk.

Reading

fileEntries :: MonadThrow m => MonadResource m => Serialize e => Path Absolute -> Stream (Of e) m ()

A stream of all entries, decoded from a file.

Throws exceptions if the path given doesn't exist, or if certain ByteStrings within failed to decode.

restore :: (TreeDb db, Serialize (DbEntry db)) => Path Absolute -> db -> IO ()

Given a path to a database written to file via persist and a freshly initialized TreeDb (i.e. it only contains the genesis block), update the TreeDb to contain all entries in the file.

Throws exceptions if the path given doesn't exist, or if certain ByteStrings within failed to decode.

Note: This assumes that reinserting the genesis block into a TreeDb that already contains it is a no-op. This invariant is enforced by tests.