Chainweb.TreeDB.Validation

Description

TODO

Validation

isValidEntryDb

Arguments

:: (TreeDb db, DbEntry db ~ BlockHeader)

=> db
the database

-> BlockHeader
The block header to be checked

-> IO Bool
True if validation succeeded

Tests if the block header is valid (i.e. validateBlockHeader produces an empty list)

This doesn't include any payload validation.

validateEntryDb

Arguments

:: (TreeDb db, DbEntry db ~ BlockHeader)

=> db
the database

-> BlockHeader
The block header to be checked

-> IO [ValidationFailureType]
A list of ways in which the block header isn't valid

Validate properties of the block header, producing a list of the validation failures

This doesn't include any payload validation.

validateEntryDbM

Arguments

:: (TreeDb db, DbEntry db ~ BlockHeader)

=> db
the database

-> BlockHeader
The block header to be checked

-> IO ()

Validate properties of the block header, throwing an exception detailing the failures if any.

This doesn't include any payload validation.

validateAdditionsDbM

Arguments

:: (TreeDb db, DbEntry db ~ BlockHeader)

=> db
the database

-> HashMap BlockHash BlockHeader
The set of block headers to be checked

-> IO ()

Validate a set of additions that are supposed to be added atomically to the database.

This doesn't include any payload validation.