Chainweb.Utils

Description

Utilities used by various modules of the chainweb package

SI unit prefixes

milli :: Fractional a => a

micro :: Fractional a => a

nano :: Fractional a => a

pico :: Fractional a => a

femto :: Fractional a => a

kilo :: Num a => a

mega :: Num a => a

giga :: Num a => a

tera :: Num a => a

peta :: Num a => a

exa :: Num a => a

Misc

int :: Integral a => Num b => a -> b

A shorter alias for fromIntegral

len :: Integral a => [b] -> a

A generalization of length that returns any type that is an instance of Integral.

(==>) :: Bool -> Bool -> Bool infixr 1

Boolean implication operator.

keySet :: HashMap a b -> HashSet a

The set of keys of a HashMap.

minimumsOf :: Ord a => Getting (Endo (Endo [a])) s a -> s -> [a]

The the minimum elements of a list.

minimumsByOf :: Getting (Endo (Endo [a])) s a -> (a -> a -> Ordering) -> s -> [a]

The the minimum elements of a list by some comparision function.

leadingZeros :: Integral int => ByteString -> int

Count leading zero bits of a bytestring

maxBy :: (a -> a -> Ordering) -> a -> a -> a

The maximum of two value by some comparision function.

minBy :: (a -> a -> Ordering) -> a -> a -> a

The minimum of two value by some comparision function.

allEqOn :: Foldable f => Eq b => (a -> b) -> f a -> Bool

Checks that all elements of foldable structure are equal under the given mapping.

roundBy :: Integral a => a -> a -> a

Round an integral n up to the nearest multiple of an integral m

unlessM :: Monad m => m Bool -> m () -> m ()

A version of unless with a monadic predicate.

whenM :: Monad m => m Bool -> m () -> m ()

A version of when with a monadic predicate.

ebool_ :: e -> Bool -> Either e ()

partitionEithersNEL :: NonEmpty (Either a b) -> These (NonEmpty a) (NonEmpty b)

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

class IxedGet a where

Provides a simple Fold lets you fold the value at a given key in a Map or element at an ordinal position in a list or Seq.

This is a restrictec version of Ixed from the lens package that prevents the value at the key from being modified.

Minimal complete definition

Nothing

Methods

ixg :: Index a -> Fold a (IxValue a)

Provide a Fold for a value at a given key.

default ixg :: Ixed a => Index a -> Fold a (IxValue a)

Encoding and Serialization

data EncodingException where

Exceptions that are thrown when encoding or decoding values.

Constructors

EncodeException :: Text -> EncodingException

DecodeException :: Text -> EncodingException

Base64DecodeException :: Text -> EncodingException

ItemCountDecodeException :: Expected Natural -> Actual Natural -> EncodingException

TextFormatException :: Text -> EncodingException

JsonDecodeException :: Text -> EncodingException

X509CertificateDecodeException :: Text -> EncodingException

X509KeyDecodeException :: Text -> EncodingException

Binary

runGet :: MonadThrow m => Get a -> ByteString -> m a

Decode a value from a ByteString. In case of a failure a DecodeException is thrown.

runPut :: Put -> ByteString

Encode a value into a ByteString.

runGetEither :: Get a -> ByteString -> Either EncodingException a

Decode a value from a ByteString and return either the result or a DecodeException.

Codecs

data Codec t

A binary codec.

TODO: maybe use Put/Get ?

Constructors

Codec

Fields

codecEncode :: t -> ByteString

codecDecode :: ByteString -> Either String t

Text

sshow :: Show a => IsString b => a -> b

Show a value as any type that is an instance of IsString.

tread :: Read a => Text -> Either Text a

Read a value from a textual encoding using its Read instance. Returns and textual error message if the operation fails.

treadM :: MonadThrow m => Read a => Text -> m a

Throws TextFormatException on failure.

class HasTextRepresentation a where

Class of types that have an textual representation.

Methods

toText :: a -> Text

fromText :: MonadThrow m => Text -> m a

eitherFromText :: HasTextRepresentation a => Text -> Either String a

Decode a value from its textual representation.

unsafeFromText :: HasCallStack => HasTextRepresentation a => Text -> a

Unsafely decode a value rom its textual representation. It is an program error if decoding fails.

parseM :: MonadThrow m => Parser a -> Text -> m a

Run a Parser on a text input. All input must be consume by the parser. A TextFormatException is thrown if parsing fails.

parseText :: HasTextRepresentation a => Parser Text -> Parser a

A parser for types with an HasTextRepresentation instance.

Base64

encodeB64Text :: ByteString -> Text

Encode a binary value to a textual base64 representation.

decodeB64Text :: MonadThrow m => Text -> m ByteString

Decode a binary value from a textual base64 representation. A Base64DecodeException is thrown if the input is not a valid base64 encoding.

encodeB64UrlText :: ByteString -> Text

Encode a binary value to a textual base64-url representation.

decodeB64UrlText :: MonadThrow m => Text -> m ByteString

Decode a binary value from a textual base64-url representation. A Base64DecodeException is thrown if the input is not a valid base64-url encoding.

encodeB64UrlNoPaddingText :: ByteString -> Text

Encode a binary value to a textual base64-url without padding representation.

decodeB64UrlNoPaddingText :: MonadThrow m => Text -> m ByteString

Decode a binary value from a textual base64-url without padding representation. A Base64DecodeException is thrown if the input is not a valid base64-url without padding encoding.

JSON

encodeToText :: ToJSON a => a -> Text

Encode a value to a JSON text.

encodeToByteString :: ToJSON a => a -> ByteString

Encode a value to a strict ByteString.

decodeOrThrow :: MonadThrow m => FromJSON a => ByteString -> m a

Decode a JSON value from a lazy ByteString. If decoding fails a JsonDecodeException is thrown.

decodeStrictOrThrow :: MonadThrow m => FromJSON a => ByteString -> m a

Decode a JSON value from a strict ByteString. If decoding fails a JsonDecodeException is thrown.

decodeFileStrictOrThrow :: MonadIO m => MonadThrow m => FromJSON a => FilePath -> m a

Decode a JSON value from the contents of a file. If decoding fails a JsonDecodeException is thrown.

This function parses immediately, but defers conversion. See json for details.

decodeOrThrow' :: MonadThrow m => FromJSON a => ByteString -> m a

Strictly decode a JSON value from a lazy ByteString. If decoding fails a JsonDecodeException is thrown.

decodeStrictOrThrow' :: MonadThrow m => FromJSON a => ByteString -> m a

Strictly decode a JSON value from a strict ByteString. If decoding fails a JsonDecodeException is thrown.

decodeFileStrictOrThrow' :: forall a m. MonadIO m => MonadThrow m => FromJSON a => FilePath -> m a

Strictly decode a JSON value from the content of a file. If decoding fails a JsonDecodeException is thrown.

parseJsonFromText :: HasTextRepresentation a => String -> Value -> Parser a

A json parser for types with an instance of HasTextRepresentation.

Error Handling

newtype Expected a

A newtype wrapper for tagger values as "expected" outcomes of some computation.

Constructors

Expected

Fields

getExpected :: a

newtype Actual a

A newtype wrapper for tagger values as "actual" outcomes of some computation.

Constructors

Actual

Fields

getActual :: a

unexpectedMsg :: Show a => Text -> Expected a -> Actual a -> Text

A textual message that describes the Expected and the Actual outcome of some computation.

(==?) :: Eq a => Expected a -> Actual a -> Bool

Compare an Expected with an Actual value.

check :: MonadThrow m => Eq a => Exception e => (Expected a -> Actual a -> e) -> Expected a -> Actual a -> m a

Compare an Expected with an Actual value and raise an exception if they are different.

fromMaybeM :: MonadThrow m => Exception e => e -> Maybe a -> m a

Throw an exception if a value is Nothing.

fromJuste :: HasCallStack => Maybe a -> a

Like fromJust, but carries forward the HasCallStack constraint. Juste is French for Just.

(???) :: MonadThrow m => Exception e => Maybe a -> e -> m a infixl 0

Flipped infix version of fromMaybeM

fromEitherM :: MonadThrow m => Exception e => Either e a -> m a

Throw an exception if a value is a Left result.

newtype InternalInvariantViolation

An exeption to indicate an violation of an internal code invariants. Throwing this type of exception means that there is a bug in the code.

Constructors

InternalInvariantViolation Text

eatIOExceptions :: IO () -> IO ()

Catch and strictly evaluate any IOExceptions.

This function should be used with great care because operation may silently fail without leaving a trace. This can hide issues in the code making them very difficult to debug.

Synchronous Exceptions

catchSynchronous :: MonadCatch m => Exception e => NFData a => m a -> (e -> m a) -> m a

Catch and handle exception that are not contained in SomeAsyncException.

catchAllSynchronous :: MonadCatch m => NFData a => m a -> (SomeException -> m a) -> m a

A version of catchSynchronous that doesn't discriminate on the type of the exception but handles all exceptions in SomeExeption that are not contained in SomeAsyncException.

trySynchronous :: MonadCatch m => Exception e => NFData a => m a -> m (Either e a)

Catch all exceptions and return 'Left e' for all exceptions e that are not contained in SomeAsyncException. Asynchronous exceptions are re-thrown.

tryAllSynchronous :: MonadCatch m => NFData a => m a -> m (Either SomeException a)

A version of tryhSynchronous that doesn't discriminate on the type of the exception but handles all exceptions in SomeExeption that are not contained in SomeAsyncException.

runForever :: (LogLevel -> Text -> IO ()) -> Text -> IO () -> IO ()

Repeatedly run a computation forever until it is stopped by receiving SomeAsyncException.

If the computation throws an exception that is not contained in SomeAsyncException an error message is logged and the function continues to repeat the computation forever.

An info-level message is logged when processing starts and stops.

runForeverThrottled

Arguments

:: (LogLevel -> Text -> IO ())

-> Text

-> Word64
burst size (number of calls)

-> Word64
rate limit (usec / call)

-> IO ()

-> IO ()

Repeatedly run a computation forever at the given rate until it is stopped by receiving SomeAsyncException.

If the computation throws an exception that is not contained in SomeAsyncException an error message is logged and the function continues to repeat the computation forever.

An info-level message is logged when processing starts and stops.

Command Line Options

type OptionParser a = Parser a

Type of parsers for simple compandline options.

prefixLong :: HasName f => Maybe String -> String -> Mod f a

An option modifier that defines a long option name that may be prefixed.

prefixLong (Just "component") "option" results in the command line option --component-option.

suffixHelp :: Maybe String -> String -> Mod f a

An option modifier that defines an help message to which a suffix may be appended.

suffixHelp (Just "component") "Help message" result in the help text "Help message for component"

textReader :: HasTextRepresentation a => ReadM a

A command line option reader for types with an instance of HasTextRepresentation.

textOption :: HasTextRepresentation a => Mod OptionFields a -> Parser a

A option parser for types with an instance of HasTextRepresentation.

jsonOption :: FromJSON a => Mod OptionFields a -> Parser a

Configuration to Enable/Disable Components

data EnableConfig a

Configuration wrapper to enable and disable components

Constructors

EnableConfig

Fields

_enableConfigEnabled :: !Bool

A flag that indicates whether the component is enabled.

_enableConfigConfig :: !a

The configuration value of the component.

enableConfigConfig :: forall a a. Lens (EnableConfig a) (EnableConfig a) a a

enableConfigEnabled :: forall a. Lens' (EnableConfig a) Bool

defaultEnableConfig :: a -> EnableConfig a

The default is that the configured component is enabled.

pEnableConfig :: String -> MParser a -> MParser (EnableConfig a)

Command line parser for the configuration of a component that can be enabled or disabled.

enabledConfig :: EnableConfig a -> Maybe a

Access wrapped config as Maybe if enabled.

validateEnableConfig :: ConfigValidation a l -> ConfigValidation (EnableConfig a) l

Configuration Exception

newtype ConfigurationException

Exeption that is thrown when a configuration value is invalid.

Constructors

ConfigurationException Text

Streaming

streamToHashSet :: Monad m => Eq a => Hashable a => Stream (Of a) m r -> m (Of (HashSet a) r)

Fold the items of a stream into a HashSet.

streamToHashSet_ :: Monad m => Eq a => Hashable a => Stream (Of a) m r -> m (HashSet a)

Fold the items of a stream into a HashSet and discard the result of the stream.

nub :: Monad m => Eq a => Stream (Of a) m r -> Stream (Of a) m r

Drop successive equal items from a stream.

S.toList_ (nub (S.each [1,1,2,2,3,4,4,4])) == [1,2,3,4]

timeoutStream :: Int -> Stream (Of a) IO r -> Stream (Of a) IO (Maybe r)

Add a timeout to each step of a stream.

If a step takes longer than the timeout processing of the stream is terminated and Nothing is returned. After processing all items of the stream without a timeout Just the result of the stream is returned.

reverseStream :: Monad m => Stream (Of a) m () -> Stream (Of a) m ()

This function reverses the order of the items in a stream. In order to due so it store all items of the stream in a memory buffer before continuing to stream starting at the former end of the stream.

THIS FUNCTION BREAKS STREAMING! USE WITH CARE!

The function getBranch returns the keys of the branch in the order such that an key of an entry is returned before the keys of the dependencies of the entry are returned.

Public ChainDb API functions require that items are returned in an order such that an item is returned after all dependencies of the item are returned.

Storing the complete result of a stream query in memory is problematic for streams of unbounded length. It it is particularly problematic in a server application. Beside of potential long latencies and GC issue overhead it can also represent a DOS attack vulnerability. A server that uses this function should use paging aggressively.

Filesystem

withTempDir :: String -> (Path Absolute -> IO a) -> IO a

Perform an action over a random path under /tmp. Example path:

Path "tmpchainweb-git-store-test-8086816238120523704"

Type Level

symbolText :: forall s a. KnownSymbol s => IsString a => a

Return the value of a type level symbol as a value of a type that is an instance of IsString.

optics

locally :: MonadReader s m => ASetter s s a b -> (a -> b) -> m r -> m r

Resource Management

concurrentWith :: forall a b t d. Traversable t => (forall c. a -> (b -> IO c) -> IO c) -> (t b -> IO d) -> t a -> IO d

Bracket style resource managment uses CPS style which only supports sequential componsition of different allocation functions.

This function is a hack that allows to allocate and deallocate several resources concurrently and provide them to a single inner computation.

withLink :: forall a. IO a -> IO (Async a)

Run an async IO action, link it back to the main thread, and return the async result

Strict Tuples

sfst :: T2 a b -> a

ssnd :: T2 a b -> b

scurry :: (T2 a b -> c) -> a -> b -> c

suncurry :: (a -> b -> c) -> T2 a b -> c

suncurry3 :: (a -> b -> c -> d) -> T3 a b c -> d

rwipe3 :: T3 a b c -> T2 b c

Approximate thread delays

approximateThreadDelay :: Int -> IO ()