Chainweb.Utils.Paging

Description

TODO

Limit

newtype Limit

Limit the result of a query to a maximum number of items

Constructors

Limit

Fields

_getLimit :: Natural

Page

data Page k a

Constructors

Page

Fields

_pageLimit :: !Limit

The number of items in the page

_pageItems :: ![a]

The items of the page

_pageNext :: !(Maybe k)

A cursor for querying the next page, if there is any. The value is given the next parameter of the respective query interface.

pageLimit :: forall k a. Lens' (Page k a) Limit

pageItems :: forall k a a. Lens (Page k a) (Page k a) [a] [a]

pageNext :: forall k a k. Lens (Page k a) (Page k a) (Maybe k) (Maybe k)

Next Item

data NextItem k

When seeking a position in a stream, define if the given position is inclusive or exclusive.

Inclusive: return all items of the stream starting with the given key. Exclusive: return all items of the stream starting immidiately after the given key.

Constructors

Inclusive k

Exclusive k

_getNextItem :: NextItem k -> k

getNextItem :: Getter (NextItem k) k

isExclusive :: NextItem k -> Bool

isInclusive :: NextItem k -> Bool

nextItemToText :: HasTextRepresentation k => NextItem k -> Text

nextItemFromText :: MonadThrow m => HasTextRepresentation k => Text -> m (NextItem k)

End-Of-Stream

newtype Eos

Data type to indicate end of stream

Constructors

Eos

Fields

_getEos :: Bool

isEos :: Eos -> Bool

atEos :: Monad m => Stream (Of a) m () -> m Eos

Tools for creating pages from streams

finitePrefixOfInfiniteStreamToPage :: MonadThrow m => (a -> k) -> Maybe Limit -> Stream (Of a) m () -> m (Page (NextItem k) a)

Create page from a non-empty stream that is a non-blocking finite prefix of a possibly blocking infinite stream.

If the given stream contains more items than requested by the caller an Inclusive cursor is added to the page. Otherwise the last item of the stream is added as Exclusive cursor.

If the input stream is empty we assume that it is because of a limiting filter that results in a query for a finite stream. No cursor is returned.

For an empty input we can't return a next cursor. We can't return just Nothing because that is used in a Page to signal the end of the stream, which contradicts the assumption that the input stream is the prefix of an infinite stream. So, when we see an empty stream we assume that it's empty because of some filter and return Nothing

finiteStreamToPage :: Monad m => (a -> k) -> Maybe Limit -> Stream (Of a) m () -> m (Page (NextItem k) a)

Create Page from a (possibly empty) prefix of a non-blocking finite stream. If the input stream has more than the requested number of items an Inclusive cursor is added. Otherwise it is assumed that the stream has ended and Nothing is returned as cursor.

seekFiniteStreamToPage :: Monad m => Eq k => (a -> k) -> Maybe (NextItem k) -> Maybe Limit -> Stream (Of a) m () -> m (Page (NextItem k) a)

Quick and dirty pagin implementation. Usage should be avoided.

Properties

properties :: [(String, Property)]