Chainweb.Version

Description

TODO

Documentation

data ChainwebVersion

Generally, a chain is uniquely identified by it's genesis block. For efficiency and convenience we explicitely propagate ChainwebVersion and the ChainId to all blocks in the chain. At runtime the ChainId is represented at the type level (but included as value in serialized representations). Thus, the ChainwebVersion identifies a chain at runtime at the value level.

We assume that values that are identified through different Chainweb versions are not mixed at runtime. This is not enforced at the type level.

Constructors

Test ChainGraph
General-purpose test instance, where:

the underlying ChainGraph is configurable,

the genesis block time is the Linux epoch,

each HashTarget is maxBound,

each mining Nonce is constant,

the creationTime of BlockHeaders is the parent time plus one second, and

POW is simulated by poison process thread delay.

This is primarily used in unit tests.

TimedConsensus ChainGraph
Test instance for confirming the behaviour of our Consensus mechanisms (Cut processing, Header validation, etc.), where:

the underlying ChainGraph is configurable,

the genesis block time is the Linux epoch,

each HashTarget is maxBound,

each mining Nonce is constant,

the creationTime of BlockHeaders is the actual time,

POW is simulated by poison process thread delay, and

there are no Pact or mempool operations running.

This is primarily used in our slow-tests executable.

PowConsensus ChainGraph
Test instance for confirming the behaviour of the Proof-of-Work mining algorithm and Difficulty Adjustment, where:

the underlying ChainGraph is configurable,

the genesis block time the current time,

the genesis HashTarget is 7 bits lower than maxBound,

the Nonce changes with each mining attempt,

creationTime of BlockHeaders is the actual time, and

there are no Pact or mempool operations running.

This is primarily used in our slow-tests executable.

TimedCPM ChainGraph
Test instance for confirming the combined behaviour of our Consensus mechanisms, Pact code processing and validation, and Mempool, where:

the underlying ChainGraph is configurable,

the genesis block time is the Linux epoch,

each HashTarget is maxBound,

each mining Nonce is constant,

the creationTime of BlockHeaders is the actual time,

POW is simulated by poison process thread delay, and

the Pact Service and Mempool operations are running.

This is primarily used in our run-nodes executable.

FastTimedCPM ChainGraph
Test instance for confirming the combined behaviour of our Consensus mechanisms, Pact code processing and validation, and Mempool, where:

the underlying ChainGraph is configurable,

the genesis block time is the Linux epoch,

each HashTarget is maxBound,

each mining Nonce is constant,

the creationTime of BlockHeader's is the actual time,

POW is not simulated by poison process thread delay, and

the Pact Service and Mempool operations are running.

This is primarily used in our standalone executable.

Development
An instance which has no guarantees about the long-term stability of its parameters. They are free to change as developers require.

Testnet02

Mainnet01

encodeChainwebVersion :: MonadPut m => ChainwebVersion -> m ()

decodeChainwebVersion :: MonadGet m => m ChainwebVersion

chainwebVersionFromText :: MonadThrow m => Text -> m ChainwebVersion

Read textual representation of a ChainwebVersion.

chainwebVersionToText :: HasCallStack => ChainwebVersion -> Text

chainwebVersionId :: ChainwebVersion -> Word32

This function and its dual fromChainwebVersionId are used to efficiently serialize a ChainwebVersion and its associated internal ChainGraph value. This function must be injective (one-to-one)! The scheme is as follows:

Production ChainwebVersions start from 0x00000001 and count upwards. Their value must be less than 0x8000000, but this limit is unlikely to ever be reached.

ChainwebVersions for testing begin at 0x80000000, as can be seen in toTestChainwebVersion. This value is combined (via .|.) with the "code" of their associated ChainGraph (as seen in graphToCode). Such codes start at 0x00010000 and count upwards.

Properties of Chainweb Version

POW

newtype BlockRate

The gap in SECONDS that we desire between the Creation Time of subsequent blocks in some chain.

Constructors

BlockRate Seconds

blockRate :: ChainwebVersion -> BlockRate

The Proof-of-Work BlockRate for each ChainwebVersion. This is the number of seconds we expect to pass while a miner mines on various chains, eventually succeeding on one.

newtype WindowWidth

The number of blocks to be mined after a difficulty adjustment, before considering a further adjustment. Critical for the "epoch-based" adjustment algorithm seen in adjust.

Constructors

WindowWidth Natural

window :: ChainwebVersion -> Maybe WindowWidth

The Proof-of-Work WindowWidth for each ChainwebVersion. For chainwebs that do not expect to perform POW, this should be Nothing.

Date-based Transaction Disabling

txSilenceEndDate :: ChainwebVersion -> Maybe (Time Micros)

Typelevel ChainwebVersion

newtype ChainwebVersionT

Constructors

ChainwebVersionT Symbol

type family ChainwebVersionSymbol n :: Symbol

chainwebVersionSymbolVal :: KnownChainwebVersionSymbol n => Proxy n -> Text

data SomeChainwebVersionT

Constructors

forall (a :: ChainwebVersionT).KnownChainwebVersionSymbol a => SomeChainwebVersionT (Proxy a)

class KnownSymbol (ChainwebVersionSymbol n) => KnownChainwebVersionSymbol (n :: ChainwebVersionT)

Minimal complete definition

chainwebVersionSymbolVal

someChainwebVersionVal :: ChainwebVersion -> SomeChainwebVersionT

Singletons

data family Sing :: k -> Type

Kind-indexed family of singleton types.

In order to define singletons for types of a data kind A, one defines an instance that has a value constructor for each type constructor of A. The following is an example how to define singletons for types of kind Maybe a:

data instance Sing :: forall a . Maybe a -> Type where
   SNothing :: Sing 'Nothing
   SJust :: Sing a -> Sing ('Just a)

type SChainwebVersion (v :: ChainwebVersionT) = Sing v

HasChainwebVersion

class HasChainwebVersion a where

Minimal complete definition

_chainwebVersion | chainwebVersion

Methods

_chainwebVersion :: a -> ChainwebVersion

chainwebVersion :: Getter a ChainwebVersion

mkChainId :: MonadThrow m => HasChainwebVersion v => Integral i => v -> i -> m ChainId

chainIds :: HasChainwebVersion v => v -> HashSet ChainId

someChainId :: HasCallStack => HasChainwebVersion v => v -> ChainId

Sometimes, in particular for testing and examples, some fixed chain id is needed, but it doesn't matter which one. This function provides some valid chain ids.

randomChainId :: HasChainwebVersion v => v -> IO ChainId

Uniformily get a random ChainId

ChainId

module Chainweb.ChainId

Re-exports from Chainweb.ChainGraph

Chain Graph

data ChainGraph

class HasChainGraph a where

Minimal complete definition

_chainGraph | chainGraph

Methods

_chainGraph :: a -> ChainGraph

chainGraph :: Getter a ChainGraph

adjacentChainIds :: HasChainId p => ChainGraph -> p -> HashSet ChainId

Graph Properties

order :: ChainGraph -> Natural

The order of the graph refers to the number of vertices |V| of a given graph.

diameter :: ChainGraph -> Natural

degree :: ChainGraph -> Natural

shortestPath :: ChainId -> ChainId -> ChainGraph -> [ChainId]

Undirected Edges

data AdjPair a

_getAdjPair :: AdjPair a -> (a, a)

pattern Adj :: HasChainId a => a -> a -> AdjPair a

adjs :: ChainGraph -> HashSet (AdjPair ChainId)

adjsOfVertex :: HasChainId p => ChainGraph -> p -> HashSet (AdjPair ChainId)

checkAdjacentChainIds :: MonadThrow m => HasChainGraph g => HasChainId cid => HasChainId adj => g -> cid -> Expected (HashSet adj) -> m (HashSet adj)

Given a ChainGraph g, checkAdjacentChainIds cid as checks that the ChainId cid is in g and the set of adjacents chain ids of cid is the expected set as.