Chainweb.WebBlockHeaderDB

Description

TODO

Documentation

data WebBlockHeaderDb

Every WebChain has the following properties

All entires of _webBlockHeaderDb are valid BlockHeaderDbs

There are no dangling adjacent parent hashes

The adjacent hashes of all block headers conform with the chain graph of the web chain.

TODO: in order to enforce these invariants the insertion to the dbs must be guarded see issue #123.

mkWebBlockHeaderDb :: ChainwebVersion -> HashMap ChainId BlockHeaderDb -> WebBlockHeaderDb

FIXME: this needs some consistency checks

initWebBlockHeaderDb :: RocksDb -> ChainwebVersion -> IO WebBlockHeaderDb

getWebBlockHeaderDb :: MonadThrow m => HasChainId p => Given WebBlockHeaderDb => p -> m BlockHeaderDb

webBlockHeaderDb :: Getter WebBlockHeaderDb (HashMap ChainId BlockHeaderDb)

webEntries :: WebBlockHeaderDb -> (Stream (Of BlockHeader) IO () -> IO a) -> IO a

Returns all blocks in all block header databases.

lookupWebBlockHeaderDb :: Given WebBlockHeaderDb => ChainId -> BlockHash -> IO BlockHeader

lookupAdjacentParentHeader :: Given WebBlockHeaderDb => BlockHeader -> ChainId -> IO BlockHeader

lookupParentHeader :: Given WebBlockHeaderDb => BlockHeader -> IO BlockHeader

insertWebBlockHeaderDb :: Given WebBlockHeaderDb => BlockHeader -> IO ()

blockAdjacentParentHeaders :: Given WebBlockHeaderDb => BlockHeader -> IO (HashMap ChainId BlockHeader)

checkBlockHeaderGraph :: MonadThrow m => BlockHeader -> m ()

Given a ChainGraph g, checkBlockHeaderGraph h checks that the _chainId h is a vertex in g and that the adjacent hashes of h correspond exactly to the adjacent vertices of h in g.

TODO: move this to Chainweb.BlockHeader?

checkBlockAdjacentParents :: Given WebBlockHeaderDb => BlockHeader -> IO ()

Given a WebBlockHeaderDb db, checkBlockAdjacentParents h checks that all referenced adjacent parents block headers exist in db.

Orphan instances

IxedGet WebBlockHeaderDb

Instance details

Methods

ixg :: Index WebBlockHeaderDb -> Fold WebBlockHeaderDb (IxValue WebBlockHeaderDb)