Chainweb.WebBlockHeaderDB.Types

Description

TODO

Documentation

data WebBlockHeaderDb

Every WebChain has the following properties

All entires of _webBlockHeaderDb are valid BlockHeaderDbs

There are no dangling adjacent parent hashes

The adjacent hashes of all block headers conform with the chain graph of the web chain.

TODO: in order to enforce these invariants the insertion to the dbs must be guarded see issue #123.

Constructors

WebBlockHeaderDb

Fields

_webBlockHeaderDb :: !(HashMap ChainId BlockHeaderDb)

_webChainwebVersion :: !ChainwebVersion