Control.Concurrent.FixedThreadPool

Documentation

data ThreadPool

data Action

newThreadPool :: Int -> IO ThreadPool

stopThreadPool :: ThreadPool -> IO ()

killThreadPool :: ThreadPool -> IO ()

waitThreadPool :: ThreadPool -> IO ()

withThreadPool :: Int -> (ThreadPool -> IO a) -> IO a

runAction :: ThreadPool -> IO () -> IO Action

Runs an IO action asynchronously. Returns an Action you can wait on with waitAction.

runActionGroup :: Traversable t => ThreadPool -> t (IO ()) -> IO (Maybe SomeException)

Runs a group of IO actions and waits for them to finish. If any of the IO actions throws an exception, runActionGroup will return it. If more than one action throws an exception, runActionGroup will arbitrarily return one of the exceptions. The runActionGroup function will wait until all of the actions have run, even if an exception is thrown.

mapAction :: Traversable t => ThreadPool -> (a -> IO b) -> t a -> IO (t (Either SomeException b))

mapAction_ :: Traversable t => ThreadPool -> (a -> IO b) -> t a -> IO ()

waitAction :: Action -> IO (Maybe SomeException)