Data.IVar

Description

Strict immutable variables with distinguished read and write end points.

Classes

class IsRIVar v

The class of readable immutable variables.

Minimal complete definition

tryReadIVar, awaitIVar

awaitIVar :: IsRIVar v => v a -> IO a

tryReadIVar :: IsRIVar v => v a -> IO (Maybe a)

class IsWIVar v

The class of writeable immutable variables.

Minimal complete definition

putIVar, putIVar'

type IsIVar v = (IsRIVar v, IsWIVar v)

The class of immutable variables that are readable and writeable.

putIVar :: IsWIVar v => v a -> a -> IO Bool

IVar

data IVar a

An immutable MVar. Once a value has been assigned to the variable it can't be changed any more.

newIVar :: IO (IVar a)

Create a new immutable variable. Intitially the variable is empty. Once a value got assigned that value can't be changed any more.

mkWeakIVar :: IVar a -> IO () -> IO (Weak (IVar a))

Create an weak IVar. For details and caveats regarding weak references, please, see the documentation for mkWeakMVar.

Writable IVar

data WIVar a

An immutable variable that one can write to (exactly once), but that isn't readable. This is, for instance, useful to model one end of a one-directional communitcation channel.

wIVar :: IVar a -> WIVar a

Restrict an IVar to be only writeabel but not readable.

Readble IVar

data RIVar a

An immutable variable that one can write to (exactly once), but that isn't readable. This is, for instance, useful to model one end of a one-directional communitcation channel.

rIVar :: IVar a -> RIVar a

Restrict an IVar to be only readable but not writeable.