Data.LogMessage

Description

This module defines log messages that are similar to Haskell exceptions from Exception. Like exceptions, log messages in this module are typed dynamically and classes of log messages types are extensible.

Log messages and exceptions are similar in that they can be emitted/thrown anywhere in the code base (in the IO monad, for logs) and are propagated to handlers that are defined upward in the call stack until they are eventually picked up. The difference is that exceptions synchronously interrupt the computation that throws them, while log messages are usually handled asynchronously and the computation that emits them continues while the message is handled.

Log messages are usually handled only at the top level by a global handler (or stack of handlers), but that depends on the implementation of the logger (usually a queue, but sometimes just an IO callback), which is orthorgonal to the definitions in this module.

Documentation

data SomeLogMessage

The SomeLogMessage type is the root of the log message type hierarchy. When a log message of type a is emitted, behind the scenes it is encapsulated in a SomeLogMessage.

Constructors

forall a.LogMessage a => SomeLogMessage a

class (NFData a, Typeable a) => LogMessage a where

The class of log messages.

Log messages must be instances of NFData and Typeable and must have a textual representation. The default instance uses the Show instance of a type.

Minimal complete definition

Nothing

Methods

logText :: a -> Text

default logText :: Show a => a -> Text

toLogMessage :: a -> SomeLogMessage

fromLogMessage :: SomeLogMessage -> Maybe a

Log Function

type LogFunction = forall a. LogMessage a => LogLevel -> a -> IO ()

Type of a log functions

type LogFunctionText = LogLevel -> Text -> IO ()

LogFunction type specialized to Text

type LogFunctionJson a = (Typeable a, NFData a, ToJSON a) => LogLevel -> a -> IO ()

LogFunction type specialized to JSON

newtype ALogFunction

A newtype wrapper that allows to store a LogFunction without running into impredicative types.

Constructors

ALogFunction

Fields

_getLogFunction :: LogFunction

alogFunction :: forall a. LogMessage a => ALogFunction -> LogLevel -> a -> IO ()

Get a LogFunction from ALogFunction

aNoLog :: ALogFunction

ALogFunction that discards all log messages

LogMessage types

newtype JsonLog a

A newtype wrapper for log messages types with a ToJSON instance.

Constructors

JsonLog a

data SomeJsonLog

A dynamically polymorphic wrapper for any log message type that has a ToJSON instance.

Constructors

forall a.(NFData a, ToJSON a) => SomeJsonLog a

newtype TextLog

A newtype wrapper for textual log messages.

Constructors

TextLog Text

data BinaryLog

Binary log messages.

Constructors

BinaryLog ByteString

BinaryLogLazy ByteString

data SomeSymbolLog

Static textual log messages using Symbol literals from TypeLits.

Constructors

forall (a :: Symbol).KnownSymbol a => SomeSymbolLog (Proxy a)