Data.PQueue

Description

TODO

Documentation

data PQueue a

A thread safe priority queue.

The queue is fair for users of the queue. It does not guarantee progress for items in the queue. An item of low priority my starve in the queue if higher priority items are added a rate at least as high as items are removed.

newEmptyPQueue :: IO (PQueue a)

pQueueInsert :: Ord a => PQueue a -> a -> IO ()

pQueueInsertLimit :: Ord a => PQueue a -> Natural -> a -> IO ()

pQueueRemove :: PQueue a -> IO a

It the queue is empty it blocks and races for new items

pQueueIsEmpty :: PQueue a -> IO Bool

pQueueSize :: PQueue a -> IO Natural