Data.Singletons

Description

This module contains a selection of the definitions from the module Data.Singletons.Internal from the singletons package.

The copyright and license of the singletons package applies to the respective definitions in this module.

Data family of singletons

data family Sing :: k -> Type

Kind-indexed family of singleton types.

In order to define singletons for types of a data kind A, one defines an instance that has a value constructor for each type constructor of A. The following is an example how to define singletons for types of kind Maybe a:

data instance Sing :: forall a . Maybe a -> Type where
   SNothing :: Sing 'Nothing
   SJust :: Sing a -> Sing ('Just a)

class SingI (a :: k) where

The class of types that have singletons.

Methods

sing :: Sing a

pattern Sing :: forall k (a :: k). () => SingI a => Sing a

A pattern for converting between a singlton and the corresponding SingInstance.

Sing Kind

class SingKind k where

The class of kinds for which singletons are defined.

Associated Types

type Demote k = (r :: Type) | r -> k

Methods

fromSing :: Sing (a :: k) -> Demote k

toSing :: Demote k -> SomeSing k

data SomeSing k where

Existentially quantified singleton type. This allows to hide the type of the singleton from the copmiler. The type can be brought into scope at runtime by pattern matching on the SomeSing constructor.

This can, for instance, be used for creating singletons from user provided values are deserialized values.

Constructors

SomeSing :: Sing (a :: k) -> SomeSing k

pattern FromSing :: SingKind k => forall (a :: k). Sing a -> Demote k

A pattern for converting betwen a singleton and it's demoted value.

Sing Instance

data SingInstance (a :: k) where

A data type to enclose and explicitly pass around an SingI dictionary.

Constructors

SingInstance :: SingI a => SingInstance a

singInstance :: forall k (a :: k). Sing a -> SingInstance a

Create a SingInstance value from a singleton. Pattern matching on the resulting value brings the respective SingI constraint into scope.

Utils

type KindOf (a :: k) = k

Obtain the kind of a type variable.

demote :: forall a. SingKind (KindOf a) => SingI a => Demote (KindOf a)

Return the demoted value of a singleton.

withSomeSing :: forall k r. SingKind k => Demote k -> (forall (a :: k). Sing a -> r) -> r

Provide a computation with the singlton for a value.

withSing :: SingI a => (Sing a -> b) -> b

Given a SingI instance provide an inner computation with an explicit singlton value.

withSingI :: Sing n -> (SingI n => r) -> r

Given an singleton value, provide an inner computation with a SingI insteance.

singThat :: forall k (a :: k). SingKind k => SingI a => (Demote k -> Bool) -> Maybe (Sing a)

Provide Just a singleton for a value that satisfies a predicate. If the value doesn't satify the predicate Nothing is returned.

singByProxy :: SingI a => proxy a -> Sing a

Get a singleton value from a type proxy