Data.TaskMap

Description

TODO

Documentation

data TaskMap k v

new :: Eq k => Hashable k => IO (TaskMap k v)

insert :: Eq k => Hashable k => TaskMap k v -> k -> IO v -> IO (Async v)

delete :: Eq k => Hashable k => TaskMap k v -> k -> IO ()

lookup :: Eq k => Hashable k => TaskMap k v -> k -> IO (Maybe (Async v))

await :: Eq k => Hashable k => TaskMap k v -> k -> IO (Maybe v)

If the key is in the map this blocks and awaits the results, otherwise it returns Nothing immediately.

size :: TaskMap k v -> IO Int

null :: TaskMap k v -> IO Bool

memo

Arguments

:: Eq k

=> Hashable k

=> TaskMap k v

-> k

-> (k -> IO v)
an action that is used to produce the value if the key isn't in the map.

-> IO v