Network.X509.SelfSigned

Description

TODO

Supported X509 Certificate Types

type Ed25519Cert = SecretKey

The signature algorithm Ed25519. EdDSA signature algorithms are not yet supported by most clients. It is part of TLSv1.3 and supported in openssl-1.1 (with TLSv1.2 and TLS1.3) and the master branch (as of 2019-01-10) of Haskell tls (with TLSv1.2 and TLSv1.3). Also, the under high load the performance of the Haskell tls package is better with 4096 bit RSA certificates than with ED25519 certificates.

type Ed448Cert = SecretKey

The signature algorithm Ed25519. EdDSA signature algorithms are not yet supported by most clients. It is part of TLSv1.3 and supported in openssl-1.1 (with TLSv1.2 and TLS1.3) and the master branch (as of 2019-01-10) of Haskell tls (with TLSv1.2 and TLSv1.3).

type RsaCert = PrivateKey

RSA signature algorithm. Generating RSA keys is somewhat slow and the (sufficiently secure) public keys are large. But currently it is still the most widely supported certificate type. For the Haskell tls <= 1.4.1 it seems that this is the only certificate type that is fully supported on the server side.

type P256Cert = KeyPair

ECDSA signature algorithm. p-256 (or prime256v1 or secp256r1) is used with ecdsa. it is supported by most standard clients, but disabled as signing algorithm on the server-side of the Haskell tls package due to an timing attack vulnerability in the generic ECDSA implementation.

https://github.com/ocheron/cryptonite/commits/tc-ecdsa https://github.com/ocheron/hs-tls/commits/ecdsa-signing

The former resolves the timing attack vulnerability for P-256 (SEC_p521r1) in the cryptonite package. The latter implements signing with ECDSA certificates in the tls package. With the increasing support for Ed25519 and Ed448 accross the internet it is not clear if support for ECDSA with NIST curves will ever be added to the Haskell tls package.

type DefCertType = RsaCert

class X509Key k

Minimal complete definition

generateKey, publicKey, getPubKey, getPrivKey, sigAlg, signIO, pemKeyHeader

Generate PEM encoded self-signed X509 Certificates

generateSelfSignedCertificate

Arguments

:: forall k. X509Key k

=> Natural
days of validity

-> DistinguishedName
the subject of the certificate. This should include at least the CN name, which can be created with name.

-> Maybe (NonEmpty AltName)
For self-signed certificates it's usually fine to leave this empty. If it's not empty it should include the CN value from the distinguished name.

-> IO (Fingerprint, X509CertPem, X509KeyPem)

Generate a self-signed Certificate for a given distinguished name and optionally additional alternative names. Note, however, that for a most uses of a self-signed cerificate the names don't matter. Instead the certificate is authenticated by checking its fingerprint.

generateLocalhostCertificate

Arguments

:: forall k. X509Key k

=> Natural
days of validity

-> IO (Fingerprint, X509CertPem, X509KeyPem)

Generate a self-signed Certificate for localhost. Note, however, that for a most uses of a self-signed cerificate the names don't matter. Instead the certificate is authenticated by checking its fingerprint.

makeCertificate

Arguments

:: forall k. X509Key k

=> DateTime

-> DateTime

-> DistinguishedName

-> DistinguishedName

-> Maybe (NonEmpty AltName)
the list of altnames must include the CN name from the subject distinguished name.

-> PubKey

-> Certificate

Client Support

data TlsPolicy

A certificate policy for using self-signed certifcates with a connection manager

Constructors

TlsInsecure

TlsSecure

Fields

_tlsPolicyUseSystemStore :: !Bool

whether to use the system certificate store.

_tlsPolicyCallback :: !(ServiceID -> IO (Maybe Fingerprint))

a callback for looking up certificate fingerprints for known service endpoints.

type ServiceID = (HostName, ByteString)

unsafeMakeCredential :: HasCallStack => X509CertChainPem -> X509KeyPem -> Credential

Create a client credential for a certificate

certificateCacheManagerSettings :: TlsPolicy -> Maybe Credential -> IO ManagerSettings

Create a connection manager for a given TlsPolicy.

isCertificateMismatchException :: HttpException -> Bool

Check whether a connection failed due to an certificate missmatch

Server Settings

tlsServerSettings :: X509CertPem -> X509KeyPem -> TLSSettings

TLS server settings

tlsServerChainSettings :: X509CertChainPem -> X509KeyPem -> TLSSettings

TLS server settings

Low level Utils

Distinguished Name and Alt Names

name :: String -> DistinguishedName

org :: String -> DistinguishedName

data AltName

Constructors

AltNameRFC822 String

AltNameDNS String

AltNameURI String

AltNameIP ByteString

AltNameXMPP String

AltNameDNSSRV String

Certificate Fingerprints

type FingerprintByteCount = 32

fingerprintByteCount :: Natural

newtype Fingerprint

Constructors

Fingerprint ByteString

fingerprint :: SignedExact Certificate -> Fingerprint

fingerprintToText :: Fingerprint -> Text

fingerprintFromText :: MonadThrow m => Text -> m Fingerprint

unsafeFingerprintFromText :: HasCallStack => String -> Fingerprint

fingerprintPem :: MonadThrow m => X509CertPem -> m Fingerprint

unsafeFingerprintPem :: HasCallStack => X509CertPem -> Fingerprint

PEM encoded Certificate

newtype X509CertPem

Constructors

X509CertPem ByteString

x509CertPemToText :: X509CertPem -> Text

x509CertPemFromText :: MonadThrow m => Text -> m X509CertPem

pX509CertPem :: Maybe String -> OptionParser X509CertPem

unsafeX509CertPemFromText :: HasCallStack => String -> X509CertPem

validateX509CertPem :: MonadError Text m => X509CertPem -> m ()

decodePemX509Cert :: MonadThrow m => X509CertPem -> m (SignedExact Certificate)

PEM encoded Key

newtype X509KeyPem

Constructors

X509KeyPem ByteString

x509KeyPemToText :: X509KeyPem -> Text

x509KeyPemFromText :: MonadThrow m => Text -> m X509KeyPem

pX509KeyPem :: Maybe String -> OptionParser X509KeyPem

unsafeX509KeyPemFromText :: HasCallStack => String -> X509KeyPem

validateX509KeyPem :: MonadError Text m => X509KeyPem -> m ()

decodePemX509Key :: MonadThrow m => X509KeyPem -> m PrivKey

PEM encode certificate chain

data X509CertChainPem

Constructors

X509CertChainPem X509CertPem ![X509CertPem]

x509CertChainPemToText :: X509CertChainPem -> Text

x509CertChainPemFromText :: MonadThrow m => Text -> m X509CertChainPem

pX509CertChainPem :: Maybe String -> OptionParser X509CertChainPem

unsafeX509CertChainPemFromText :: HasCallStack => String -> X509CertChainPem

validateX509CertChainPem :: MonadError Text m => X509CertChainPem -> m ()