Numeric.Additive

Description

Haskell's Num class doesn't support fine grained control over what arithmetic operations are defined for a type. Sometimes only some operations have a well defined semantics and Num instances are notorious for including undefined/error values or unlawful workarounds.

Additive Semigroup

class AdditiveSemigroup g where

Additive Semigroup

(a `plus` b) `plus` c == a `plus` (b `plus` c)

Methods

plus :: g -> g -> g

class AdditiveSemigroup g => AdditiveAbelianSemigroup g

Additive Abelian Semigroup

a `plus` b == b `plus` a

(^+^) :: AdditiveAbelianSemigroup g => g -> g -> g infixl 6

Additive Monoid

class AdditiveSemigroup g => AdditiveMonoid g where

Additive Monoid

a `plus` zero == a

zero `plus` a == a

Methods

zero :: g

type AdditiveAbelianMonoid g = (AdditiveMonoid g, AdditiveAbelianSemigroup g)

Additive Group

class AdditiveMonoid g => AdditiveGroup g where

Additive Group

a `plus` inverse a == zero

inverse a `plus` a == zero

Minimal complete definition

invert | minus

Methods

invert :: g -> g

minus :: g -> g -> g

Additive Abelian Group

type AdditiveAbelianGroup g = (AdditiveGroup g, AdditiveAbelianMonoid g)

Additive Abelian Group

(^-^) :: AdditiveAbelianGroup g => g -> g -> g infix 6