Numeric.AffineSpace

Description

Torsor

class AdditiveGroup (Diff t) => LeftTorsor t where

A torsor is a generalization of affine spaces. It doesn't require the underlying structure to be vector space, but an additive group suffices. This means that it doesn't support scalar multiplication. In particular it doesn't require an inverse operation to multiplication, which would add unneeded complexity to the formal definition of the operational semantics.

A Torsor is also called principal homogeous space.

zero `add` a == a

(a `plus` b) `add` t == a `add` (b `add` t)

(s `diff` t) `add` t == s

The last property is states that add is a bijection.

Associated Types

type Diff t

Methods

add :: Diff t -> t -> t

diff :: t -> t -> Diff t

(.+^) :: AdditiveAbelianGroup (Diff t) => LeftTorsor t => t -> Diff t -> t infixl 6

(^+.) :: AdditiveAbelianGroup (Diff t) => LeftTorsor t => Diff t -> t -> t infixl 6

(.-.) :: AdditiveAbelianGroup (Diff t) => LeftTorsor t => t -> t -> Diff t infix 6

(.-^) :: AdditiveAbelianGroup (Diff t) => LeftTorsor t => t -> Diff t -> t infixl 6

Vector Space

class (AdditiveAbelianGroup v, Fractional (Scalar v)) => FractionalVectorSpace v where

Vector Space over Fractional Numbers

A real vector space is an additive abelian group that forms an module with the field of real numbers.

a * (b `scale` c) == (a * b) `scale` c

1 `scale` a == a

a `scale` (b `plus` c) == (a `scale` b) `plus` (a `scale` c)

(a + b) `scale` c == (a `scale` c) `plus` (b `scale` c)

Associated Types

type Scalar v

Methods

scale :: Scalar v -> v -> v

AfficeSpace

type AffineSpace t = (FractionalVectorSpace (Diff t), LeftTorsor t)

An affine space is a torsor for the action of the additive group of a vector space.