P2P.Node.Configuration

Description

TODO

P2P Configuration

data P2pConfiguration

Configuration of the Network

TODO: add ChainwebVersion?

Constructors

P2pConfiguration

Fields

_p2pConfigPeer :: !PeerConfig

the local peer.

_p2pConfigMaxSessionCount :: !Natural

the number of active peers.

_p2pConfigMaxPeerCount :: !Natural

total number of peers

_p2pConfigSessionTimeout :: !Seconds

interval at which peers are rotated out of the active set

_p2pConfigKnownPeers :: ![PeerInfo]

List of known peers. Must not be empty.

_p2pConfigPeerDbFilePath :: !(Maybe FilePath)

the path where the peer database is persisted

_p2pConfigIgnoreBootstrapNodes :: !Bool

p2pConfigPeer :: Lens' P2pConfiguration PeerConfig

p2pConfigMaxSessionCount :: Lens' P2pConfiguration Natural

p2pConfigMaxPeerCount :: Lens' P2pConfiguration Natural

p2pConfigSessionTimeout :: Lens' P2pConfiguration Seconds

p2pConfigKnownPeers :: Lens' P2pConfiguration [PeerInfo]

p2pConfigPeerDbFilePath :: Lens' P2pConfiguration (Maybe FilePath)

p2pConfigIgnoreBootstrapNodes :: Lens' P2pConfiguration Bool

defaultP2pConfiguration :: P2pConfiguration

These are acceptable values for both test and production chainwebs.

pP2pConfiguration :: Maybe NetworkId -> MParser P2pConfiguration