P2P.Node.PeerDB

Description

TODO

Index Values

newtype LastSuccess

Constructors

LastSuccess

Fields

_getLastSuccess :: Maybe UTCTime

newtype SuccessiveFailures

Constructors

SuccessiveFailures

Fields

_getSuccessiveFailures :: Natural

newtype AddedTime

Constructors

AddedTime

Fields

_getAddedTime :: UTCTime

newtype ActiveSessionCount

Constructors

ActiveSessionCount

Fields

_getActiveSessionCount :: Natural

Peer Entry

data PeerEntry

Constructors

PeerEntry

Fields

_peerEntryInfo :: !PeerInfo

There must be only one peer per peer address. A peer id can be updated from Nothing to Just some value. If a peer id of Just some value changes, it is considered a new peer and the existing value is replaced.

_peerEntrySuccessiveFailures :: !SuccessiveFailures

The number of successive failure for this peer. If this number execeeds a certain threshold we drop the peer from the database.

_peerEntryLastSuccess :: !LastSuccess

The time of the last successful interaction with this peer.

_peerEntryNetworkIds :: !(Set NetworkId)

The set of networks that this peer supports.

_peerEntryActiveSessionCount :: !ActiveSessionCount

number of currently active sessions with this peer. By trying to maximize this number, the sharing of peers between different network ids is increased. There should be only one active session per network id.

peerEntryInfo :: Lens' PeerEntry PeerInfo

peerEntrySuccessiveFailures :: Lens' PeerEntry SuccessiveFailures

peerEntryLastSuccess :: Lens' PeerEntry LastSuccess

peerEntryNetworkIds :: Lens' PeerEntry (Set NetworkId)

Peer Database

data PeerDb

Constructors

PeerDb (MVar ()) (TVar PeerSet)

peerDbSnapshot :: PeerDb -> IO PeerSet

peerDbSnapshotSTM :: PeerDb -> STM PeerSet

peerDbSize :: PeerDb -> IO Natural

peerDbSizeSTM :: PeerDb -> STM Natural

peerDbInsert :: PeerDb -> NetworkId -> PeerInfo -> IO ()

Adds new PeerInfo values for a given chain id.

This function is fair. If there are multiple concurrent writers each writer is guaranteed to eventually write. It has also robust performance under contention.

peerDbInsertList :: [PeerEntry] -> PeerDb -> IO ()

peerDbInsertPeerInfoList :: NetworkId -> [PeerInfo] -> PeerDb -> IO ()

peerDbInsertSet :: Set PeerEntry -> PeerDb -> IO ()

peerDbDelete :: PeerDb -> PeerInfo -> IO ()

Delete a peer, identified by its host address, from the peer database.

newEmptyPeerDb :: IO PeerDb

fromPeerEntryList :: [PeerEntry] -> IO PeerDb

fromPeerInfoList :: NetworkId -> [PeerInfo] -> IO PeerDb

Update PeerDb Entries

updateLastSuccess :: PeerDb -> PeerInfo -> IO ()

resetSuccessiveFailures :: PeerDb -> PeerInfo -> IO ()

incrementSuccessiveFailures :: PeerDb -> PeerInfo -> IO ()

incrementActiveSessionCount :: PeerDb -> PeerInfo -> IO ()

decrementActiveSessionCount :: PeerDb -> PeerInfo -> IO ()

Persistence

storePeerDb :: FilePath -> PeerDb -> IO ()

Atomically store the database to a file.

TODO: Currently, this is somewhat inefficient, in particular at shutdown. It would be better to stream a transaction log to the temporary file and consolidate in the background only when there is enough time.

loadPeerDb :: FilePath -> IO PeerDb

loadIntoPeerDb :: FilePath -> PeerDb -> IO ()

New entries overwrite existing entries

Some PeerDb

newtype PeerDbT (v :: ChainwebVersionT) (n :: NetworkIdT)

PeerDb with type level ChainwebVersion and NetworkIdT indexes

Constructors

PeerDbT PeerDb

data SomePeerDb

Constructors

forall v n.(KnownChainwebVersionSymbol v, SingI n) => SomePeerDb (PeerDbT v n)

somePeerDbVal :: ChainwebVersion -> NetworkId -> PeerDb -> SomePeerDb

properties

properties :: [(String, Property)]