P2P.Peer

Description

TODO

Peer Id

newtype PeerId

Constructors

PeerId ByteString

peerIdToText :: PeerId -> Text

shortPeerId :: PeerId -> Text

peerIdFromText :: MonadThrow m => Text -> m PeerId

unsafePeerIdFromText :: HasCallStack => String -> PeerId

pPeerId :: Maybe String -> OptionParser PeerId

peerIdToFingerprint :: PeerId -> Fingerprint

peerIdFromFingerprint :: Fingerprint -> PeerId

Peer Info

data PeerInfo

TODO: eventually this should have more information, like, for instance, the API version that the peer supports, public key, etc.

Constructors

PeerInfo

Fields

_peerId :: !(Maybe PeerId)

The peer id wraps the SHA256 fingerprint of the X509 certificate.

If this is Nothing we assume that the HostAddress fully identifies the peer and it has certificate that can be validated from the system CA store.

_peerAddr :: !HostAddress

peerId :: Lens' PeerInfo (Maybe PeerId)

peerAddr :: Lens' PeerInfo HostAddress

pPeerInfo :: Maybe String -> MParser PeerInfo

pPeerInfoCompact :: Maybe String -> OptionParser PeerInfo

Parser Peer Id as a single option

arbitraryPeerInfo :: Gen PeerInfo

peerInfoClientEnv :: Manager -> PeerInfo -> ClientEnv

Create a ClientEnv for querying HTTP API of a PeerInfo

Peer Configuration

data PeerConfig

Constructors

PeerConfig

Fields

_peerConfigAddr :: !HostAddress

The public host address of the peer. A port number of 0 means that a free port is assigned by the system.

_peerConfigInterface :: !HostPreference

The network interface that the peer binds to. Default is to bind to all available interfaces (*).

_peerConfigCertificateChain :: !(Maybe X509CertChainPem)

The X509 certificate chain of the peer. If this is Nothing a new ephemeral certificate is generated on startup and discarded on exit.

_peerConfigCertificateChainFile :: !(Maybe FilePath)

A file with the X509 certificate chain of the peer. If this is Nothing a new ephemeral certificate is generated on startup and discarded on exit.

If _peerConfigCertificateChain is not Nothing it has precedence.

_peerConfigKey :: !(Maybe X509KeyPem)

The key for the X509 certificate. If no certificate is provided the key is ignored. It is an error if a certificate is provided but no key.

_peerConfigKeyFile :: !(Maybe FilePath)

A file with key for the X509 certificate. If no certificate is provided the key is ignored. It is an error if a certificate is provided but no key.

If _peerConfigKey is not Nothing it has precedence.

peerConfigAddr :: Lens' PeerConfig HostAddress

peerConfigInterface :: Lens' PeerConfig HostPreference

peerConfigCertificateChain :: Lens' PeerConfig (Maybe X509CertChainPem)

peerConfigKey :: Lens' PeerConfig (Maybe X509KeyPem)

defaultPeerConfig :: PeerConfig

_peerConfigPort :: PeerConfig -> Port

peerConfigPort :: Lens' PeerConfig Port

_peerConfigHost :: PeerConfig -> Hostname

peerConfigHost :: Lens' PeerConfig Hostname

pPeerConfig :: Maybe String -> MParser PeerConfig

peerInfoHostname :: Lens' PeerInfo Hostname

peerInfoPort :: Lens' PeerInfo Port

shortPeerInfo :: PeerInfo -> Text

Peer

data Peer

A peer in a P2P network

Constructors

Peer

Fields

_peerInfo :: !PeerInfo

The peer id is the SHA256 fingerprint of the certificate

_peerInterface :: !HostPreference

_peerCertificateChain :: !X509CertChainPem

_peerKey :: !X509KeyPem

peerInfo :: Lens' Peer PeerInfo

peerInterface :: Lens' Peer HostPreference

peerCertificateChain :: Lens' Peer X509CertChainPem

peerKey :: Lens' Peer X509KeyPem

unsafeCreatePeer :: PeerConfig -> IO Peer

getPeerCertificate :: PeerConfig -> IO (Fingerprint, X509CertChainPem, X509KeyPem)

Bootstrap Peer Infos

bootstrapPeerInfos :: ChainwebVersion -> [PeerInfo]

For each chainweb version there is a hardcoded set of bootstrap nodes for the P2P network.

If a bootstrap node has an public DNS name with an official TLS certificate the peer-id should be omitted. For bootstrap nodes without an proper certificate, the peer id is the SHA256 hash of the X509 certificate.