P2P.TaskQueue

Description

TODO

Tasks

type TaskAction env a = LogFunction -> env -> IO a

newtype TaskId

Constructors

TaskId Text

newtype AttemptsCount

Constructors

AttemptsCount Natural

newtype Priority

Constructors

Priority Int

data Task env a

Constructors

Task

Fields

_taskId :: !TaskId

_taskAction :: !(TaskAction env a)

_taskPriority :: !Priority

_taskAttemptsCount :: !AttemptsCount

_taskResult :: !(IVar (Either [SomeException] a))

_taskFailures :: ![SomeException]

newTask :: TaskId -> Priority -> TaskAction env a -> IO (Task env a)

awaitTask :: Task env a -> IO a

taskResult :: Task env a -> RIVar (Either [SomeException] a)

P2P Session for Running Queued Tasks

session :: AttemptsCount -> PQueue (Task env a) -> LogFunction -> env -> PeerInfo -> IO Bool

This session is intended to be used by a P2P node. It's type is general enough to be useful in other contexts, too.

session_ :: AttemptsCount -> PQueue (Task env a) -> LogFunction -> env -> IO Bool

Exceptions

data TaskException

Constructors

TaskFailed [SomeException]