Utils.Logging

Description

This module provides backends for handling log messages from the Data.LogMessage. Log messages that are similar to Haskell exceptions from Exception. Like exceptions, log messages in this module are typed dynamically and classes of log messages types are extensible.

Log messages and exceptions are similar in that they can be emitted/thrown anywhere in the code base (in the IO monad, for logs) and are propagated to handlers that are defined upward in the call stack until they are eventually picked up. The difference is that exceptions synchronously interrupt the computation that throws them, while log messages are usually handled asynchronously and the computation that emits them continues while the message is handled.

Log messages are usually handled only at the top level by a global handler (or stack of handlers), but that depends on the implementation of the logger (usually a queue, but sometimes just an IO callback), which is orthorgonal to the definitions in this module. The backends in this module use the type from the package yet-another-logger, which also provides an implementation of a logger queue.

Unlike exceptions, log messages must be handled. The type systems ensures that there is a base log handler that catches messages of any type, even if it just discards all messages.

Base Logger Backend

type SomeBackend = GenericBackend (Maybe (BackendLogMessage SomeLogMessage))

A backend that is used for partial log message handlers that may pass messages on to subsequent handlers.

Log Handlers Backends

type Backend a = LogMessage a -> IO ()

type TextBackend = Backend TextLog

type JsonBackend a = Backend (JsonLog a)

type SomeJsonBackend = Backend SomeJsonLog

Log Message Handlers

logHandle :: Monoid b => LogMessage a => Backend a -> GenericBackend b -> GenericBackend b

A log handle that handles all messages of a given type.

maybeLogHandle :: Monoid b => LogMessage a => (LogMessage a -> IO (Maybe (LogMessage SomeLogMessage))) -> GenericBackend b -> GenericBackend b

A log message handle that may emit a new log message. It can for instance be used to forward unhandled log messages.

genericLogHandle :: Monoid b => LogMessage a => (BackendLogMessage a -> IO (Maybe (BackendLogMessage SomeLogMessage))) -> GenericBackend b -> GenericBackend b

This is most the powerful handle function that allows to implement generic base backends and to inspect internal log messages of the logging system.

Logging Backend Stacks

data LogHandler

Constructors

forall a.LogMessage a => LogHandler (BackendLogMessage a -> IO (Maybe (BackendLogMessage SomeLogMessage)))

logHandler :: forall a. LogMessage a => Backend a -> LogHandler

maybeLogHandler :: forall a. LogMessage a => (LogMessage a -> IO (Maybe (LogMessage SomeLogMessage))) -> LogHandler

logHandles :: Monoid b => Foldable f => f LogHandler -> GenericBackend b -> GenericBackend b

Filter LogScope Backend

data LogFilter

A filter for log messages.

Tese are the rules for processing a log message:

If a log label of a message matches the key and value of a rule, the message is discarded if the log level of the message is larger than the log level of the rule.

If, after applying all rules, no rule matched any label of the log message, the message is discarded if the log level of the message is larger than the default level of the filter.

These semantics seem to be useful under certain circumstances. At least, the order of the rules doesn't matter.

When a filter is specified more than once in a configuration, filters are merged by concatenating the lists of rules and by taking the minimum log level.

The default log filter has no rules and log level debug as default level.

Constructors

LogFilter

Fields

_logFilterRules :: ![(LogLabel, LogLevel)]

_logFilterDefault :: !LogLevel

logFilterRules :: Lens' LogFilter [(LogLabel, LogLevel)]

logFilterDefault :: Lens' LogFilter LogLevel

logFilterHandle :: LogFilter -> LogHandler

A handle that applies a log filter.

The filter is applied after log messages have been emitted according to the log level of the respective logger.

Base Backends

withBaseHandleBackend :: forall b. Text -> Manager -> [(Text, Text)] -> BackendConfig -> (BaseBackend -> IO b) -> IO b

A Base Backend for arbitrary log messages. All message bodies are encoded using logText.

If a logfile is used, the file is opend in write mode and previous content is deleted.

If the Elasticsearch handle is used, the logs are send to an Elasticsearchserver and the format setting is ignored.

Text Formatting:

A textual representation is produced for each log message. Log messages are separated by newline charachters. No escaping is performed on newline characters that are embedded in the logs.

JSON Formatting:

One JSON value is produced for each log message. JSON values are separated by newline characters. Note that the output of this logger as a whole doesn't represent a valid JSON document.

Message bodies are encoded as a JSON string. A possibly existing ToJSON instance is ignored. For formatting the message body using the ToJSON instances log messages can be wrapped into JsonLog or instance the or the function withJsonHandleBackend can be used.

TODO: should we try to cast to 'SomeJsonLog a' or is that to much magic?

Specialized Backends

withTextHandleBackend :: forall a. Text -> Manager -> [(Text, Text)] -> BackendConfig -> (TextBackend -> IO a) -> IO a

This Logger logs only TextLog messages.

withJsonHandleBackend :: forall a b. ToJSON a => Text -> Manager -> [(Text, Text)] -> BackendConfig -> (Backend a -> IO b) -> IO b

This logger encodes the bodies of the log messages using the ToJSON instance of the respective type.

If a logfile is used, the file is opend in write mode and previous content is deleted.

If the Elasticsearch handle is used, the logs are send to an Elasticsearchserver and the format setting is ignored.

Text Formatting:

A textual representation is produced for each log message. Log messages are separated by newline charachters. Message bodies are encoded using the respective ToJSON instance. The formatted log messages contain no unescaped line break characters.

JSON Formatting:

One JSON value is produced for each log message. JSON values are separated by newline characters. Note that the output of this logger as a whole doesn't represent a valid JSON document.

Message bodies are encoded using the respective ToJSON instance.

withJsonEventSourceBackend :: ToJSON a => Port -> (Backend a -> IO b) -> IO b

A backend for JSON log messages that publishes messages via an HTTP event source on the given port.

Of-the-shelf logger

withFileHandleLogger :: (MonadIO m, MonadBaseControl IO m) => LogConfig -> (Logger SomeLogMessage -> m α) -> m α

Example Logger

withExampleLogger

Arguments

:: Port

-> LoggerConfig
Logger configuration

-> BackendConfig
Logger backend configurationjk

-> FilePath

-> (Logger SomeLogMessage -> IO a)

-> IO a

Configuration

configureHandler :: (c -> (Backend b -> IO a) -> IO a) -> EnableConfig c -> (Backend b -> IO a) -> IO a

Enables a logger in a log-handler stack based on its EnabledConfig wrapper. If the logger is disabled, messages are passed to the inner backend.

Usage Example:

withEnabledJsonHandleBackend mgr = configureHandler (withJsonHandleBackend mgr)

Utils

toBackendLogMessage :: LogMessage a => BackendLogMessage a -> BackendLogMessage SomeLogMessage

fromBackendLogMessage :: LogMessage a => BackendLogMessage SomeLogMessage -> Maybe (BackendLogMessage a)

Orphan instances

Functor LogMessage

Instance details

Methods

fmap :: (a -> b) -> LogMessage a -> LogMessage b

(<$) :: a -> LogMessage b -> LogMessage a