Utils.Logging.Config

Description

Configuration for generic logging utils

Log Format

data LogFormat

Constructors

LogFormatText

LogFormatJson

logFormatToText :: IsString p => LogFormat -> p

logFormatFromText :: MonadThrow m => Text -> m LogFormat

pLogFormat :: OptionParser LogFormat

pLogFormat_

Arguments

:: Text
prefix for the command line options.

-> OptionParser LogFormat

Logging Backend Handle

data HandleConfig

Constructors

StdOut

StdErr

FileHandle FilePath

ElasticSearch Text

pHandleConfig :: OptionParser HandleConfig

pHandleConfig_

Arguments

:: Text
prefix for the command line options.

-> OptionParser HandleConfig

validateHandleConfig :: ConfigValidation HandleConfig l

Logger Backend

data BackendConfig

BackendConfig

Constructors

BackendConfig

Fields

_backendConfigColor :: !ColorOption

_backendConfigFormat :: !LogFormat

_backendConfigHandle :: !HandleConfig

backendConfigColor :: Lens' BackendConfig ColorOption

backendConfigHandle :: Lens' BackendConfig HandleConfig

defaultBackendConfig :: BackendConfig

validateBackendConfig :: ConfigValidation BackendConfig []

pBackendConfig :: MParser BackendConfig

pBackendConfig_

Arguments

:: Text
prefix for this and all subordinate command line options.

-> MParser BackendConfig

A version of pLoggerBackendConfig that takes a prefix for the command line option.